Odds of correct classification (OCC)

Computes the odds of correct classification (OCC) for each cluster. In other words, it computes the proportion of trajectories that can be expected to be correctly classified by the model for each cluster.

OCC(object)

Arguments

object

The model, of type lcModel.

Value

The OCC per cluster, as a numeric vector of length nClusters(object). Empty clusters will output NA.

Details

An OCC of 1 indicates that the cluster assignment is no better than by random chance.

References

Nagin DS (2005). Group-based modeling of development. Harvard University Press. ISBN 9780674041318, doi:10.4159/9780674041318 . Klijn SL, Weijenberg MP, Lemmens P, van den Brandt PA, Passos VL (2017). “Introducing the fit-criteria assessment plot - A visualisation tool to assist class enumeration in group-based trajectory modelling.” Statistical Methods in Medical Research, 26(5), 2424-2436. doi:10.1177/0962280215598665 . van der Nest G, Lima Passos V, Candel MJ, van Breukelen GJ (2020). “An overview of mixture modelling for latent evolutions in longitudinal data: Modelling approaches, fit statistics and software.” Advances in Life Course Research, 43, 100323. ISSN 1040-2608, doi:10.1016/j.alcr.2019.100323 .

See also