Extract the cluster trajectories

Extracts a data frame of all cluster trajectories.

# S4 method for lcModel
clusterTrajectories(object, at = time(object), what = "mu", ...)

Arguments

object
The lcModel object.

at
An optional vector of the times at which to compute the cluster trajectory predictions.

what
The distributional parameter to predict. By default, the mean response 'mu' is predicted. The cluster membership predictions can be obtained by specifying what = 'mb'.

...
Additional arguments.

Value

A data.frame of the estimated values at the given times. The first column should be named "Cluster". The second column should be time, with the name matching the timeVariable(object). The third column should be the expected value of the observations, named after the responseVariable(object).

See also

Examples

model <- latrend(method = lcMethodLcmmGMM(fixed = Y ~ Time, mixture = fixed),
  id = "Id", time = "Time", data = latrendData)

#> Be patient, hlme is running ... 
#> The program took 0.09 seconds 
#> Be patient, hlme is running ... 
#> The program took 0.4 seconds 

clusterTrajectories(model)

#>    Cluster      Time           Y
#> 1        A 0.0000000 -0.72947437
#> 2        A 0.2222222 -0.57446041
#> 3        A 0.4444444 -0.41944644
#> 4        A 0.6666667 -0.26443248
#> 5        A 0.8888889 -0.10941852
#> 6        A 1.1111111  0.04559545
#> 7        A 1.3333333  0.20060941
#> 8        A 1.5555556  0.35562338
#> 9        A 1.7777778  0.51063734
#> 10       A 2.0000000  0.66565131
#> 11       B 0.0000000  0.11281051
#> 12       B 0.2222222  0.08293362
#> 13       B 0.4444444  0.05305673
#> 14       B 0.6666667  0.02317984
#> 15       B 0.8888889 -0.00669705
#> 16       B 1.1111111 -0.03657394
#> 17       B 1.3333333 -0.06645083
#> 18       B 1.5555556 -0.09632772
#> 19       B 1.7777778 -0.12620461
#> 20       B 2.0000000 -0.15608150

clusterTrajectories(model, at = c(0, .5, 1))

#>   Cluster Time           Y
#> 1       A  0.0 -0.72947437
#> 2       A  0.5 -0.38069295
#> 3       A  1.0 -0.03191153
#> 4       B  0.0  0.11281051
#> 5       B  0.5  0.04558751
#> 6       B  1.0 -0.02163549