Extract lcModel coefficients

Extract the coefficients of the lcModel object, if defined. The returned set of coefficients depends on the underlying type of lcModel. The default implementation checks for the existence of a coef() function for the internal model as defined in the @model slot, returning the output if available.

# S3 method for lcModel
coef(object, ...)

Arguments

object

The lcModel object.

...

Additional arguments.

Value

A named numeric vector with all coefficients, or a matrix with each column containing the cluster-specific coefficients. If coef() is not defined for the given model, an empty numeric vector is returned.

Implementation

Classes extending lcModel can override this method to return model-specific coefficients.

coef.lcModelExt <- function(object, ...) {
  # return model coefficients
}

See also

Examples

data(latrendData)
method <- lcMethodLcmmGBTM(fixed = Y ~ Time, mixture = ~ 1,
  id = "Id", time = "Time", nClusters = 3)
gbtm <- latrend(method, data = latrendData)
#> Be patient, hlme is running ... 
#> The program took 0.25 seconds 
coef(gbtm)
#> intercept class1 intercept class2 intercept class1 intercept class2 
#>      -0.90377813       1.62577286      -0.75800717      -0.29874492 
#> intercept class3             Time           stderr 
#>       0.09769888       0.24004793       0.35704201