Compute the posterior confusion matrix

Compute the posterior confusion matrix (PCM). The entry \((i,j)\) represents the probability (or number, in case of scale = TRUE) of a trajectory belonging to cluster \(i\) is assigned to cluster \(j\) under the specified trajectory cluster assignment strategy.

confusionMatrix(object, strategy = which.max, scale = TRUE, ...)

Arguments

object

The model, of type lcModel.

strategy

The strategy for assigning trajectories to a specific cluster, see trajectoryAssignments(). If strategy = NULL, the posterior probabilities are used as weights (analogous to a repeated evaluation of strategy = which.weight).

scale

Whether to express the confusion in probabilities (scale = TRUE), or in terms of the number of trajectories.

...

Arguments passed on to trajectoryAssignments

Value

A K-by-K confusion matrix with K = nClusters(object).

See also

Examples

data(latrendData)

if (rlang::is_installed("lcmm")) {
  method <- lcMethodLcmmGMM(
    fixed = Y ~ Time,
    mixture = ~ Time,
    random = ~ 1,
    id = "Id",
    time = "Time"
  )
  model <- latrend(method, latrendData)
  confusionMatrix(model)
}
#> ------------------ iteration  0 ------------------
#> Function value  -2214.962 
#> Convergence criteria: parameters stability= 1.0001 
#>                     : function stability= 1.0001 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef   SE.coef   Var.coef
#> parameter1 -0.02626082  3.672290   13.48571
#> parameter2  0.00000000  5.009514   25.09523
#> parameter3  1.00000000 11.170348 -124.77667
#> parameter4  1.00000000 31.494353 -991.89424
#> 
#> ------------------ iteration  1 ------------------
#> Function value  -2214.946 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 0.01528693 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef   SE.coef   Var.coef
#> parameter1 -0.02626170  3.669229   13.46324
#> parameter2  0.00000196  4.698516   22.07605
#> parameter3  0.99999000 11.170594 -124.78217
#> parameter4  0.99999151 31.494423 -991.89866
#> 
#> ------------------ iteration  2 ------------------
#> Function value  -2214.929 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 0.01761045 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef   SE.coef   Var.coef
#> parameter1 -0.02626236  3.673180   13.49225
#> parameter2  0.00000453  4.258475   18.13461
#> parameter3  0.99998085 11.170692 -124.78436
#> parameter4  0.99998151 31.494499 -991.90344
#> 
#> ------------------ iteration  3 ------------------
#> Function value  -1384.589 
#> Convergence criteria: parameters stability= 0.3102247 
#>                     : function stability= 830.34 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef   SE.coef   Var.coef
#> parameter1 -0.04536491  4.601141   21.17050
#> parameter2  0.11953703  6.947529   48.26816
#> parameter3  0.64134243 14.223751 -202.31508
#> parameter4  0.59138599 22.660503  513.49841
#> 
#> ------------------ iteration  4 ------------------
#> Function value  -1111.675 
#> Convergence criteria: parameters stability= 1.057287 
#>                     : function stability= 272.9141 
#>                     : relative distance to maximum(RDM)= 77.40388 
#>                  coef    SE.coef   Var.coef
#> parameter1 -0.1229702 0.01534980 0.00023562
#> parameter2  0.2002027 0.00989126 0.00009784
#> parameter3  0.1084268 0.00580726 0.00003372
#> parameter4 -0.2808289 0.00360838 0.00001302
#> 
#> ------------------ iteration  5 ------------------
#> Function value  -956.2987 
#> Convergence criteria: parameters stability= 0.00358795 
#>                     : function stability= 155.3761 
#>                     : relative distance to maximum(RDM)= 29.15557 
#>                  coef    SE.coef   Var.coef
#> parameter1 -0.1705625 0.01636548 0.00026783
#> parameter2  0.2037941 0.01086859 0.00011813
#> parameter3  0.1313634 0.00705489 0.00004977
#> parameter4 -0.3088278 0.00441799 0.00001952
#> 
#> ------------------ iteration  6 ------------------
#> Function value  -889.2709 
#> Convergence criteria: parameters stability= 0.00542335 
#>                     : function stability= 67.0278 
#>                     : relative distance to maximum(RDM)= 5.203307 
#>                  coef    SE.coef   Var.coef
#> parameter1 -0.2265610 0.01824680 0.00033295
#> parameter2  0.2269771 0.01175687 0.00013822
#> parameter3  0.1636821 0.00939243 0.00008822
#> parameter4 -0.3353904 0.00532726 0.00002838
#> 
#> ------------------ iteration  7 ------------------
#> Function value  -877.4981 
#> Convergence criteria: parameters stability= 0.00204746 
#>                     : function stability= 11.77276 
#>                     : relative distance to maximum(RDM)= 0.388275 
#>                  coef    SE.coef   Var.coef
#> parameter1 -0.2575585 0.01981871 0.00039278
#> parameter2  0.2387577 0.01212680 0.00014706
#> parameter3  0.1925249 0.01187135 0.00014093
#> parameter4 -0.3461571 0.00573541 0.00003289
#> 
#> ------------------ iteration  8 ------------------
#> Function value  -876.6699 
#> Convergence criteria: parameters stability= 0.00023298 
#>                     : function stability= 0.8281948 
#>                     : relative distance to maximum(RDM)= 0.00495316 
#>                  coef    SE.coef   Var.coef
#> parameter1 -0.2655036 0.02048768 0.00041975
#> parameter2  0.2400265 0.01217954 0.00014834
#> parameter3  0.2054081 0.01313407 0.00017250
#> parameter4 -0.3476645 0.00579217 0.00003355
#> 
#> ------------------ iteration  9 ------------------
#> Function value  -876.6599 
#> Convergence criteria: parameters stability= 3.59e-06 
#>                     : function stability= 0.00999739 
#>                     : relative distance to maximum(RDM)= 1.01e-06 
#>                  coef    SE.coef   Var.coef
#> parameter1 -0.2662980 0.02057557 0.00042335
#> parameter2  0.2400478 0.01218149 0.00014839
#> parameter3  0.2071285 0.01331116 0.00017719
#> parameter4 -0.3477202 0.00579388 0.00003357
#> 
#> ------------------ iteration  10 ------------------
#> Function value  -876.6599 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 2.03e-06 
#>                     : relative distance to maximum(RDM)= 0 
#>                  coef    SE.coef   Var.coef
#> parameter1 -0.2663087 0.02057687 0.00042341
#> parameter2  0.2400479 0.01218151 0.00014839
#> parameter3  0.2071537 0.01331377 0.00017726
#> parameter4 -0.3477207 0.00579390 0.00003357
#> 
#> The program took 0.17 seconds 
#> ------------------ iteration  0 ------------------
#> Function value  -875.8352 
#> Convergence criteria: parameters stability= 1.0001 
#>                     : function stability= 1.0001 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                  coef    SE.coef     Var.coef
#> parameter1  0.0000000   3.551890    12.615925
#> parameter2 -0.2765972   5.516118    30.427559
#> parameter3 -0.2560203   3.037092     9.223927
#> parameter4  0.2339572  88.466961 -7826.403106
#> parameter5  0.2461387  88.440425 -7821.708779
#> parameter6  0.2071537  74.680812  5577.223668
#> parameter7 -0.3477207 172.636101 29803.223496
#> 
#> ------------------ iteration  1 ------------------
#> Function value  -873.994 
#> Convergence criteria: parameters stability= 5.01e-05 
#>                     : function stability= 1.841189 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef    SE.coef    Var.coef
#> parameter1 -0.00000022   3.180447    10.11524
#> parameter2 -0.27663816   5.543917    30.73501
#> parameter3 -0.25597947   3.588780    12.87934
#> parameter4  0.22895886  85.113657 -7244.33456
#> parameter5  0.25113660  84.963084 -7218.72560
#> parameter6  0.20703717  74.367497  5530.52458
#> parameter7 -0.34737316 172.138076 29631.51721
#> 
#> ------------------ iteration  2 ------------------
#> Function value  -859.0964 
#> Convergence criteria: parameters stability= 0.00069688 
#>                     : function stability= 14.89763 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef   SE.coef   Var.coef
#> parameter1 -0.00029738  10.66240  -113.6868
#> parameter2 -0.27701579  32.56422  1060.4287
#> parameter3 -0.25565151  32.65059  1066.0608
#> parameter4  0.21033159  61.16099 -3740.6665
#> parameter5  0.26980804  58.95450 -3475.6335
#> parameter6  0.20648825  72.69177  5284.0932
#> parameter7 -0.34657255 166.43302 27699.9491
#> 
#> ------------------ iteration  3 ------------------
#> Function value  -844.5946 
#> Convergence criteria: parameters stability= 0.00028433 
#>                     : function stability= 14.50182 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef    SE.coef    Var.coef
#> parameter1 -0.00078499   7.539457   -56.84342
#> parameter2 -0.27629067  33.638259  1131.53250
#> parameter3 -0.25655277  33.852721  1146.00673
#> parameter4  0.19880810  40.073633 -1605.89610
#> parameter5  0.28203338  34.527363 -1192.13878
#> parameter6  0.20626924  71.765445  5150.27909
#> parameter7 -0.34589921 161.720276 26153.44766
#> 
#> ------------------ iteration  4 ------------------
#> Function value  -716.1336 
#> Convergence criteria: parameters stability= 0.01387405 
#>                     : function stability= 128.4609 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef    SE.coef     Var.coef
#> parameter1 -0.00501281   2.460415     6.053643
#> parameter2 -0.26937013   6.133523    37.620109
#> parameter3 -0.26627370   5.978390    35.741152
#> parameter4  0.12455005   6.265785    39.260063
#> parameter5  0.37247275   4.847660    23.499806
#> parameter6  0.20587208   7.767472    60.333627
#> parameter7 -0.34141714 129.548623 16782.845797
#> 
#> ------------------ iteration  5 ------------------
#> Function value  -334.7966 
#> Convergence criteria: parameters stability= 0.086728 
#>                     : function stability= 381.337 
#>                     : relative distance to maximum(RDM)= 157.6855 
#>                   coef    SE.coef   Var.coef
#> parameter1  0.01143070 0.14453836 0.02089134
#> parameter2 -0.13485216 0.03282602 0.00107755
#> parameter3 -0.40648825 0.03667541 0.00134509
#> parameter4 -0.01886364 0.01447015 0.00020939
#> parameter5  0.51773548 0.01843621 0.00033989
#> parameter6  0.24073163 0.01691481 0.00028611
#> parameter7 -0.26513822 0.00768029 0.00005899
#> 
#> ------------------ iteration  6 ------------------
#> Function value  -291.6813 
#> Convergence criteria: parameters stability= 0.3068446 
#>                     : function stability= 43.11529 
#>                     : relative distance to maximum(RDM)= 40.46282 
#>                  coef    SE.coef   Var.coef
#> parameter1  0.2032109 0.14215247 0.02020733
#> parameter2  0.1284274 0.02383865 0.00056828
#> parameter3 -0.7343482 0.02694022 0.00072578
#> parameter4 -0.1703930 0.00755994 0.00005715
#> parameter5  0.7605108 0.00843444 0.00007114
#> parameter6  0.2306920 0.01231619 0.00015169
#> parameter7 -0.1590496 0.00185199 0.00000343
#> 
#> ------------------ iteration  7 ------------------
#> Function value  -123.7402 
#> Convergence criteria: parameters stability= 0.00331586 
#>                     : function stability= 167.9411 
#>                     : relative distance to maximum(RDM)= 7.79339 
#>                  coef    SE.coef   Var.coef
#> parameter1  0.2013452 0.14214392 0.02020489
#> parameter2  0.1261218 0.02369173 0.00056130
#> parameter3 -0.7528443 0.02620270 0.00068658
#> parameter4 -0.1466805 0.00871161 0.00007589
#> parameter5  0.7190565 0.00964421 0.00009301
#> parameter6  0.2236083 0.01169290 0.00013672
#> parameter7 -0.1842296 0.00258201 0.00000667
#> 
#> ------------------ iteration  8 ------------------
#> Function value  -92.77588 
#> Convergence criteria: parameters stability= 0.00118282 
#>                     : function stability= 30.96434 
#>                     : relative distance to maximum(RDM)= 0.49567 
#>                  coef    SE.coef   Var.coef
#> parameter1  0.2005246 0.14215307 0.02020749
#> parameter2  0.1153279 0.02430082 0.00059053
#> parameter3 -0.7339871 0.02689008 0.00072308
#> parameter4 -0.1368313 0.00954859 0.00009118
#> parameter5  0.7018656 0.01056544 0.00011163
#> parameter6  0.2255056 0.01217817 0.00014831
#> parameter7 -0.2019473 0.00321731 0.00001035
#> 
#> ------------------ iteration  9 ------------------
#> Function value  -90.96085 
#> Convergence criteria: parameters stability= 7.884e-05 
#>                     : function stability= 1.815033 
#>                     : relative distance to maximum(RDM)= 0.00272533 
#>                  coef    SE.coef   Var.coef
#> parameter1  0.2002665 0.14216235 0.02021013
#> parameter2  0.1129522 0.02442020 0.00059635
#> parameter3 -0.7297423 0.02703135 0.00073069
#> parameter4 -0.1345810 0.00982520 0.00009653
#> parameter5  0.6978214 0.01087440 0.00011825
#> parameter6  0.2250899 0.01222048 0.00014934
#> parameter7 -0.2077364 0.00345066 0.00001191
#> 
#> ------------------ iteration  10 ------------------
#> Function value  -90.95127 
#> Convergence criteria: parameters stability= 4e-07 
#>                     : function stability= 0.00957939 
#>                     : relative distance to maximum(RDM)= 1.1e-07 
#>                  coef    SE.coef   Var.coef
#> parameter1  0.2002341 0.14216300 0.02021032
#> parameter2  0.1128111 0.02442858 0.00059676
#> parameter3 -0.7294758 0.02704144 0.00073124
#> parameter4 -0.1344466 0.00984756 0.00009697
#> parameter5  0.6975642 0.01089945 0.00011880
#> parameter6  0.2250392 0.01222199 0.00014938
#> parameter7 -0.2082035 0.00347009 0.00001204
#> 
#> ------------------ iteration  11 ------------------
#> Function value  -90.95127 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 3.8e-07 
#>                     : relative distance to maximum(RDM)= 0 
#>                  coef    SE.coef   Var.coef
#> parameter1  0.2002339 0.14216243 0.02021016
#> parameter2  0.1128105 0.02442864 0.00059676
#> parameter3 -0.7294744 0.02704150 0.00073124
#> parameter4 -0.1344460 0.00984770 0.00009698
#> parameter5  0.6975628 0.01089961 0.00011880
#> parameter6  0.2250389 0.01222200 0.00014938
#> parameter7 -0.2082065 0.00347022 0.00001204
#> 
#> The program took 0.35 seconds 
#>              A            B
#> A 9.998034e-01 0.0001965748
#> B 3.070110e-09 0.9999999969