Get the model estimation time

Get the estimation time of the model, determined by the time taken for the associated fit() function to finish.

# S4 method for lcModel
estimationTime(object, unit = "secs", ...)

# S4 method for lcModels
estimationTime(object, unit = "secs", ...)

# S4 method for list
estimationTime(object, unit = "secs", ...)

Arguments

object

The list of lcModel objects.

unit

The time unit in which the estimation time should be outputted. By default, estimation time is in seconds. For accepted units, see base::difftime.

...

Additional arguments.

Value

A numeric representing the model estimation time, in the specified unit.

Examples

data(latrendData)
model <- latrend(lcMethodKML("Y", id = "Id", time = "Time"), latrendData)
#>  ~ Fast KmL ~
#> ********************S
estimationTime(model)
#> [1] 0.445
estimationTime(model, unit = 'mins')
#> [1] 0.007416667
estimationTime(model, unit = 'days')
#> [1] 5.150463e-06