Compute external model metric(s)

Compute one or more external metrics for two or more lcModel objects.

Call getExternalMetricNames() to retrieve the names of the defined internal metrics.

# S4 method for lcModel,lcModel
externalMetric(
  object,
  object2,
  name = getOption("latrend.externalMetric"),
  ...
)

# S4 method for lcModels,missing
externalMetric(object, object2, name = "adjustedRand")

# S4 method for lcModels,character
externalMetric(object, object2 = "adjustedRand")

# S4 method for lcModels,lcModel
externalMetric(object, object2, name, drop = TRUE)

# S4 method for list,lcModel
externalMetric(object, object2, name, drop = TRUE)

Arguments

object

The lcModel, lcModels, or list of lcModel objects to compute the metrics for.

object2

The other lcModel to compare with.

name

The name(s) of the external metric(s) to compute. If no names are given, the names specified in the latrend.externalMetric option (none by default) are used.

...

Additional arguments.

drop

Whether to return a numeric vector instead of a data.frame in case of a single metric.

Value

For externalMetric(lcModel, lcModel): A numeric vector of the computed metrics. A named numeric vector containing the computed model metrics. For externalMetric(lcModels): A distance matrix of class dist representing the pairwise comparisons. For externalMetric(lcModels, name): A distance matrix of class dist representing the pairwise comparisons. For externalMetric(lcModels, lcModel): A named numeric vector or data.framecontaining the computed model metrics. For externalMetric(list, lcModel): A named numeric vector or data.framecontaining the computed model metrics.

References

Desgraupes B (2018). clusterCrit: Clustering Indices. R package version 1.2.8, https://CRAN.R-project.org/package=clusterCrit.

You K (2018). mclustcomp: Measures for Comparing Clusters. R package version 0.3.1, https://CRAN.R-project.org/package=mclustcomp.

Csardi G, Nepusz T (2006). “The igraph software package for complex network research.” InterJournal, Complex Systems, 1695. https://igraph.org.

Hubert L, Arabie P (1985). “Comparing Partitions.” Journal of Classification, 2(1), 193--218. ISSN 1432-1343, doi: 10.1007/BF01908075 .

Revelle W (2019). psych: Procedures for Psychological, Psychometric, and Personality Research. Northwestern University, Evanston, Illinois. R package version 1.9.12, https://CRAN.R-project.org/package=psych.

Scrucca L, Fop M, Murphy TB, Raftery AE (2016). “mclust 5: clustering, classification and density estimation using Gaussian finite mixture models.” The R Journal, 8(1), 205--233. https://journal.r-project.org/archive/2016-1/scrucca-fop-murphy-etal.pdf.

See also

Examples

data(latrendData)
model1 <- latrend(lcMethodKML("Y", id = "Id", time = "Time"), latrendData)
#>  ~ Fast KmL ~
#> ********************S
model2 <- latrend(lcMethodLcmmGMM(fixed = Y ~ Time, mixture = ~ Time,
   id = "Id", time = "Time"), latrendData)
#> Be patient, hlme is running ... 
#> The program took 0.09 seconds 
#> Be patient, hlme is running ... 
#> The program took 0.29 seconds 
ari <- externalMetric(model1, model2, 'adjustedRand')