Extract lcModel fitted values

Returns the cluster-specific fitted values for the given lcModel object. The default implementation calls predict() with newdata = NULL.

# S3 method for lcModel
fitted(object, ..., clusters = trajectoryAssignments(object))

Arguments

object
The lcModel object.

...
Additional arguments.

clusters
Optional cluster assignments per id. If unspecified, a matrix is returned containing the cluster-specific predictions per column.

Value

A numeric vector of the fitted values for the respective class, or a matrix of fitted values for each cluster.

Implementation

Classes extending lcModel can override this method to adapt the computation of the predicted values for the training data. Note that the implementation of this function is only needed when predict() and predictForCluster() are not defined for the lcModel subclass.

fitted.lcModelExt <- function(object, ..., clusters = trajectoryAssignments(object)) {
  pred = predict(object, newdata = NULL)
  transformFitted(pred = pred, model = object, clusters = clusters)
}

The transformFitted() function takes care of transforming the prediction input to the right output format.

See also

Examples

data(latrendData)
model <- latrend(lcMethodKML("Y", id = "Id", time = "Time"), latrendData)

#>  ~ Fast KmL ~
#> ********************S

fitted(model)

#>    [1] -0.98189066 -0.62598349 -0.38037545 -0.16078630  0.06417402  0.18682776
#>    [7]  0.31634026  0.39384185  0.43595169  0.43175688 -0.98189066 -0.62598349
#>   [13] -0.38037545 -0.16078630  0.06417402  0.18682776  0.31634026  0.39384185
#>   [19]  0.43595169  0.43175688 -0.98189066 -0.62598349 -0.38037545 -0.16078630
#>   [25]  0.06417402  0.18682776  0.31634026  0.39384185  0.43595169  0.43175688
#>   [31] -0.98189066 -0.62598349 -0.38037545 -0.16078630  0.06417402  0.18682776
#>   [37]  0.31634026  0.39384185  0.43595169  0.43175688 -0.98189066 -0.62598349
#>   [43] -0.38037545 -0.16078630  0.06417402  0.18682776  0.31634026  0.39384185
#>   [49]  0.43595169  0.43175688 -0.98189066 -0.62598349 -0.38037545 -0.16078630
#>   [55]  0.06417402  0.18682776  0.31634026  0.39384185  0.43595169  0.43175688
#>   [61] -0.98189066 -0.62598349 -0.38037545 -0.16078630  0.06417402  0.18682776
#>   [67]  0.31634026  0.39384185  0.43595169  0.43175688 -0.98189066 -0.62598349
#>   [73] -0.38037545 -0.16078630  0.06417402  0.18682776  0.31634026  0.39384185
#>   [79]  0.43595169  0.43175688 -0.98189066 -0.62598349 -0.38037545 -0.16078630
#>   [85]  0.06417402  0.18682776  0.31634026  0.39384185  0.43595169  0.43175688
#>   [91] -0.98189066 -0.62598349 -0.38037545 -0.16078630  0.06417402  0.18682776
#>   [97]  0.31634026  0.39384185  0.43595169  0.43175688 -0.98189066 -0.62598349
#>  [103] -0.38037545 -0.16078630  0.06417402  0.18682776  0.31634026  0.39384185
#>  [109]  0.43595169  0.43175688 -0.98189066 -0.62598349 -0.38037545 -0.16078630
#>  [115]  0.06417402  0.18682776  0.31634026  0.39384185  0.43595169  0.43175688
#>  [121] -0.98189066 -0.62598349 -0.38037545 -0.16078630  0.06417402  0.18682776
#>  [127]  0.31634026  0.39384185  0.43595169  0.43175688 -0.98189066 -0.62598349
#>  [133] -0.38037545 -0.16078630  0.06417402  0.18682776  0.31634026  0.39384185
#>  [139]  0.43595169  0.43175688 -0.98189066 -0.62598349 -0.38037545 -0.16078630
#>  [145]  0.06417402  0.18682776  0.31634026  0.39384185  0.43595169  0.43175688
#>  [151] -0.98189066 -0.62598349 -0.38037545 -0.16078630  0.06417402  0.18682776
#>  [157]  0.31634026  0.39384185  0.43595169  0.43175688 -0.98189066 -0.62598349
#>  [163] -0.38037545 -0.16078630  0.06417402  0.18682776  0.31634026  0.39384185
#>  [169]  0.43595169  0.43175688 -0.98189066 -0.62598349 -0.38037545 -0.16078630
#>  [175]  0.06417402  0.18682776  0.31634026  0.39384185  0.43595169  0.43175688
#>  [181] -0.98189066 -0.62598349 -0.38037545 -0.16078630  0.06417402  0.18682776
#>  [187]  0.31634026  0.39384185  0.43595169  0.43175688 -0.98189066 -0.62598349
#>  [193] -0.38037545 -0.16078630  0.06417402  0.18682776  0.31634026  0.39384185
#>  [199]  0.43595169  0.43175688 -0.98189066 -0.62598349 -0.38037545 -0.16078630
#>  [205]  0.06417402  0.18682776  0.31634026  0.39384185  0.43595169  0.43175688
#>  [211] -0.98189066 -0.62598349 -0.38037545 -0.16078630  0.06417402  0.18682776
#>  [217]  0.31634026  0.39384185  0.43595169  0.43175688 -0.98189066 -0.62598349
#>  [223] -0.38037545 -0.16078630  0.06417402  0.18682776  0.31634026  0.39384185
#>  [229]  0.43595169  0.43175688 -0.98189066 -0.62598349 -0.38037545 -0.16078630
#>  [235]  0.06417402  0.18682776  0.31634026  0.39384185  0.43595169  0.43175688
#>  [241] -0.98189066 -0.62598349 -0.38037545 -0.16078630  0.06417402  0.18682776
#>  [247]  0.31634026  0.39384185  0.43595169  0.43175688 -0.98189066 -0.62598349
#>  [253] -0.38037545 -0.16078630  0.06417402  0.18682776  0.31634026  0.39384185
#>  [259]  0.43595169  0.43175688 -0.98189066 -0.62598349 -0.38037545 -0.16078630
#>  [265]  0.06417402  0.18682776  0.31634026  0.39384185  0.43595169  0.43175688
#>  [271] -0.98189066 -0.62598349 -0.38037545 -0.16078630  0.06417402  0.18682776
#>  [277]  0.31634026  0.39384185  0.43595169  0.43175688 -0.98189066 -0.62598349
#>  [283] -0.38037545 -0.16078630  0.06417402  0.18682776  0.31634026  0.39384185
#>  [289]  0.43595169  0.43175688 -0.98189066 -0.62598349 -0.38037545 -0.16078630
#>  [295]  0.06417402  0.18682776  0.31634026  0.39384185  0.43595169  0.43175688
#>  [301] -0.98189066 -0.62598349 -0.38037545 -0.16078630  0.06417402  0.18682776
#>  [307]  0.31634026  0.39384185  0.43595169  0.43175688 -0.98189066 -0.62598349
#>  [313] -0.38037545 -0.16078630  0.06417402  0.18682776  0.31634026  0.39384185
#>  [319]  0.43595169  0.43175688 -0.98189066 -0.62598349 -0.38037545 -0.16078630
#>  [325]  0.06417402  0.18682776  0.31634026  0.39384185  0.43595169  0.43175688
#>  [331] -0.98189066 -0.62598349 -0.38037545 -0.16078630  0.06417402  0.18682776
#>  [337]  0.31634026  0.39384185  0.43595169  0.43175688 -0.98189066 -0.62598349
#>  [343] -0.38037545 -0.16078630  0.06417402  0.18682776  0.31634026  0.39384185
#>  [349]  0.43595169  0.43175688 -0.98189066 -0.62598349 -0.38037545 -0.16078630
#>  [355]  0.06417402  0.18682776  0.31634026  0.39384185  0.43595169  0.43175688
#>  [361] -0.98189066 -0.62598349 -0.38037545 -0.16078630  0.06417402  0.18682776
#>  [367]  0.31634026  0.39384185  0.43595169  0.43175688 -0.98189066 -0.62598349
#>  [373] -0.38037545 -0.16078630  0.06417402  0.18682776  0.31634026  0.39384185
#>  [379]  0.43595169  0.43175688 -0.98189066 -0.62598349 -0.38037545 -0.16078630
#>  [385]  0.06417402  0.18682776  0.31634026  0.39384185  0.43595169  0.43175688
#>  [391] -0.98189066 -0.62598349 -0.38037545 -0.16078630  0.06417402  0.18682776
#>  [397]  0.31634026  0.39384185  0.43595169  0.43175688 -0.98189066 -0.62598349
#>  [403] -0.38037545 -0.16078630  0.06417402  0.18682776  0.31634026  0.39384185
#>  [409]  0.43595169  0.43175688 -0.98189066 -0.62598349 -0.38037545 -0.16078630
#>  [415]  0.06417402  0.18682776  0.31634026  0.39384185  0.43595169  0.43175688
#>  [421] -0.98189066 -0.62598349 -0.38037545 -0.16078630  0.06417402  0.18682776
#>  [427]  0.31634026  0.39384185  0.43595169  0.43175688 -0.98189066 -0.62598349
#>  [433] -0.38037545 -0.16078630  0.06417402  0.18682776  0.31634026  0.39384185
#>  [439]  0.43595169  0.43175688 -0.98189066 -0.62598349 -0.38037545 -0.16078630
#>  [445]  0.06417402  0.18682776  0.31634026  0.39384185  0.43595169  0.43175688
#>  [451] -0.98189066 -0.62598349 -0.38037545 -0.16078630  0.06417402  0.18682776
#>  [457]  0.31634026  0.39384185  0.43595169  0.43175688 -0.98189066 -0.62598349
#>  [463] -0.38037545 -0.16078630  0.06417402  0.18682776  0.31634026  0.39384185
#>  [469]  0.43595169  0.43175688 -0.98189066 -0.62598349 -0.38037545 -0.16078630
#>  [475]  0.06417402  0.18682776  0.31634026  0.39384185  0.43595169  0.43175688
#>  [481] -0.98189066 -0.62598349 -0.38037545 -0.16078630  0.06417402  0.18682776
#>  [487]  0.31634026  0.39384185  0.43595169  0.43175688 -0.98189066 -0.62598349
#>  [493] -0.38037545 -0.16078630  0.06417402  0.18682776  0.31634026  0.39384185
#>  [499]  0.43595169  0.43175688 -0.98189066 -0.62598349 -0.38037545 -0.16078630
#>  [505]  0.06417402  0.18682776  0.31634026  0.39384185  0.43595169  0.43175688
#>  [511] -0.98189066 -0.62598349 -0.38037545 -0.16078630  0.06417402  0.18682776
#>  [517]  0.31634026  0.39384185  0.43595169  0.43175688 -0.98189066 -0.62598349
#>  [523] -0.38037545 -0.16078630  0.06417402  0.18682776  0.31634026  0.39384185
#>  [529]  0.43595169  0.43175688 -0.98189066 -0.62598349 -0.38037545 -0.16078630
#>  [535]  0.06417402  0.18682776  0.31634026  0.39384185  0.43595169  0.43175688
#>  [541] -0.98189066 -0.62598349 -0.38037545 -0.16078630  0.06417402  0.18682776
#>  [547]  0.31634026  0.39384185  0.43595169  0.43175688 -0.98189066 -0.62598349
#>  [553] -0.38037545 -0.16078630  0.06417402  0.18682776  0.31634026  0.39384185
#>  [559]  0.43595169  0.43175688 -0.98189066 -0.62598349 -0.38037545 -0.16078630
#>  [565]  0.06417402  0.18682776  0.31634026  0.39384185  0.43595169  0.43175688
#>  [571] -0.98189066 -0.62598349 -0.38037545 -0.16078630  0.06417402  0.18682776
#>  [577]  0.31634026  0.39384185  0.43595169  0.43175688 -0.98189066 -0.62598349
#>  [583] -0.38037545 -0.16078630  0.06417402  0.18682776  0.31634026  0.39384185
#>  [589]  0.43595169  0.43175688 -0.98189066 -0.62598349 -0.38037545 -0.16078630
#>  [595]  0.06417402  0.18682776  0.31634026  0.39384185  0.43595169  0.43175688
#>  [601] -0.98189066 -0.62598349 -0.38037545 -0.16078630  0.06417402  0.18682776
#>  [607]  0.31634026  0.39384185  0.43595169  0.43175688 -0.98189066 -0.62598349
#>  [613] -0.38037545 -0.16078630  0.06417402  0.18682776  0.31634026  0.39384185
#>  [619]  0.43595169  0.43175688 -0.98189066 -0.62598349 -0.38037545 -0.16078630
#>  [625]  0.06417402  0.18682776  0.31634026  0.39384185  0.43595169  0.43175688
#>  [631] -0.98189066 -0.62598349 -0.38037545 -0.16078630  0.06417402  0.18682776
#>  [637]  0.31634026  0.39384185  0.43595169  0.43175688 -0.98189066 -0.62598349
#>  [643] -0.38037545 -0.16078630  0.06417402  0.18682776  0.31634026  0.39384185
#>  [649]  0.43595169  0.43175688 -0.98189066 -0.62598349 -0.38037545 -0.16078630
#>  [655]  0.06417402  0.18682776  0.31634026  0.39384185  0.43595169  0.43175688
#>  [661] -0.98189066 -0.62598349 -0.38037545 -0.16078630  0.06417402  0.18682776
#>  [667]  0.31634026  0.39384185  0.43595169  0.43175688 -0.98189066 -0.62598349
#>  [673] -0.38037545 -0.16078630  0.06417402  0.18682776  0.31634026  0.39384185
#>  [679]  0.43595169  0.43175688 -0.98189066 -0.62598349 -0.38037545 -0.16078630
#>  [685]  0.06417402  0.18682776  0.31634026  0.39384185  0.43595169  0.43175688
#>  [691] -0.98189066 -0.62598349 -0.38037545 -0.16078630  0.06417402  0.18682776
#>  [697]  0.31634026  0.39384185  0.43595169  0.43175688 -0.98189066 -0.62598349
#>  [703] -0.38037545 -0.16078630  0.06417402  0.18682776  0.31634026  0.39384185
#>  [709]  0.43595169  0.43175688 -0.98189066 -0.62598349 -0.38037545 -0.16078630
#>  [715]  0.06417402  0.18682776  0.31634026  0.39384185  0.43595169  0.43175688
#>  [721] -0.98189066 -0.62598349 -0.38037545 -0.16078630  0.06417402  0.18682776
#>  [727]  0.31634026  0.39384185  0.43595169  0.43175688 -0.98189066 -0.62598349
#>  [733] -0.38037545 -0.16078630  0.06417402  0.18682776  0.31634026  0.39384185
#>  [739]  0.43595169  0.43175688 -0.98189066 -0.62598349 -0.38037545 -0.16078630
#>  [745]  0.06417402  0.18682776  0.31634026  0.39384185  0.43595169  0.43175688
#>  [751] -0.98189066 -0.62598349 -0.38037545 -0.16078630  0.06417402  0.18682776
#>  [757]  0.31634026  0.39384185  0.43595169  0.43175688 -0.98189066 -0.62598349
#>  [763] -0.38037545 -0.16078630  0.06417402  0.18682776  0.31634026  0.39384185
#>  [769]  0.43595169  0.43175688 -0.98189066 -0.62598349 -0.38037545 -0.16078630
#>  [775]  0.06417402  0.18682776  0.31634026  0.39384185  0.43595169  0.43175688
#>  [781] -0.98189066 -0.62598349 -0.38037545 -0.16078630  0.06417402  0.18682776
#>  [787]  0.31634026  0.39384185  0.43595169  0.43175688 -0.98189066 -0.62598349
#>  [793] -0.38037545 -0.16078630  0.06417402  0.18682776  0.31634026  0.39384185
#>  [799]  0.43595169  0.43175688 -0.98189066 -0.62598349 -0.38037545 -0.16078630
#>  [805]  0.06417402  0.18682776  0.31634026  0.39384185  0.43595169  0.43175688
#>  [811] -0.98189066 -0.62598349 -0.38037545 -0.16078630  0.06417402  0.18682776
#>  [817]  0.31634026  0.39384185  0.43595169  0.43175688 -0.98189066 -0.62598349
#>  [823] -0.38037545 -0.16078630  0.06417402  0.18682776  0.31634026  0.39384185
#>  [829]  0.43595169  0.43175688 -0.98189066 -0.62598349 -0.38037545 -0.16078630
#>  [835]  0.06417402  0.18682776  0.31634026  0.39384185  0.43595169  0.43175688
#>  [841] -0.98189066 -0.62598349 -0.38037545 -0.16078630  0.06417402  0.18682776
#>  [847]  0.31634026  0.39384185  0.43595169  0.43175688 -0.98189066 -0.62598349
#>  [853] -0.38037545 -0.16078630  0.06417402  0.18682776  0.31634026  0.39384185
#>  [859]  0.43595169  0.43175688 -0.98189066 -0.62598349 -0.38037545 -0.16078630
#>  [865]  0.06417402  0.18682776  0.31634026  0.39384185  0.43595169  0.43175688
#>  [871] -0.98189066 -0.62598349 -0.38037545 -0.16078630  0.06417402  0.18682776
#>  [877]  0.31634026  0.39384185  0.43595169  0.43175688 -0.98189066 -0.62598349
#>  [883] -0.38037545 -0.16078630  0.06417402  0.18682776  0.31634026  0.39384185
#>  [889]  0.43595169  0.43175688 -0.98189066 -0.62598349 -0.38037545 -0.16078630
#>  [895]  0.06417402  0.18682776  0.31634026  0.39384185  0.43595169  0.43175688
#>  [901]  0.04839039  0.05189640  0.06509638  0.06698159  0.03735418  0.01209756
#>  [907] -0.03916995 -0.08030881 -0.17162641 -0.20624529  0.04839039  0.05189640
#>  [913]  0.06509638  0.06698159  0.03735418  0.01209756 -0.03916995 -0.08030881
#>  [919] -0.17162641 -0.20624529  0.04839039  0.05189640  0.06509638  0.06698159
#>  [925]  0.03735418  0.01209756 -0.03916995 -0.08030881 -0.17162641 -0.20624529
#>  [931]  0.04839039  0.05189640  0.06509638  0.06698159  0.03735418  0.01209756
#>  [937] -0.03916995 -0.08030881 -0.17162641 -0.20624529  0.04839039  0.05189640
#>  [943]  0.06509638  0.06698159  0.03735418  0.01209756 -0.03916995 -0.08030881
#>  [949] -0.17162641 -0.20624529  0.04839039  0.05189640  0.06509638  0.06698159
#>  [955]  0.03735418  0.01209756 -0.03916995 -0.08030881 -0.17162641 -0.20624529
#>  [961]  0.04839039  0.05189640  0.06509638  0.06698159  0.03735418  0.01209756
#>  [967] -0.03916995 -0.08030881 -0.17162641 -0.20624529  0.04839039  0.05189640
#>  [973]  0.06509638  0.06698159  0.03735418  0.01209756 -0.03916995 -0.08030881
#>  [979] -0.17162641 -0.20624529  0.04839039  0.05189640  0.06509638  0.06698159
#>  [985]  0.03735418  0.01209756 -0.03916995 -0.08030881 -0.17162641 -0.20624529
#>  [991]  0.04839039  0.05189640  0.06509638  0.06698159  0.03735418  0.01209756
#>  [997] -0.03916995 -0.08030881 -0.17162641 -0.20624529  0.04839039  0.05189640
#> [1003]  0.06509638  0.06698159  0.03735418  0.01209756 -0.03916995 -0.08030881
#> [1009] -0.17162641 -0.20624529  0.04839039  0.05189640  0.06509638  0.06698159
#> [1015]  0.03735418  0.01209756 -0.03916995 -0.08030881 -0.17162641 -0.20624529
#> [1021]  0.04839039  0.05189640  0.06509638  0.06698159  0.03735418  0.01209756
#> [1027] -0.03916995 -0.08030881 -0.17162641 -0.20624529  0.04839039  0.05189640
#> [1033]  0.06509638  0.06698159  0.03735418  0.01209756 -0.03916995 -0.08030881
#> [1039] -0.17162641 -0.20624529  0.04839039  0.05189640  0.06509638  0.06698159
#> [1045]  0.03735418  0.01209756 -0.03916995 -0.08030881 -0.17162641 -0.20624529
#> [1051]  0.04839039  0.05189640  0.06509638  0.06698159  0.03735418  0.01209756
#> [1057] -0.03916995 -0.08030881 -0.17162641 -0.20624529  0.04839039  0.05189640
#> [1063]  0.06509638  0.06698159  0.03735418  0.01209756 -0.03916995 -0.08030881
#> [1069] -0.17162641 -0.20624529  0.04839039  0.05189640  0.06509638  0.06698159
#> [1075]  0.03735418  0.01209756 -0.03916995 -0.08030881 -0.17162641 -0.20624529
#> [1081]  0.04839039  0.05189640  0.06509638  0.06698159  0.03735418  0.01209756
#> [1087] -0.03916995 -0.08030881 -0.17162641 -0.20624529  0.04839039  0.05189640
#> [1093]  0.06509638  0.06698159  0.03735418  0.01209756 -0.03916995 -0.08030881
#> [1099] -0.17162641 -0.20624529  0.04839039  0.05189640  0.06509638  0.06698159
#> [1105]  0.03735418  0.01209756 -0.03916995 -0.08030881 -0.17162641 -0.20624529
#> [1111]  0.04839039  0.05189640  0.06509638  0.06698159  0.03735418  0.01209756
#> [1117] -0.03916995 -0.08030881 -0.17162641 -0.20624529  0.04839039  0.05189640
#> [1123]  0.06509638  0.06698159  0.03735418  0.01209756 -0.03916995 -0.08030881
#> [1129] -0.17162641 -0.20624529  0.04839039  0.05189640  0.06509638  0.06698159
#> [1135]  0.03735418  0.01209756 -0.03916995 -0.08030881 -0.17162641 -0.20624529
#> [1141]  0.04839039  0.05189640  0.06509638  0.06698159  0.03735418  0.01209756
#> [1147] -0.03916995 -0.08030881 -0.17162641 -0.20624529  0.04839039  0.05189640
#> [1153]  0.06509638  0.06698159  0.03735418  0.01209756 -0.03916995 -0.08030881
#> [1159] -0.17162641 -0.20624529  0.04839039  0.05189640  0.06509638  0.06698159
#> [1165]  0.03735418  0.01209756 -0.03916995 -0.08030881 -0.17162641 -0.20624529
#> [1171]  0.04839039  0.05189640  0.06509638  0.06698159  0.03735418  0.01209756
#> [1177] -0.03916995 -0.08030881 -0.17162641 -0.20624529  0.04839039  0.05189640
#> [1183]  0.06509638  0.06698159  0.03735418  0.01209756 -0.03916995 -0.08030881
#> [1189] -0.17162641 -0.20624529  0.04839039  0.05189640  0.06509638  0.06698159
#> [1195]  0.03735418  0.01209756 -0.03916995 -0.08030881 -0.17162641 -0.20624529
#> [1201]  0.04839039  0.05189640  0.06509638  0.06698159  0.03735418  0.01209756
#> [1207] -0.03916995 -0.08030881 -0.17162641 -0.20624529  0.04839039  0.05189640
#> [1213]  0.06509638  0.06698159  0.03735418  0.01209756 -0.03916995 -0.08030881
#> [1219] -0.17162641 -0.20624529  0.04839039  0.05189640  0.06509638  0.06698159
#> [1225]  0.03735418  0.01209756 -0.03916995 -0.08030881 -0.17162641 -0.20624529
#> [1231]  0.04839039  0.05189640  0.06509638  0.06698159  0.03735418  0.01209756
#> [1237] -0.03916995 -0.08030881 -0.17162641 -0.20624529  0.04839039  0.05189640
#> [1243]  0.06509638  0.06698159  0.03735418  0.01209756 -0.03916995 -0.08030881
#> [1249] -0.17162641 -0.20624529  0.04839039  0.05189640  0.06509638  0.06698159
#> [1255]  0.03735418  0.01209756 -0.03916995 -0.08030881 -0.17162641 -0.20624529
#> [1261]  0.04839039  0.05189640  0.06509638  0.06698159  0.03735418  0.01209756
#> [1267] -0.03916995 -0.08030881 -0.17162641 -0.20624529  0.04839039  0.05189640
#> [1273]  0.06509638  0.06698159  0.03735418  0.01209756 -0.03916995 -0.08030881
#> [1279] -0.17162641 -0.20624529  0.04839039  0.05189640  0.06509638  0.06698159
#> [1285]  0.03735418  0.01209756 -0.03916995 -0.08030881 -0.17162641 -0.20624529
#> [1291]  0.04839039  0.05189640  0.06509638  0.06698159  0.03735418  0.01209756
#> [1297] -0.03916995 -0.08030881 -0.17162641 -0.20624529  0.04839039  0.05189640
#> [1303]  0.06509638  0.06698159  0.03735418  0.01209756 -0.03916995 -0.08030881
#> [1309] -0.17162641 -0.20624529  0.04839039  0.05189640  0.06509638  0.06698159
#> [1315]  0.03735418  0.01209756 -0.03916995 -0.08030881 -0.17162641 -0.20624529
#> [1321]  0.04839039  0.05189640  0.06509638  0.06698159  0.03735418  0.01209756
#> [1327] -0.03916995 -0.08030881 -0.17162641 -0.20624529  0.04839039  0.05189640
#> [1333]  0.06509638  0.06698159  0.03735418  0.01209756 -0.03916995 -0.08030881
#> [1339] -0.17162641 -0.20624529  0.04839039  0.05189640  0.06509638  0.06698159
#> [1345]  0.03735418  0.01209756 -0.03916995 -0.08030881 -0.17162641 -0.20624529
#> [1351]  0.04839039  0.05189640  0.06509638  0.06698159  0.03735418  0.01209756
#> [1357] -0.03916995 -0.08030881 -0.17162641 -0.20624529  0.04839039  0.05189640
#> [1363]  0.06509638  0.06698159  0.03735418  0.01209756 -0.03916995 -0.08030881
#> [1369] -0.17162641 -0.20624529  0.04839039  0.05189640  0.06509638  0.06698159
#> [1375]  0.03735418  0.01209756 -0.03916995 -0.08030881 -0.17162641 -0.20624529
#> [1381]  0.04839039  0.05189640  0.06509638  0.06698159  0.03735418  0.01209756
#> [1387] -0.03916995 -0.08030881 -0.17162641 -0.20624529  0.04839039  0.05189640
#> [1393]  0.06509638  0.06698159  0.03735418  0.01209756 -0.03916995 -0.08030881
#> [1399] -0.17162641 -0.20624529  0.04839039  0.05189640  0.06509638  0.06698159
#> [1405]  0.03735418  0.01209756 -0.03916995 -0.08030881 -0.17162641 -0.20624529
#> [1411]  0.04839039  0.05189640  0.06509638  0.06698159  0.03735418  0.01209756
#> [1417] -0.03916995 -0.08030881 -0.17162641 -0.20624529  0.04839039  0.05189640
#> [1423]  0.06509638  0.06698159  0.03735418  0.01209756 -0.03916995 -0.08030881
#> [1429] -0.17162641 -0.20624529  0.04839039  0.05189640  0.06509638  0.06698159
#> [1435]  0.03735418  0.01209756 -0.03916995 -0.08030881 -0.17162641 -0.20624529
#> [1441]  0.04839039  0.05189640  0.06509638  0.06698159  0.03735418  0.01209756
#> [1447] -0.03916995 -0.08030881 -0.17162641 -0.20624529  0.04839039  0.05189640
#> [1453]  0.06509638  0.06698159  0.03735418  0.01209756 -0.03916995 -0.08030881
#> [1459] -0.17162641 -0.20624529  0.04839039  0.05189640  0.06509638  0.06698159
#> [1465]  0.03735418  0.01209756 -0.03916995 -0.08030881 -0.17162641 -0.20624529
#> [1471]  0.04839039  0.05189640  0.06509638  0.06698159  0.03735418  0.01209756
#> [1477] -0.03916995 -0.08030881 -0.17162641 -0.20624529  0.04839039  0.05189640
#> [1483]  0.06509638  0.06698159  0.03735418  0.01209756 -0.03916995 -0.08030881
#> [1489] -0.17162641 -0.20624529  0.04839039  0.05189640  0.06509638  0.06698159
#> [1495]  0.03735418  0.01209756 -0.03916995 -0.08030881 -0.17162641 -0.20624529
#> [1501]  0.04839039  0.05189640  0.06509638  0.06698159  0.03735418  0.01209756
#> [1507] -0.03916995 -0.08030881 -0.17162641 -0.20624529  0.04839039  0.05189640
#> [1513]  0.06509638  0.06698159  0.03735418  0.01209756 -0.03916995 -0.08030881
#> [1519] -0.17162641 -0.20624529  0.04839039  0.05189640  0.06509638  0.06698159
#> [1525]  0.03735418  0.01209756 -0.03916995 -0.08030881 -0.17162641 -0.20624529
#> [1531]  0.04839039  0.05189640  0.06509638  0.06698159  0.03735418  0.01209756
#> [1537] -0.03916995 -0.08030881 -0.17162641 -0.20624529  0.04839039  0.05189640
#> [1543]  0.06509638  0.06698159  0.03735418  0.01209756 -0.03916995 -0.08030881
#> [1549] -0.17162641 -0.20624529  0.04839039  0.05189640  0.06509638  0.06698159
#> [1555]  0.03735418  0.01209756 -0.03916995 -0.08030881 -0.17162641 -0.20624529
#> [1561]  0.04839039  0.05189640  0.06509638  0.06698159  0.03735418  0.01209756
#> [1567] -0.03916995 -0.08030881 -0.17162641 -0.20624529  0.04839039  0.05189640
#> [1573]  0.06509638  0.06698159  0.03735418  0.01209756 -0.03916995 -0.08030881
#> [1579] -0.17162641 -0.20624529  0.04839039  0.05189640  0.06509638  0.06698159
#> [1585]  0.03735418  0.01209756 -0.03916995 -0.08030881 -0.17162641 -0.20624529
#> [1591]  0.04839039  0.05189640  0.06509638  0.06698159  0.03735418  0.01209756
#> [1597] -0.03916995 -0.08030881 -0.17162641 -0.20624529  0.04839039  0.05189640
#> [1603]  0.06509638  0.06698159  0.03735418  0.01209756 -0.03916995 -0.08030881
#> [1609] -0.17162641 -0.20624529  0.04839039  0.05189640  0.06509638  0.06698159
#> [1615]  0.03735418  0.01209756 -0.03916995 -0.08030881 -0.17162641 -0.20624529
#> [1621]  0.04839039  0.05189640  0.06509638  0.06698159  0.03735418  0.01209756
#> [1627] -0.03916995 -0.08030881 -0.17162641 -0.20624529  0.04839039  0.05189640
#> [1633]  0.06509638  0.06698159  0.03735418  0.01209756 -0.03916995 -0.08030881
#> [1639] -0.17162641 -0.20624529  0.04839039  0.05189640  0.06509638  0.06698159
#> [1645]  0.03735418  0.01209756 -0.03916995 -0.08030881 -0.17162641 -0.20624529
#> [1651]  0.04839039  0.05189640  0.06509638  0.06698159  0.03735418  0.01209756
#> [1657] -0.03916995 -0.08030881 -0.17162641 -0.20624529  0.04839039  0.05189640
#> [1663]  0.06509638  0.06698159  0.03735418  0.01209756 -0.03916995 -0.08030881
#> [1669] -0.17162641 -0.20624529  0.04839039  0.05189640  0.06509638  0.06698159
#> [1675]  0.03735418  0.01209756 -0.03916995 -0.08030881 -0.17162641 -0.20624529
#> [1681]  0.04839039  0.05189640  0.06509638  0.06698159  0.03735418  0.01209756
#> [1687] -0.03916995 -0.08030881 -0.17162641 -0.20624529  0.04839039  0.05189640
#> [1693]  0.06509638  0.06698159  0.03735418  0.01209756 -0.03916995 -0.08030881
#> [1699] -0.17162641 -0.20624529  0.04839039  0.05189640  0.06509638  0.06698159
#> [1705]  0.03735418  0.01209756 -0.03916995 -0.08030881 -0.17162641 -0.20624529
#> [1711]  0.04839039  0.05189640  0.06509638  0.06698159  0.03735418  0.01209756
#> [1717] -0.03916995 -0.08030881 -0.17162641 -0.20624529  0.04839039  0.05189640
#> [1723]  0.06509638  0.06698159  0.03735418  0.01209756 -0.03916995 -0.08030881
#> [1729] -0.17162641 -0.20624529  0.04839039  0.05189640  0.06509638  0.06698159
#> [1735]  0.03735418  0.01209756 -0.03916995 -0.08030881 -0.17162641 -0.20624529
#> [1741]  0.04839039  0.05189640  0.06509638  0.06698159  0.03735418  0.01209756
#> [1747] -0.03916995 -0.08030881 -0.17162641 -0.20624529  0.04839039  0.05189640
#> [1753]  0.06509638  0.06698159  0.03735418  0.01209756 -0.03916995 -0.08030881
#> [1759] -0.17162641 -0.20624529  0.04839039  0.05189640  0.06509638  0.06698159
#> [1765]  0.03735418  0.01209756 -0.03916995 -0.08030881 -0.17162641 -0.20624529
#> [1771]  0.04839039  0.05189640  0.06509638  0.06698159  0.03735418  0.01209756
#> [1777] -0.03916995 -0.08030881 -0.17162641 -0.20624529  0.04839039  0.05189640
#> [1783]  0.06509638  0.06698159  0.03735418  0.01209756 -0.03916995 -0.08030881
#> [1789] -0.17162641 -0.20624529  0.04839039  0.05189640  0.06509638  0.06698159
#> [1795]  0.03735418  0.01209756 -0.03916995 -0.08030881 -0.17162641 -0.20624529
#> [1801]  0.04839039  0.05189640  0.06509638  0.06698159  0.03735418  0.01209756
#> [1807] -0.03916995 -0.08030881 -0.17162641 -0.20624529  0.04839039  0.05189640
#> [1813]  0.06509638  0.06698159  0.03735418  0.01209756 -0.03916995 -0.08030881
#> [1819] -0.17162641 -0.20624529  0.04839039  0.05189640  0.06509638  0.06698159
#> [1825]  0.03735418  0.01209756 -0.03916995 -0.08030881 -0.17162641 -0.20624529
#> [1831]  0.04839039  0.05189640  0.06509638  0.06698159  0.03735418  0.01209756
#> [1837] -0.03916995 -0.08030881 -0.17162641 -0.20624529  0.04839039  0.05189640
#> [1843]  0.06509638  0.06698159  0.03735418  0.01209756 -0.03916995 -0.08030881
#> [1849] -0.17162641 -0.20624529  0.04839039  0.05189640  0.06509638  0.06698159
#> [1855]  0.03735418  0.01209756 -0.03916995 -0.08030881 -0.17162641 -0.20624529
#> [1861]  0.04839039  0.05189640  0.06509638  0.06698159  0.03735418  0.01209756
#> [1867] -0.03916995 -0.08030881 -0.17162641 -0.20624529  0.04839039  0.05189640
#> [1873]  0.06509638  0.06698159  0.03735418  0.01209756 -0.03916995 -0.08030881
#> [1879] -0.17162641 -0.20624529  0.04839039  0.05189640  0.06509638  0.06698159
#> [1885]  0.03735418  0.01209756 -0.03916995 -0.08030881 -0.17162641 -0.20624529
#> [1891]  0.04839039  0.05189640  0.06509638  0.06698159  0.03735418  0.01209756
#> [1897] -0.03916995 -0.08030881 -0.17162641 -0.20624529  0.04839039  0.05189640
#> [1903]  0.06509638  0.06698159  0.03735418  0.01209756 -0.03916995 -0.08030881
#> [1909] -0.17162641 -0.20624529  0.04839039  0.05189640  0.06509638  0.06698159
#> [1915]  0.03735418  0.01209756 -0.03916995 -0.08030881 -0.17162641 -0.20624529
#> [1921]  0.04839039  0.05189640  0.06509638  0.06698159  0.03735418  0.01209756
#> [1927] -0.03916995 -0.08030881 -0.17162641 -0.20624529  0.04839039  0.05189640
#> [1933]  0.06509638  0.06698159  0.03735418  0.01209756 -0.03916995 -0.08030881
#> [1939] -0.17162641 -0.20624529  0.04839039  0.05189640  0.06509638  0.06698159
#> [1945]  0.03735418  0.01209756 -0.03916995 -0.08030881 -0.17162641 -0.20624529
#> [1951]  0.04839039  0.05189640  0.06509638  0.06698159  0.03735418  0.01209756
#> [1957] -0.03916995 -0.08030881 -0.17162641 -0.20624529  0.04839039  0.05189640
#> [1963]  0.06509638  0.06698159  0.03735418  0.01209756 -0.03916995 -0.08030881
#> [1969] -0.17162641 -0.20624529  0.04839039  0.05189640  0.06509638  0.06698159
#> [1975]  0.03735418  0.01209756 -0.03916995 -0.08030881 -0.17162641 -0.20624529
#> [1981]  0.04839039  0.05189640  0.06509638  0.06698159  0.03735418  0.01209756
#> [1987] -0.03916995 -0.08030881 -0.17162641 -0.20624529  0.04839039  0.05189640
#> [1993]  0.06509638  0.06698159  0.03735418  0.01209756 -0.03916995 -0.08030881
#> [1999] -0.17162641 -0.20624529