Extract the fitted trajectories for all strata

Extract the fitted trajectories for all strata

# S4 method for lcModel
fittedTrajectories(
  object,
  at = time(object),
  what = "mu",
  clusters = trajectoryAssignments(object),
  ...
)

Arguments

object
The model.

at
The time points at which to compute the id-specific trajectories.

what
The distributional parameter to compute the response for.

clusters
The cluster assignments for the strata to base the trajectories on.

...
Additional arguments.

Value

A data.frame representing the fitted response per trajectory per moment in time.

See also

Examples

data(latrendData)
m <- lcMethodKML("Y", id = "Id", time = "Time")
model <- latrend(method = m, data = latrendData)
#>  ~ Fast KmL ~
#> ********************S
fittedTrajectories(model)
#>        Id Cluster      Time           Y
#>    1:   1       B 0.0000000 -0.98189066
#>    2:   1       B 0.2222222 -0.62598349
#>    3:   1       B 0.4444444 -0.38037545
#>    4:   1       B 0.6666667 -0.16078630
#>    5:   1       B 0.8888889  0.06417402
#>   ---                                  
#> 1996: 200       A 1.1111111  0.01209756
#> 1997: 200       A 1.3333333 -0.03916995
#> 1998: 200       A 1.5555556 -0.08030881
#> 1999: 200       A 1.7777778 -0.17162641
#> 2000: 200       A 2.0000000 -0.20624529

fittedTrajectories(model, at = c(0, .5, 1))
#>       Id Cluster Time           Y
#>   1:   1       B  0.0 -0.98189066
#>   2:   1       B  0.5 -0.32547817
#>   3:   1       B  1.0  0.12550089
#>   4:   2       B  0.0 -0.98189066
#>   5:   2       B  0.5 -0.32547817
#>  ---                             
#> 596: 199       A  0.5  0.06556768
#> 597: 199       A  1.0  0.02472587
#> 598: 200       A  0.0  0.04839039
#> 599: 200       A  0.5  0.06556768
#> 600: 200       A  1.0  0.02472587