custom interface

custom interface

# S4 method for lcMethodCustom
getArgumentDefaults(object)

# S4 method for lcMethodCustom
getName(object)

# S4 method for lcMethodCustom
getShortName(object)

# S4 method for lcMethodCustom
prepareData(method, data, verbose)

# S4 method for lcMethodCustom
fit(method, data, envir, verbose)

# S4 method for lcModelCustom
getName(object, ...)

# S4 method for lcModelCustom
getShortName(object, ...)

# S4 method for lcModelCustom
converged(object, ...)

# S4 method for lcModelCustom
postprob(object, ...)

# S3 method for lcModelCustom
predict(object, ..., newdata = NULL, what = "mu")

# S4 method for lcModelCustom
predictPostprob(object, newdata = NULL, ...)

# S4 method for lcModelCustom
clusterTrajectories(object, at = time(object), ...)

# S4 method for lcModelCustom
fittedTrajectories(
  object,
  at = time(object),
  what = "mu",
  clusters = trajectoryAssignments(object),
  ...
)

# S4 method for lcMethodRandom
getArgumentDefaults(object)

# S4 method for lcMethodRandom
getName(object)

# S4 method for lcMethodRandom
getShortName(object)

# S4 method for lcMethodRandom
fit(method, data, envir, verbose, ...)

# S4 method for lcModelPartition
clusterTrajectories(
  object,
  at = time(object),
  center = object@center,
  approxFun = approx,
  ...
)

# S4 method for lcModelPartition
converged(object, ...)

# S4 method for lcModelPartition
getName(object, ...)

# S4 method for lcModelPartition
getShortName(object, ...)

# S4 method for lcModelPartition
postprob(object, ...)

# S4 method for lcModelStratify
converged(object, ...)

# S4 method for lcModelStratify
predictPostprob(object, newdata = NULL, ...)

# S4 method for lcModelWeightedPartition
clusterTrajectories(
  object,
  at = time(object),
  center = weighted.meanNA,
  approxFun = approx,
  ...
)

# S4 method for lcModelWeightedPartition
converged(object, ...)

# S4 method for lcModelWeightedPartition
getName(object, ...)

# S4 method for lcModelWeightedPartition
getShortName(object, ...)

# S4 method for lcModelWeightedPartition
postprob(object, ...)

Arguments

object
The lcMethod or lcModel object.

method
An object inheriting from lcMethod with all its arguments having been evaluated and finalized.

data
A data.frame representing the transformed training data.

verbose
A R.utils::Verbose object indicating the level of verbosity.

envir
The environment containing variables generated by prepareData() and preFit().

...
Additional arguments.

newdata
Optional data frame for which to compute the posterior probability. If omitted, the model training data is used.

what
The distributional parameter to predict. By default, the mean response 'mu' is predicted. The cluster membership predictions can be obtained by specifying what = 'mb'.

at
An optional vector of the times at which to compute the cluster trajectory predictions.

center
The function to use to compute the cluster trajectory center at the respective moment in time.

See also