mixtools interface

mixtools interface

# S4 method for class 'lcMethodMixtoolsGMM'
getArgumentDefaults(object)

# S4 method for class 'lcMethodMixtoolsGMM'
getArgumentExclusions(object)

# S4 method for class 'lcMethodMixtoolsGMM'
getCitation(object, ...)

# S4 method for class 'lcMethodMixtoolsGMM'
getName(object)

# S4 method for class 'lcMethodMixtoolsGMM'
getShortName(object)

# S4 method for class 'lcMethodMixtoolsGMM'
preFit(method, data, envir, verbose, ...)

# S4 method for class 'lcMethodMixtoolsGMM'
fit(method, data, envir, verbose, ...)

# S4 method for class 'lcMethodMixtoolsNPRM'
getArgumentDefaults(object)

# S4 method for class 'lcMethodMixtoolsNPRM'
getArgumentExclusions(object)

# S4 method for class 'lcMethodMixtoolsNPRM'
getCitation(object, ...)

# S4 method for class 'lcMethodMixtoolsNPRM'
getName(object)

# S4 method for class 'lcMethodMixtoolsNPRM'
getShortName(object)

# S4 method for class 'lcMethodMixtoolsNPRM'
fit(method, data, envir, verbose, ...)

# S4 method for class 'lcModelMixtoolsGMM'
predictForCluster(object, newdata, cluster, what = "mu", ...)

# S4 method for class 'lcModelMixtoolsGMM'
postprob(object, ...)

# S3 method for class 'lcModelMixtoolsGMM'
logLik(object, ...)

# S3 method for class 'lcModelMixtoolsGMM'
coef(object, ...)

# S3 method for class 'lcModelMixtoolsGMM'
sigma(object, ...)

# S4 method for class 'lcModelMixtoolsRM'
clusterTrajectories(
  object,
  at = time(object),
  what = "mu",
  se = TRUE,
  ci = c(0.025, 0.975),
  ...
)

# S4 method for class 'lcModelMixtoolsRM'
postprob(object, ...)

# S3 method for class 'lcModelMixtoolsRM'
logLik(object, ...)

# S4 method for class 'lcModelMixtoolsRM'
converged(object, ...)

Arguments

object

The object.

...

Not used.

method

An object inheriting from lcMethod with all its arguments having been evaluated and finalized.

data

A data.frame representing the transformed training data.

envir

The environment containing variables generated by prepareData() and preFit().

verbose

A R.utils::Verbose object indicating the level of verbosity.

newdata

A data.frame of trajectory data for which to compute trajectory assignments.

cluster

The cluster name (as character) to predict for.

what

The distributional parameter to predict. By default, the mean response 'mu' is predicted. The cluster membership predictions can be obtained by specifying what = 'mb'.

at

A numeric vector of the times at which to compute the cluster trajectories.

se

Whether to compute the standard error of the prediction.

ci

The confidence interval to compute.

See also