Supported methods for longitudinal clustering

This page provides an overview of the currently supported methods for longitudinal clustering. For general recommendations on which method to apply to your dataset, see here.

Supported methods

Method	Description	Source

lcMethodAkmedoids	Anchored k-medoids adepeju2020akmedoidslatrend	akmedoids

lcMethodCrimCV	Group-based trajectory modeling of count data nielsen2018crimcvlatrend	crimCV

lcMethodDtwclust	Methods for distance-based clustering, including dynamic time warping sardaespinosa2019timelatrend	dtwclust

lcMethodFeature	Feature-based clustering

lcMethodFlexmix	Interface to the FlexMix framework gruen2008flexmixlatrend	flexmix

lcMethodFlexmixGBTM	Group-based trajectory modeling	flexmix

lcMethodFunFEM	Model-based clustering using funFEM bouveyron2015funfemlatrend	funFEM

lcMethodGCKM	Growth-curve modeling and k-means	lme4

lcMethodKML	Longitudinal k-means genolini2015kmllatrend	kml

lcMethodLcmmGBTM	Group-based trajectory modeling proustlima2017estimationlatrend	lcmm

lcMethodLcmmGMM	Growth mixture modeling proustlima2017estimationlatrend	lcmm

lcMethodLMKM	Feature-based clustering using linear regression and k-means

lcMethodMclustLLPA	Longitudinal latent profile analysis scrucca2016mclustlatrend	mclust

lcMethodMixAK_GLMM	Mixture of generalized linear mixed models	mixAK

lcMethodMixtoolsGMM	Growth mixture modeling	mixtools

lcMethodMixtoolsNPRM	Non-parametric repeated measures clustering benaglia2009mixtoolslatrend	mixtools

lcMethodMixTVEM	Mixture of time-varying effects models

lcMethodRandom	Random partitioning

lcMethodStratify	Stratification rule

In addition, the functionality of any method can be extended via meta methods. This is used for extending the estimation procedure of a method, such as repeated fitting and selecting the best result, or fitting until convergence.

It is strongly encouraged to evaluate and compare several candidate methods in order to identify the most suitable method.

References

See also

Examples

data(latrendData)
method <- lcMethodLMKM(Y ~ Time, id = "Id", time = "Time")
model <- latrend(method, data = latrendData)