lcApproxModel class

approx models have defined cluster trajectories at fixed moments in time, which should be interpolated For a correct implementation, lcApproxModel requires the extending class to implement clusterTrajectories(at=NULL) to return the fixed cluster trajectories

# S3 method for lcApproxModel
fitted(object, ..., clusters = trajectoryAssignments(object))

# S4 method for lcApproxModel
predictForCluster(
  object,
  newdata,
  cluster,
  what = "mu",
  approxFun = approx,
  ...
)

Arguments

object

The lcModel object.

...

Additional arguments.

clusters

Optional cluster assignments per id. If unspecified, a matrix is returned containing the cluster-specific predictions per column.

newdata

Optional data.frame for which to compute the model predictions. If omitted, the model training data is used. Cluster trajectory predictions are made when ids are not specified.

cluster

The cluster name (as character) to predict for.

what

The distributional parameter to predict. By default, the mean response 'mu' is predicted. The cluster membership predictions can be obtained by specifying what = 'mb'.

approxFun

Function to interpolate between measurement moments, approx() by default.