Create a lcMethod object of the specified type and arguments

Note, see the lcMethod help page for documentation about the lcMethod class.

Provides a mechanism for creating lcMethod objects for an arbitrary class. However, it is advisable to use the subclass-specific constructors instead (e.g., lcMethodLcmmGBTM(), lcMethodKML(), or lcMethodLMKM()).

lcMethod(.class, ..., .defaults = list(), .excludeArgs = c())

Arguments

.class
The type of lcMethod class

...
Any arguments to assign to the method object.

.defaults
See defaults of lcMethod.call.

.excludeArgs
See excludeArgs of lcMethod.call.

Implementation

When creating your own lcMethod subclass, it is recommended to provide a constructor function for the subclass. Sensible default method argument values are are specified through the function arguments.

The lcMethod.call() function provides a convenient way to initialize the method according to the user-specified and default function arguments.

lcMethodExample <- function(
  response,
  time = getOption("latrend.time"),
  id = getOption("latrend.id"),
  nClusters = 2,
  # custom method argument defaults
  extraNum = 1,
  myOptAlgo = "EM",
  algoFunction = mean,
  ...
) {
  lcMethod.call(
    "lcMethodExample",
    call = stackoverflow::match.call.defaults()
  )
}

Users can then specify lcMethodExample objects using the lcMethodExample() function. For example, by calling lcMethodExample(response = "Y", nClusters = 3).

See also