Create a lcMethod object from a call

Creates a lcMethod class of the specified type Class for the given arguments given in a call, along with any default arguments from reference functions. This function is intended to be used by classes extending lcMethod to provide an easy way to construct the appropriate call object.

lcMethod.call(Class, call, defaults = list(), excludeArgs = c())

Arguments

Class
The type of lcMethod class

call
The arguments to create the lcMethod from.

defaults
List of function to obtain defaults from for arguments not defined in call.

excludeArgs
The names of the arguments to exclude from the defaults, provided as a character vector.

Value

An object of class Class that extends lcMethod.

See also

Examples

data(latrendData)
lcMethodKML2 <- function(response = "Y", id = "Id", time = "Time", nClusters = 2, ...) {
  lcMethod.call("lcMethodKML", call = stackoverflow::match.call.defaults(),
    defaults = c(kml::kml, kml::parALGO),
    excludeArgs = c("object", "nbClusters", "parAlgo", "toPlot", "saveFreq"))
}
method <- lcMethodKML2(nClusters = 3)
latrend(method, data = latrendData)

#>  ~ Fast KmL ~
#> ********************S

#> Longitudinal cluster model using longitudinal k-means (KML)
#> lcMethodKML as "longitudinal k-means (KML)"
#>  nClusters:      3
#>  time:           "Time"
#>  id:             "Id"
#>  response:       "Y"
#>  scale:          TRUE
#>  nbCriterion:    1000
#>  startingCond:   "nearlyAll"
#>  centerMethod:   function (x) {    mean(x, na.rm = TRUE)}
#>  distance:       function () {}
#>  power:          2
#>  distanceName:   "euclidean"
#>  imputationMethod:"copyMean"
#>  maxIt:          200
#>  nbRedrawing:    20
#> 
#> Cluster sizes (K=3):
#>           A           B           C 
#> 105 (52.5%)  52 (26.2%)  43 (21.3%) 
#> 
#> Number of obs: 2000, strata (Id): 200
#> 
#> Scaled residuals:
#>     Min.  1st Qu.   Median     Mean  3rd Qu.     Max. 
#> -4.32707 -0.61431 -0.01003  0.00000  0.64212  5.19705 
#>