Specify GBTM method

Group-based trajectory modeling through fixed-effects modeling.

lcMethodLcmmGBTM(
  fixed,
  mixture = ~1,
  classmb = ~1,
  time = getOption("latrend.time"),
  id = getOption("latrend.id"),
  nClusters = 2,
  init = "default",
  ...
)

Arguments

fixed

The fixed effects formula.

mixture

The mixture-specific effects formula. See lcmm::hlme for details.

classmb

The cluster membership formula for the multinomial logistic model. See lcmm::hlme for details.

time

The name of the time variable.

id

The name of the trajectory identifier variable. This replaces the subject argument of lcmm::hlme.

nClusters

The number of clusters to fit. This replaces the ng argument of lcmm::hlme.

init

Alternative for the B argument of lcmm::hlme, for initializing the hlme fitting procedure. If "lme.random" (default): random initialization through a standard linear mixed model. Assigns a fitted standard linear mixed model enclosed in a call to random() to the B argument. If "lme", fits a standard linear mixed model and passes this to the B argument. If NULL or "default", the default lcmm::hlme input for B is used.

The argument is ignored if the B argument is specified, or nClusters = 1.

...

Arguments passed to lcmm::hlme. The following arguments are ignored: data, fixed, random, mixture, subject, classmb, returndata, ng, verbose, subset.

References

Proust-Lima C, Philipps V, Liquet B (2017). “Estimation of Extended Mixed Models Using Latent Classes and Latent Processes: The R Package lcmm.” Journal of Statistical Software, 78(2), 1--56. doi:10.18637/jss.v078.i02 .

Proust-Lima C, Philipps V, Diakite A, Liquet B (2019). lcmm: Extended Mixed Models Using Latent Classes and Latent Processes. R package version: 1.8.1, https://cran.r-project.org/package=lcmm.

See also

Examples

data(latrendData)
if (rlang::is_installed("lcmm")) {
  method <- lcMethodLcmmGBTM(
    fixed = Y ~ Time,
    mixture = ~ 1,
   id = "Id",
   time = "Time",
   nClusters = 3
  )
  gbtm <- latrend(method, data = latrendData)
  summary(gbtm)

  method <- lcMethodLcmmGBTM(
    fixed = Y ~ Time,
    mixture = ~ Time,
    id = "Id",
    time = "Time",
    nClusters = 3
  )
}
#> ------------------ iteration  0 ------------------
#> Function value  -3236.988 
#> Convergence criteria: parameters stability= 1.0001 
#>                     : function stability= 1.0001 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef   SE.coef   Var.coef
#> parameter1 -0.02626082  6.687442   44.72189
#> parameter2  1.00000000  5.342833   28.54586
#> parameter3  1.00000000 79.948167 6391.70944
#> 
#> ------------------ iteration  1 ------------------
#> Function value  -1641.944 
#> Convergence criteria: parameters stability= 0.4794574 
#>                     : function stability= 1595.044 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                  coef   SE.coef  Var.coef
#> parameter1 -0.4733827  7.675149  58.90792
#> parameter2  0.5293172  6.131886  37.60003
#> parameter3  0.7591741 15.032362 225.97191
#> 
#> ------------------ iteration  2 ------------------
#> Function value  -1641.926 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 0.01839313 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                  coef   SE.coef  Var.coef
#> parameter1 -0.4733828  7.675200  58.90870
#> parameter2  0.5293152  6.131926  37.60052
#> parameter3  0.7591641 15.035399 226.06323
#> 
#> ------------------ iteration  3 ------------------
#> Function value  -1641.907 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 0.01804602 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                  coef   SE.coef  Var.coef
#> parameter1 -0.4733831  7.675251  58.90947
#> parameter2  0.5293138  6.131967  37.60102
#> parameter3  0.7591541 15.038404 226.15358
#> 
#> ------------------ iteration  4 ------------------
#> Function value  -1641.89 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 0.01774322 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                  coef   SE.coef  Var.coef
#> parameter1 -0.4733835  7.675300  58.91023
#> parameter2  0.5293129  6.132007  37.60151
#> parameter3  0.7591441 15.041479 226.24610
#> 
#> ------------------ iteration  5 ------------------
#> Function value  -1641.872 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 0.01751268 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                  coef   SE.coef  Var.coef
#> parameter1 -0.4733838  7.675352  58.91102
#> parameter2  0.5293123  6.132048  37.60201
#> parameter3  0.7591341 15.044767 226.34502
#> 
#> ------------------ iteration  6 ------------------
#> Function value  -1572.704 
#> Convergence criteria: parameters stability= 0.2070615 
#>                     : function stability= 69.16839 
#>                     : relative distance to maximum(RDM)= 260.8259 
#>                  coef    SE.coef   Var.coef
#> parameter1 -0.4839071 0.01357611 0.00018431
#> parameter2  0.5092916 0.01227396 0.00015065
#> parameter3  0.3046565 0.00342945 0.00001176
#> 
#> ------------------ iteration  7 ------------------
#> Function value  -1368.127 
#> Convergence criteria: parameters stability= 0.00138633 
#>                     : function stability= 204.5767 
#>                     : relative distance to maximum(RDM)= 170.5737 
#>                  coef    SE.coef   Var.coef
#> parameter1 -0.4882755 0.01445842 0.00020905
#> parameter2  0.4774404 0.01268166 0.00016082
#> parameter3  0.3234379 0.00388005 0.00001505
#> 
#> ------------------ iteration  8 ------------------
#> Function value  -1142.83 
#> Convergence criteria: parameters stability= 0.00908887 
#>                     : function stability= 225.2967 
#>                     : relative distance to maximum(RDM)= 64.22842 
#>                  coef    SE.coef  Var.coef
#> parameter1 -0.4405171 0.01571312 2.469e-04
#> parameter2  0.4036922 0.01339039 1.793e-04
#> parameter3  0.3604406 0.00490796 2.409e-05
#> 
#> ------------------ iteration  9 ------------------
#> Function value  -1040.808 
#> Convergence criteria: parameters stability= 0.02660229 
#>                     : function stability= 102.0227 
#>                     : relative distance to maximum(RDM)= 7.337037 
#>                  coef    SE.coef   Var.coef
#> parameter1 -0.3246307 0.01616458 0.00026129
#> parameter2  0.2919675 0.01363459 0.00018590
#> parameter3  0.3867128 0.00572969 0.00003283
#> 
#> ------------------ iteration  10 ------------------
#> Function value  -1029.112 
#> Convergence criteria: parameters stability= 0.00504075 
#>                     : function stability= 11.69597 
#>                     : relative distance to maximum(RDM)= 0.1363509 
#>                  coef    SE.coef   Var.coef
#> parameter1 -0.2728270 0.01669294 0.00027865
#> parameter2  0.2457638 0.01407102 0.00019799
#> parameter3  0.4016243 0.00627880 0.00003942
#> 
#> ------------------ iteration  11 ------------------
#> Function value  -1028.905 
#> Convergence criteria: parameters stability= 8.14e-05 
#>                     : function stability= 0.2072104 
#>                     : relative distance to maximum(RDM)= 7.265e-05 
#>                  coef    SE.coef   Var.coef
#> parameter1 -0.2664429 0.01681838 0.00028286
#> parameter2  0.2401646 0.01417667 0.00020098
#> parameter3  0.4046723 0.00639819 0.00004094
#> 
#> ------------------ iteration  12 ------------------
#> Function value  -1028.905 
#> Convergence criteria: parameters stability= 4e-08 
#>                     : function stability= 0.00010898 
#>                     : relative distance to maximum(RDM)= 0 
#>                  coef    SE.coef   Var.coef
#> parameter1 -0.2663089 0.01682159 0.00028297
#> parameter2  0.2400481 0.01417938 0.00020105
#> parameter3  0.4047497 0.00640125 0.00004098
#> 
#> ------------------ iteration  13 ------------------
#> Function value  -1028.905 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 0 
#>                     : relative distance to maximum(RDM)= 0 
#>                  coef    SE.coef   Var.coef
#> parameter1 -0.2663087 0.01682159 0.00028297
#> parameter2  0.2400479 0.01417938 0.00020105
#> parameter3  0.4047498 0.00640126 0.00004098
#> 
#> The program took 0.18 seconds 
#> ------------------ iteration  0 ------------------
#> Function value  -1026.27 
#> Convergence criteria: parameters stability= 1.0001 
#>                     : function stability= 1.0001 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                  coef    SE.coef    Var.coef
#> parameter1  0.0000000   8.259062   -68.21210
#> parameter2  0.0000000   6.743496    45.47474
#> parameter3 -0.2831303  63.734506 -4062.08720
#> parameter4 -0.2663087  68.982481 -4758.58265
#> parameter5 -0.2494872  64.584579 -4171.16787
#> parameter6  0.2400479 130.559087 17045.67513
#> parameter7  0.4047498 154.941511 24006.87170
#> 
#> ------------------ iteration  1 ------------------
#> Function value  -956.0388 
#> Convergence criteria: parameters stability= 0.02285062 
#>                     : function stability= 70.23141 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef    SE.coef     Var.coef
#> parameter1  0.00043624   3.417613    11.680077
#> parameter2 -0.00108898   2.434925     5.928859
#> parameter3 -0.37923255   6.879311    47.324915
#> parameter4 -0.28495974  13.454482   181.023086
#> parameter5 -0.13493035  51.458614  2647.988939
#> parameter6  0.23993870 117.900014 13900.413206
#> parameter7  0.39281357 145.912831 21290.554148
#> 
#> ------------------ iteration  2 ------------------
#> Function value  -942.0304 
#> Convergence criteria: parameters stability= 0.00152312 
#>                     : function stability= 14.0084 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef    SE.coef     Var.coef
#> parameter1  0.00243707   2.422769     5.869812
#> parameter2 -0.00918748   1.595478     2.545549
#> parameter3 -0.39924939   6.555062    42.968836
#> parameter4 -0.30762683  25.439361   647.161064
#> parameter5 -0.11216873  54.925525  3016.813322
#> parameter6  0.24049041 114.679565 13151.402553
#> parameter7  0.38826820 147.639791 21797.507748
#> 
#> ------------------ iteration  3 ------------------
#> Function value  -925.2262 
#> Convergence criteria: parameters stability= 0.03480229 
#>                     : function stability= 16.80425 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef    SE.coef     Var.coef
#> parameter1  0.09389055   2.352119     5.532465
#> parameter2 -0.01398510   1.707941     2.917061
#> parameter3 -0.50113656   5.555037    30.858434
#> parameter4 -0.19570462  42.051675  1768.343375
#> parameter5 -0.05682425  38.634437  1492.619712
#> parameter6  0.24113454 100.253855 10050.835464
#> parameter7  0.36718433 164.394991 27025.713204
#> 
#> ------------------ iteration  4 ------------------
#> Function value  -925.2257 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 0.00044741 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef    SE.coef     Var.coef
#> parameter1  0.09388640   2.352147     5.532593
#> parameter2 -0.01397510   1.635883     2.676112
#> parameter3 -0.50113731   5.228913    27.341529
#> parameter4 -0.19570630  42.052503  1768.413025
#> parameter5 -0.05682292  38.634566  1492.629693
#> parameter6  0.24113439 100.253792 10050.822792
#> parameter7  0.36718421 164.394990 27025.712606
#> 
#> ------------------ iteration  5 ------------------
#> Function value  -906.741 
#> Convergence criteria: parameters stability= 0.1337967 
#>                     : function stability= 18.48475 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef    SE.coef     Var.coef
#> parameter1  0.02210043   2.229236     4.969492
#> parameter2  0.32459499   2.047375     4.191743
#> parameter3 -0.50595717   3.247530    10.546454
#> parameter4 -0.28486601  63.948622  4089.426302
#> parameter5  0.02049704  42.207445  1781.468393
#> parameter6  0.23722364 100.291454 10058.375752
#> parameter7  0.36030020 168.781331 28487.137767
#> 
#> ------------------ iteration  6 ------------------
#> Function value  -890.7368 
#> Convergence criteria: parameters stability= 5.523537 
#>                     : function stability= 16.00416 
#>                     : relative distance to maximum(RDM)= 2.747751 
#>                  coef    SE.coef   Var.coef
#> parameter1 -0.8244579 0.63298977 0.40067605
#> parameter2  2.5002434 0.36246013 0.13137735
#> parameter3 -0.7408270 0.04406908 0.00194208
#> parameter4 -0.2884020 0.01646407 0.00027107
#> parameter5  0.1550190 0.03620434 0.00131075
#> parameter6  0.2360427 0.01218550 0.00014849
#> parameter7  0.3478163 0.00540948 0.00002926
#> 
#> ------------------ iteration  7 ------------------
#> Function value  -883.2225 
#> Convergence criteria: parameters stability= 0.8614301 
#>                     : function stability= 7.514296 
#>                     : relative distance to maximum(RDM)= 0.1861671 
#>                  coef    SE.coef   Var.coef
#> parameter1 -0.5967116 0.40611738 0.16493132
#> parameter2  1.6014142 0.29380636 0.08632218
#> parameter3 -0.7448678 0.05096139 0.00259706
#> parameter4 -0.2925786 0.01834999 0.00033672
#> parameter5  0.1155492 0.04388711 0.00192608
#> parameter6  0.2394306 0.01246626 0.00015541
#> parameter7  0.3558483 0.00571731 0.00003269
#> 
#> ------------------ iteration  8 ------------------
#> Function value  -882.5454 
#> Convergence criteria: parameters stability= 0.06700964 
#>                     : function stability= 0.6771088 
#>                     : relative distance to maximum(RDM)= 0.0029959 
#>                  coef    SE.coef   Var.coef
#> parameter1 -0.8525454 0.44769854 0.20043398
#> parameter2  1.6336113 0.31351803 0.09829356
#> parameter3 -0.7582006 0.04822534 0.00232568
#> parameter4 -0.2984322 0.01871005 0.00035007
#> parameter5  0.0979872 0.04380724 0.00191907
#> parameter6  0.2400050 0.01250541 0.00015639
#> parameter7  0.3569665 0.00577124 0.00003331
#> 
#> ------------------ iteration  9 ------------------
#> Function value  -882.5348 
#> Convergence criteria: parameters stability= 0.00252938 
#>                     : function stability= 0.01062605 
#>                     : relative distance to maximum(RDM)= 2.15e-06 
#>                   coef    SE.coef   Var.coef
#> parameter1 -0.90225646 0.45371263 0.20585515
#> parameter2  1.62599750 0.31271004 0.09778757
#> parameter3 -0.75795363 0.04805944 0.00230971
#> parameter4 -0.29872910 0.01874108 0.00035123
#> parameter5  0.09773208 0.04393013 0.00192986
#> parameter6  0.24004770 0.01250802 0.00015645
#> parameter7  0.35704107 0.00577488 0.00003335
#> 
#> ------------------ iteration  10 ------------------
#> Function value  -882.5348 
#> Convergence criteria: parameters stability= 2.36e-06 
#>                     : function stability= 7.55e-06 
#>                     : relative distance to maximum(RDM)= 0 
#>                   coef    SE.coef   Var.coef
#> parameter1 -0.90377582 0.45391107 0.20603526
#> parameter2  1.62577349 0.31272756 0.09779853
#> parameter3 -0.75800710 0.04805548 0.00230933
#> parameter4 -0.29874490 0.01874293 0.00035130
#> parameter5  0.09769893 0.04393471 0.00193026
#> parameter6  0.24004793 0.01250806 0.00015645
#> parameter7  0.35704201 0.00577496 0.00003335
#> 
#> The program took 0.37 seconds