Specify GMM method using lcmm

Growth mixture modeling through latent-class linear mixed modeling.

lcMethodLcmmGMM(
  fixed,
  mixture = ~1,
  random = ~1,
  classmb = ~1,
  time = getOption("latrend.time"),
  id = getOption("latrend.id"),
  init = "lme",
  nClusters = 2,
  ...
)

Arguments

fixed

The fixed effects formula.

mixture

The mixture-specific effects formula. See lcmm::hlme for details.

random

The random effects formula. See lcmm::hlme for details.

classmb

The cluster membership formula for the multinomial logistic model. See lcmm::hlme for details.

time

The name of the time variable.

id

The name of the trajectory identifier variable. This replaces the subject argument of lcmm::hlme.

init

Alternative for the B argument of lcmm::hlme, for initializing the hlme fitting procedure. This is only applicable for nClusters > 1. Options:

"lme.random" (default): random initialization through a standard linear mixed model. Assigns a fitted standard linear mixed model enclosed in a call to random() to the B argument.

"lme", fits a standard linear mixed model and passes this to the B argument.

"gridsearch", a gridsearch is used with initialization from "lme.random", following the approach used by lcmm::gridsearch. To use this initalization, specify arguments gridsearch.maxiter (max number of iterations during search), gridsearch.rep (number of fits during search), and gridsearch.parallel (whether to enable parallel computation).

NULL or "default", the default lcmm::hlme input for B is used.

The argument is ignored if the B argument is specified, or nClusters = 1.

nClusters

The number of clusters to fit. This replaces the ng argument of lcmm::hlme.

...

Arguments passed to lcmm::hlme. The following arguments are ignored: data, fixed, random, mixture, subject, classmb, returndata, ng, verbose, subset.

References

Proust-Lima C, Philipps V, Liquet B (2017). “Estimation of Extended Mixed Models Using Latent Classes and Latent Processes: The R Package lcmm.” Journal of Statistical Software, 78(2), 1--56. doi:10.18637/jss.v078.i02 .

Proust-Lima C, Philipps V, Diakite A, Liquet B (2019). lcmm: Extended Mixed Models Using Latent Classes and Latent Processes. R package version: 1.8.1, https://cran.r-project.org/package=lcmm.

See also

Examples

data(latrendData)

if (rlang::is_installed("lcmm")) {
  method <- lcMethodLcmmGMM(
    fixed = Y ~ Time,
    mixture = ~ Time,
    random = ~ 1,
    id = "Id",
    time = "Time",
    nClusters = 2
  )
  gmm <- latrend(method, data = latrendData)
  summary(gmm)

  # define method with gridsearch
  method <- lcMethodLcmmGMM(
    fixed = Y ~ Time,
    mixture = ~ Time,
    random = ~ 1,
    id = "Id",
    time = "Time",
    nClusters = 3,
    init = "gridsearch",
    gridsearch.maxiter = 10,
    gridsearch.rep = 50,
    gridsearch.parallel = TRUE
  )
}
#> ------------------ iteration  0 ------------------
#> Function value  -2211.674 
#> Convergence criteria: parameters stability= 1.0001 
#>                     : function stability= 1.0001 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef
#> parameter1 -0.03295723
#> parameter2  0.00000000
#> parameter3  1.00000000
#> parameter4  1.00000000
#> 
#> ------------------ iteration  1 ------------------
#> Function value  -2211.661 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 0.01365446 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef
#> parameter1 -0.03295781
#> parameter2  0.00000137
#> parameter3  0.99999000
#> parameter4  0.99999246
#> 
#> ------------------ iteration  2 ------------------
#> Function value  -2211.643 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 0.01733755 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef
#> parameter1 -0.03295808
#> parameter2  0.00000233
#> parameter3  0.99998055
#> parameter4  0.99998246
#> 
#> ------------------ iteration  3 ------------------
#> Function value  -1439.862 
#> Convergence criteria: parameters stability= 0.280072 
#>                     : function stability= 771.7816 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef
#> parameter1 -0.04031975
#> parameter2  0.04576697
#> parameter3  0.64986537
#> parameter4  0.60584696
#> 
#> ------------------ iteration  4 ------------------
#> Function value  -1439.841 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 0.02074656 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef
#> parameter1 -0.04032027
#> parameter2  0.04576783
#> parameter3  0.64985797
#> parameter4  0.60583696
#> 
#> ------------------ iteration  5 ------------------
#> Function value  -1259.308 
#> Convergence criteria: parameters stability= 0.1423089 
#>                     : function stability= 180.5326 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef
#> parameter1 -0.05020431
#> parameter2  0.06607079
#> parameter3  0.47182147
#> parameter4  0.27402077
#> 
#> ------------------ iteration  6 ------------------
#> Function value  -1182.724 
#> Convergence criteria: parameters stability= 1.186942 
#>                     : function stability= 76.58411 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef
#> parameter1 -0.17261108
#> parameter2  0.09024304
#> parameter3 -0.61021076
#> parameter4  0.29811228
#> 
#> ------------------ iteration  7 ------------------
#> Function value  -1182.721 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 0.00300208 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef
#> parameter1 -0.17261028
#> parameter2  0.09024312
#> parameter3 -0.61020076
#> parameter4  0.29811233
#> 
#> ------------------ iteration  8 ------------------
#> Function value  -1182.718 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 0.00380941 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef
#> parameter1 -0.17260961
#> parameter2  0.09024344
#> parameter3 -0.61019076
#> parameter4  0.29811260
#> 
#> ------------------ iteration  9 ------------------
#> Function value  -1070.899 
#> Convergence criteria: parameters stability= 0.05788247 
#>                     : function stability= 111.819 
#>                     : relative distance to maximum(RDM)= 735.0503 
#>                  coef
#> parameter1 -0.1578944
#> parameter2  0.1032584
#> parameter3 -0.3706614
#> parameter4  0.3091681
#> 
#> ------------------ iteration  10 ------------------
#> Function value  -1070.895 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 0.00369269 
#>                     : relative distance to maximum(RDM)= 734.1069 
#>                  coef
#> parameter1 -0.1578937
#> parameter2  0.1032586
#> parameter3 -0.3706514
#> parameter4  0.3091682
#> 
#> ------------------ iteration  11 ------------------
#> Function value  -1070.891 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 0.00379712 
#>                     : relative distance to maximum(RDM)= 732.7804 
#>                  coef
#> parameter1 -0.1578932
#> parameter2  0.1032587
#> parameter3 -0.3706414
#> parameter4  0.3091683
#> 
#> ------------------ iteration  12 ------------------
#> Function value  -1044.445 
#> Convergence criteria: parameters stability= 0.5764923 
#>                     : function stability= 26.44575 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                  coef
#> parameter1 -0.1270200
#> parameter2  0.1190798
#> parameter3  0.3877042
#> parameter4  0.3233399
#> 
#> ------------------ iteration  13 ------------------
#> Function value  -1044.442 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 0.00348247 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                  coef
#> parameter1 -0.1270206
#> parameter2  0.1190798
#> parameter3  0.3876942
#> parameter4  0.3233399
#> 
#> ------------------ iteration  14 ------------------
#> Function value  -1044.438 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 0.00354728 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                  coef
#> parameter1 -0.1270210
#> parameter2  0.1190798
#> parameter3  0.3876842
#> parameter4  0.3233400
#> 
#> ------------------ iteration  15 ------------------
#> Function value  -1044.435 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 0.00359666 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                  coef
#> parameter1 -0.1270214
#> parameter2  0.1190799
#> parameter3  0.3876742
#> parameter4  0.3233402
#> 
#> ------------------ iteration  16 ------------------
#> Function value  -990.0216 
#> Convergence criteria: parameters stability= 0.03537843 
#>                     : function stability= 54.41312 
#>                     : relative distance to maximum(RDM)= 12.63568 
#>                  coef
#> parameter1 -0.1312689
#> parameter2  0.1205435
#> parameter3  0.1996608
#> parameter4  0.3263760
#> 
#> ------------------ iteration  17 ------------------
#> Function value  -973.447 
#> Convergence criteria: parameters stability= 0.00032527 
#>                     : function stability= 16.5746 
#>                     : relative distance to maximum(RDM)= 5.923468 
#>                  coef
#> parameter1 -0.1424467
#> parameter2  0.1250521
#> parameter3  0.1966762
#> parameter4  0.3394560
#> 
#> ------------------ iteration  18 ------------------
#> Function value  -961.3927 
#> Convergence criteria: parameters stability= 0.00105038 
#>                     : function stability= 12.05426 
#>                     : relative distance to maximum(RDM)= 0.593486 
#>                  coef
#> parameter1 -0.1651658
#> parameter2  0.1362990
#> parameter3  0.1897438
#> parameter4  0.3584212
#> 
#> ------------------ iteration  19 ------------------
#> Function value  -960.1702 
#> Convergence criteria: parameters stability= 0.00024153 
#>                     : function stability= 1.222577 
#>                     : relative distance to maximum(RDM)= 0.00468292 
#>                  coef
#> parameter1 -0.1762450
#> parameter2  0.1435649
#> parameter3  0.1871356
#> parameter4  0.3661144
#> 
#> ------------------ iteration  20 ------------------
#> Function value  -960.1608 
#> Convergence criteria: parameters stability= 2.61e-06 
#>                     : function stability= 0.00939256 
#>                     : relative distance to maximum(RDM)= 5.3e-07 
#>                  coef
#> parameter1 -0.1773901
#> parameter2  0.1444374
#> parameter3  0.1869344
#> parameter4  0.3668221
#> 
#> ------------------ iteration  21 ------------------
#> Function value  -960.1608 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 1.06e-06 
#>                     : relative distance to maximum(RDM)= 0 
#>                  coef
#> parameter1 -0.1774069
#> parameter2  0.1444497
#> parameter3  0.1869323
#> parameter4  0.3668285
#> 
#> The program took 0.53 seconds 
#> ------------------ iteration  0 ------------------
#> Function value  -959.2723 
#> Convergence criteria: parameters stability= 1.0001 
#>                     : function stability= 1.0001 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                  coef
#> parameter1  0.0000000
#> parameter2 -0.1874958
#> parameter3 -0.1673180
#> parameter4  0.1380242
#> parameter5  0.1508751
#> parameter6  0.1869323
#> parameter7  0.3668285
#> 
#> ------------------ iteration  1 ------------------
#> Function value  -957.172 
#> Convergence criteria: parameters stability= 6.145e-05 
#>                     : function stability= 2.100332 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                  coef
#> parameter1  0.0000005
#> parameter2 -0.1875478
#> parameter3 -0.1672656
#> parameter4  0.1324900
#> parameter5  0.1564099
#> parameter6  0.1867822
#> parameter7  0.3664314
#> 
#> ------------------ iteration  2 ------------------
#> Function value  -939.807 
#> Convergence criteria: parameters stability= 0.00087023 
#>                     : function stability= 17.36502 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef
#> parameter1  0.00005488
#> parameter2 -0.18764726
#> parameter3 -0.16715930
#> parameter4  0.11164707
#> parameter5  0.17725207
#> parameter6  0.18612558
#> parameter7  0.36545858
#> 
#> ------------------ iteration  3 ------------------
#> Function value  -920.3918 
#> Convergence criteria: parameters stability= 0.00046485 
#>                     : function stability= 19.41517 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef
#> parameter1 -0.00008149
#> parameter2 -0.18638160
#> parameter3 -0.16848824
#> parameter4  0.09655969
#> parameter5  0.19251663
#> parameter6  0.18589413
#> parameter7  0.36457657
#> 
#> ------------------ iteration  4 ------------------
#> Function value  -870.4228 
#> Convergence criteria: parameters stability= 0.00225473 
#>                     : function stability= 49.96899 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef
#> parameter1 -0.00013076
#> parameter2 -0.18534977
#> parameter3 -0.16962638
#> parameter4  0.06420283
#> parameter5  0.22723004
#> parameter6  0.18585681
#> parameter7  0.36396508
#> 
#> ------------------ iteration  5 ------------------
#> Function value  -807.3687 
#> Convergence criteria: parameters stability= 0.00288987 
#>                     : function stability= 63.05413 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef
#> parameter1  0.00154924
#> parameter2 -0.17307886
#> parameter3 -0.18296970
#> parameter4  0.03017905
#> parameter5  0.26391804
#> parameter6  0.18744110
#> parameter7  0.35673374
#> 
#> ------------------ iteration  6 ------------------
#> Function value  -234.0025 
#> Convergence criteria: parameters stability= 0.2354196 
#>                     : function stability= 573.3662 
#>                     : relative distance to maximum(RDM)= 88.25512 
#>                   coef
#> parameter1  0.15233864
#> parameter2  0.03617987
#> parameter3 -0.40084166
#> parameter4 -0.17597225
#> parameter5  0.51667596
#> parameter6  0.24031620
#> parameter7  0.24608093
#> 
#> ------------------ iteration  7 ------------------
#> Function value  -116.8424 
#> Convergence criteria: parameters stability= 0.2750103 
#>                     : function stability= 117.1601 
#>                     : relative distance to maximum(RDM)= 14.02445 
#>                  coef
#> parameter1  0.3350642
#> parameter2  0.2213973
#> parameter3 -0.7824800
#> parameter4 -0.2305114
#> parameter5  0.7425480
#> parameter6  0.1864792
#> parameter7  0.1769633
#> 
#> ------------------ iteration  8 ------------------
#> Function value  -60.66954 
#> Convergence criteria: parameters stability= 0.00715973 
#>                     : function stability= 56.17281 
#>                     : relative distance to maximum(RDM)= 1.323995 
#>                  coef
#> parameter1  0.3942235
#> parameter2  0.1995742
#> parameter3 -0.7462780
#> parameter4 -0.2256329
#> parameter5  0.7082258
#> parameter6  0.2036181
#> parameter7  0.1963930
#> 
#> ------------------ iteration  9 ------------------
#> Function value  -55.73164 
#> Convergence criteria: parameters stability= 0.00050486 
#>                     : function stability= 4.937899 
#>                     : relative distance to maximum(RDM)= 0.01668516 
#>                  coef
#> parameter1  0.4052535
#> parameter2  0.1937467
#> parameter3 -0.7346090
#> parameter4 -0.2240926
#> parameter5  0.6980466
#> parameter6  0.2096016
#> parameter7  0.2048358
#> 
#> ------------------ iteration  10 ------------------
#> Function value  -55.67275 
#> Convergence criteria: parameters stability= 4.07e-06 
#>                     : function stability= 0.05888992 
#>                     : relative distance to maximum(RDM)= 3.06e-06 
#>                  coef
#> parameter1  0.4054646
#> parameter2  0.1932927
#> parameter3 -0.7334642
#> parameter4 -0.2239387
#> parameter5  0.6970418
#> parameter6  0.2101172
#> parameter7  0.2059342
#> 
#> ------------------ iteration  11 ------------------
#> Function value  -55.67274 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 1.071e-05 
#>                     : relative distance to maximum(RDM)= 0 
#>                  coef
#> parameter1  0.4054651
#> parameter2  0.1932891
#> parameter3 -0.7334510
#> parameter4 -0.2239371
#> parameter5  0.6970299
#> parameter6  0.2101202
#> parameter7  0.2059496
#> 
#> The program took 0.48 seconds