Specify a GLMM iwht a normal mixture in the random effects

Specify a GLMM iwht a normal mixture in the random effects

lcMethodMixAK_GLMM(
  fixed,
  random,
  time = getOption("latrend.time"),
  id = getOption("latrend.id"),
  nClusters = 2,
  ...
)

Arguments

fixed
A formula specifying the fixed effects of the model, including the response. Creates the y and x arguments for the call to mixAK::GLMM_MCMC.

random
A formula specifying the random effects of the model, including the random intercept. Creates the z and random.intercept arguments for the call to mixAK::GLMM_MCMC.

time
The name of the time variable.

id
The name of the trajectory identifier variable. This is used to generate the id vector argument for the call to mixAK::GLMM_MCMC.

nClusters
The number of clusters.

...
Arguments passed to mixAK::GLMM_MCMC. The following external arguments are ignored: y, x, z, random.intercept, silent.

References

Komárek A (2009). “A New R Package for Bayesian Estimation of Multivariate Normal Mixtures Allowing for Selection of the Number of Components and Interval-Censored Data.” Computational Statistics \& Data Analysis, 53(12), 3932--3947. doi: 10.1016/j.csda.2009.05.006 .

See also

Examples

data(latrendData)
# this example only runs when the mixAK package is installed
try({
 method <- lcMethodMixAK_GLMM(fixed = Y ~ 1, random = ~ Time,
  id = "Id", time = "Time", nClusters = 3)
 model <- latrend(method, latrendData)
 summary(model)
})

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Longitudinal cluster model using generalized linear mixed model with normal random effects mixture
#> lcMethodMixAK_GLMM as "generalized linear mixed model with normal random effects mixture"
#>  time:           "Time"
#>  id:             "Id"
#>  nClusters:      3
#>  dist:           "gaussian"
#>  nMCMC:          10, 10, 1, 10
#>  tuneMCMC:       list(alpha = 1, b = 1)
#>  store:          FALSE
#>  PED:            TRUE
#>  keep.chains:    TRUE
#>  dens.zero:      1e-300
#>  parallel:       FALSE
#>  fixed:          Y ~ 1
#>  random:         ~Time
#> 
#> Cluster sizes (K=3):
#>           A           B           C 
#>  78 (27.4%)   9 (16.1%) 113 (56.5%) 
#> 
#> Number of obs: 2000, strata (Id): 200
#> 
#> Scaled residuals:
#>     Min.  1st Qu.   Median     Mean  3rd Qu.     Max. 
#> -4.24257 -0.58708  0.01048  0.00000  0.65141  2.65457 
#>