lcModel class

Abstract class for defining estimated longitudinal cluster models.

Arguments

object

The lcModel object.

...

Any additional arguments.

Details

An extending class must implement the following methods to ensure basic functionality:

predict.lcModelExt: Used to obtain the fitted cluster trajectories and trajectories.

postprob(lcModelExt): The posterior probability matrix is used to determine the cluster assignments of the trajectories.

For predicting the posterior probability for unseen data, the predictPostprob() should be implemented.

Slots

method

The lcMethod-class object specifying the arguments under which the model was fitted.

call

The call that was used to create this lcModel object. Typically, this is the call to latrend() or any of the other fitting functions.

model

An arbitrary underlying model representation.

data

A data.frame object, or an expression to resolves to the data.frame object.

date

The date-time when the model estimation was initiated.

id

The name of the trajectory identifier column.

time

The name of the time variable.

response

The name of the response variable.

label

The label assigned to this model.

ids

The trajectory identifier values the model was fitted on.

times

The exact times on which the model has been trained

clusterNames

The names of the clusters.

estimationTime

The time, in seconds, that it took to fit the model.

tag

An arbitrary user-specified data structure. This slot may be accessed and updated directly.

See also