Create a lcModel with pre-defined partitioning

Represents an arbitrary partitioning of a set of trajectories. As such, this model has no predictive capabilities. The cluster trajectories are represented by the specified center function (mean by default).

lcModelPartition(
  data,
  response,
  trajectoryAssignments,
  nClusters = NA,
  center = meanNA,
  clusterNames = NULL,
  time = getOption("latrend.time"),
  id = getOption("latrend.id"),
  name = "part",
  envir = parent.frame()
)

Arguments

data
A data.frame representing the trajectory data.

response
The name of the response variable.

trajectoryAssignments
A vector of cluster membership per trajectory, a data.frame with an id column and "Cluster" column, or the name of the cluster membership column in the data argument.. For vector input, the type must be factor, character, or integer (1 to nClusters). The order of the trajectory, and thus the respective assignments, is determined by the id column of the data. Provide a factor id column for the input data to ensure that the ordering is as you aspect.

nClusters
The number of clusters. Should be NA for trajectory assignments of type factor.

center
The function for computing the longitudinal cluster centers, used for representing the cluster trajectories.

clusterNames
The names of the clusters, or a function with input n outputting a character vector of names. If unspecified, the names are determined from the trajectoryAssignments argument.

time
The name of the time variable.

id
The name of the trajectory identification variable.

name
The name of the method.

envir
The environment associated with the model. Used for evaluating the assigned data object by model.data.lcModel.

Examples

# comparing a model to the ground truth using the adjusted Rand index
data(latrendData)
model <- latrend(lcMethodKML(), data = latrendData, response = "Y")

#>  ~ Fast KmL ~
#> ********************S

# extract the reference class from the Class column
trajLabels <- aggregate(Class ~ Id, head, 1, data = latrendData)
trajLabels$Cluster <- trajLabels$Class
refModel <- lcModelPartition(latrendData, response = "Y", trajectoryAssignments = trajLabels)
externalMetric(model, refModel, 'adjustedRand') # 0.76

#> adjustedRand 
#>    0.7590799