Compute internal model metric(s)

Compute one or more internal metrics for the given lcModel object.

Call getInternalMetricNames() to retrieve the names of the defined internal metrics.

# S4 method for lcModel
metric(
  object,
  name = getOption("latrend.metric", c("WRSS", "APPA", "AIC", "BIC")),
  ...
)

# S4 method for list
metric(object, name, drop = TRUE)

# S4 method for lcModels
metric(object, name, drop = TRUE)

Arguments

object
The lcModel, lcModels, or list of lcModel objects to compute the metrics for.

name
The name(s) of the metric(s) to compute. If no names are given, the names specified in the latrend.metric option (WRSS, APPA, AIC, BIC) are used.

...
Additional arguments.

drop
Whether to return a numeric vector instead of a data.frame in case of a single metric.

Value

For metric(lcModel): A named numeric vector with the computed model metrics.

For metric(list): A data.frame with a metric per column.

For metric(lcModels): A data.frame with a metric per column.

See also

Examples

data(latrendData)
model <- latrend(lcMethodLcmmGMM(fixed = Y ~ Time, mixture = ~ Time,
   id = "Id", time = "Time"), latrendData)

#> Be patient, hlme is running ... 
#> The program took 0.09 seconds 
#> Be patient, hlme is running ... 
#> The program took 0.29 seconds 

bic <- metric(model, "BIC")

ic <- metric(model, c("AIC", "BIC"))