Prepare environment for fitting

Perform preparatory work that is needed for fitting the method, but should not be counted towards the method estimation time. The work is added to the provided environment, allowing the fit() function to make use of the prepared work.

# S4 method for lcMethod
preFit(method, data, envir, verbose)

Arguments

method
An object inheriting from lcMethod with all its arguments having been evaluated and finalized.

data
A data.frame representing the transformed training data.

envir
The environment containing additional data variables returned by prepareData().

verbose
A R.utils::Verbose object indicating the level of verbosity.

Value

The updated environment that will be passed to fit().

Implementation

setMethod("preFit", "lcMethodExample", function(method, data, envir, verbose) {
  # update envir with additional computed work
  envir$x <- INTENSIVE_OPERATION
  return(envir)
})

Fitting procedure

Each lcMethod subclass defines a type of methods in terms of a series of steps for estimating the method. These steps, as part of the fitting procedure, are executed by latrend() in the following order:

compose(): Evaluate and finalize the method argument values.

validate(): Check the validity of the method argument values in relation to the dataset.

prepareData(): Process the training data for fitting.

preFit(): Prepare environment for estimation, independent of training data.

fit(): Estimate the specified method on the training data, outputting an object inheriting from lcModel.

postFit(): Post-process the outputted lcModel object.

The result of the fitting procedure is an lcModel object that inherits from the lcModel class.