Extract the response variable

Extracts the response variable from the given object.

# S4 method for lcMethod
responseVariable(object, ...)

# S4 method for lcModel
responseVariable(object, ...)

Arguments

object

The object to extract the response variable from.

...

Additional arguments.

Value

The response variable name as a character.

Details

If the lcMethod object specifies a formula argument, then the response is extracted from the response term of the formula.

See also

Examples

method <- lcMethodKML("Value")
responseVariable(method) # "Value"
#> [1] "Value"

method <- lcMethodLcmmGBTM(fixed = Value ~ Time, mixture = ~ Time)
responseVariable(method) # "Value"
#> [1] "Value"

data(latrendData)
model <- latrend(lcMethodKML("Y", id = "Id", time = "Time"), latrendData)
#>  ~ Fast KmL ~
#> ********************S
responseVariable(model) # "Value"
#> [1] "Y"