Reduce the lcModel memory footprint for serialization

Strip a lcModel of non-essential variables and environments in order to reduce the model size for serialization.

# S4 method for lcMethod
strip(object, ..., classes = "formula")

# S4 method for ANY
strip(object, ..., classes = "formula")

# S4 method for lcModel
strip(object, ..., classes = "formula")

Arguments

object
The lcModel object.

...
Additional arguments.

classes
The object classes for which to remove their assigned environment. By default, only environments from formula are removed.

Value

An lcModel object of the same type as the object argument.

Implementation

Classes extending lcModel can override this method to remove additional non-essentials.

setMethod("strip", "lcModelExt", function(object, ..., classes = "formula") {
  object <- callNextMethod()
  # further process the object
  return (object)
})

Examples

data(latrendData)
model <- latrend(lcMethodLcmmGMM(fixed = Y ~ Time, mixture = ~ Time,
   id = "Id", time = "Time"), data = latrendData)

#> Be patient, hlme is running ... 
#> The program took 0.07 seconds 
#> Be patient, hlme is running ... 
#> The program took 0.21 seconds 

strip(model)

#> Longitudinal cluster model using growth mixture model
#> lcMethodLcmmGMM as "growth mixture model"
#>  idiag:          FALSE
#>  nwg:            FALSE
#>  cor:            NULL
#>  convB:          1e-04
#>  convL:          1e-04
#>  convG:          1e-04
#>  maxiter:        500
#>  na.action:      1
#>  posfix:         NULL
#>  fixed:          Y ~ Time
#>  mixture:        ~Time
#>  random:         ~1
#>  classmb:        ~1
#>  time:           "Time"
#>  id:             "Id"
#>  init:           "lme.random"
#>  nClusters:      2
#> 
#> Cluster sizes (K=2):
#>         A         B 
#>  90 (45%) 110 (55%) 
#> 
#> Number of obs: 2000, strata (Id): 200
#> 
#> Scaled residuals:
#>     Min.  1st Qu.   Median     Mean  3rd Qu.     Max. 
#> -4.73994 -0.58402  0.08016  0.00000  0.69619  2.93434 
#>