Extract the cluster trajectories

Extracts a data frame of all cluster trajectories.

# S4 method for lcModel
clusterTrajectories(object, at = time(object), what = "mu", ...)

Arguments

object

The lcModel object.

at

An optional vector of the times at which to compute the cluster trajectory predictions.

what

The distributional parameter to predict. By default, the mean response 'mu' is predicted. The cluster membership predictions can be obtained by specifying what = 'mb'.

...

Additional arguments.

Value

A data.frame of the estimated values at the given times. The first column should be named "Cluster". The second column should be time, with the name matching the timeVariable(object). The third column should be the expected value of the observations, named after the responseVariable(object).

See also

Examples

model <- latrend(method = lcMethodLcmmGMM(fixed = Y ~ Time, mixture = fixed),
  id = "Id", time = "Time", data = latrendData)
#> Be patient, hlme is running ... 
#> The program took 0.08 seconds 
#> Be patient, hlme is running ... 
#> The program took 0.33 seconds 
clusterTrajectories(model)
#>    Cluster      Time            Y
#> 1        A 0.0000000  0.112810504
#> 2        A 0.2222222  0.082933615
#> 3        A 0.4444444  0.053056726
#> 4        A 0.6666667  0.023179837
#> 5        A 0.8888889 -0.006697052
#> 6        A 1.1111111 -0.036573942
#> 7        A 1.3333333 -0.066450831
#> 8        A 1.5555556 -0.096327720
#> 9        A 1.7777778 -0.126204609
#> 10       A 2.0000000 -0.156081498
#> 11       B 0.0000000 -0.729474365
#> 12       B 0.2222222 -0.574460401
#> 13       B 0.4444444 -0.419446436
#> 14       B 0.6666667 -0.264432472
#> 15       B 0.8888889 -0.109418508
#> 16       B 1.1111111  0.045595456
#> 17       B 1.3333333  0.200609420
#> 18       B 1.5555556  0.355623385
#> 19       B 1.7777778  0.510637349
#> 20       B 2.0000000  0.665651313

clusterTrajectories(model, at = c(0, .5, 1))
#>   Cluster Time           Y
#> 1       A  0.0  0.11281050
#> 2       A  0.5  0.04558750
#> 3       A  1.0 -0.02163550
#> 4       B  0.0 -0.72947436
#> 5       B  0.5 -0.38069295
#> 6       B  1.0 -0.03191153