Compute the posterior confusion matrix

Compute the posterior confusion matrix (PCM). The entry \((i,j)\) represents the probability (or number, in case of scale = TRUE) of a trajectory belonging to cluster \(i\) is assigned to cluster \(j\) under the specified trajectory cluster assignment strategy.

confusionMatrix(object, strategy = which.max, scale = TRUE, ...)

Arguments

object

The model, of type lcModel.

strategy

The strategy for assigning trajectories to a specific cluster, see trajectoryAssignments(). If strategy = NULL, the posterior probabilities are used as weights (analogous to a repeated evaluation of strategy = which.weight).

scale

Whether to express the confusion in probabilities (scale = TRUE), or in terms of the number of trajectories.

...

Arguments passed on to trajectoryAssignments

Value

A K-by-K confusion matrix with K = nClusters(object).

See also

Examples

data(latrendData)

if (rlang::is_installed("lcmm")) {
  method <- lcMethodLcmmGMM(
    fixed = Y ~ Time,
    mixture = ~ Time,
    random = ~ 1,
    id = "Id",
    time = "Time"
  )
  model <- latrend(method, latrendData)
  confusionMatrix(model)
}
#> ------------------ iteration  0 ------------------
#> Function value  -2211.674 
#> Convergence criteria: parameters stability= 1.0001 
#>                     : function stability= 1.0001 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef   SE.coef    Var.coef
#> parameter1 -0.03295723  3.671255    13.47811
#> parameter2  0.00000000  4.663262    21.74601
#> parameter3  1.00000000 11.298389  -127.65358
#> parameter4  1.00000000 31.748083 -1007.94077
#> 
#> ------------------ iteration  1 ------------------
#> Function value  -2211.661 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 0.01365446 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef   SE.coef    Var.coef
#> parameter1 -0.03295781  3.672703    13.48875
#> parameter2  0.00000137  5.416630    29.33988
#> parameter3  0.99999000 11.298556  -127.65737
#> parameter4  0.99999246 31.748115 -1007.94279
#> 
#> ------------------ iteration  2 ------------------
#> Function value  -2211.643 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 0.01733755 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef   SE.coef    Var.coef
#> parameter1 -0.03295808  3.671208    13.47777
#> parameter2  0.00000233  4.946477    24.46763
#> parameter3  0.99998055 11.298637  -127.65919
#> parameter4  0.99998246 31.748227 -1007.94995
#> 
#> ------------------ iteration  3 ------------------
#> Function value  -1439.862 
#> Convergence criteria: parameters stability= 0.280072 
#>                     : function stability= 771.7816 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef   SE.coef   Var.coef
#> parameter1 -0.04031975  4.568936   20.87517
#> parameter2  0.04576697  6.864072   47.11549
#> parameter3  0.64986537 15.463936 -239.13331
#> parameter4  0.60584696 26.426104  698.33896
#> 
#> ------------------ iteration  4 ------------------
#> Function value  -1439.841 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 0.02074656 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef   SE.coef   Var.coef
#> parameter1 -0.04032027  4.569053   20.87625
#> parameter2  0.04576783  6.864238   47.11776
#> parameter3  0.64985797 15.464033 -239.13633
#> parameter4  0.60583696 26.429943  698.54191
#> 
#> ------------------ iteration  5 ------------------
#> Function value  -1259.308 
#> Convergence criteria: parameters stability= 0.1423089 
#>                     : function stability= 180.5326 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef    SE.coef     Var.coef
#> parameter1 -0.05020431   5.429509     29.47957
#> parameter2  0.06607079  10.206393    104.17045
#> parameter3  0.47182147  16.189455   -262.09845
#> parameter4  0.27402077 328.006561 107588.30435
#> 
#> ------------------ iteration  6 ------------------
#> Function value  -1182.724 
#> Convergence criteria: parameters stability= 1.186942 
#>                     : function stability= 76.58411 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef    SE.coef    Var.coef
#> parameter1 -0.17261108   4.785824    22.90411
#> parameter2  0.09024304   9.785319    95.75248
#> parameter3 -0.61021076  17.228002  -296.80404
#> parameter4  0.29811228 269.571111 72668.58395
#> 
#> ------------------ iteration  7 ------------------
#> Function value  -1182.721 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 0.00300208 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef    SE.coef    Var.coef
#> parameter1 -0.17261028   4.785870    22.90455
#> parameter2  0.09024312   9.785316    95.75240
#> parameter3 -0.61020076  17.228015  -296.80448
#> parameter4  0.29811233 269.571036 72668.54354
#> 
#> ------------------ iteration  8 ------------------
#> Function value  -1182.718 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 0.00380941 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef    SE.coef    Var.coef
#> parameter1 -0.17260961   4.785916    22.90499
#> parameter2  0.09024344   9.785311    95.75231
#> parameter3 -0.61019076  17.228004  -296.80414
#> parameter4  0.29811260 269.570458 72668.23197
#> 
#> ------------------ iteration  9 ------------------
#> Function value  -1070.899 
#> Convergence criteria: parameters stability= 0.05788247 
#>                     : function stability= 111.819 
#>                     : relative distance to maximum(RDM)= 735.0503 
#>                  coef    SE.coef   Var.coef
#> parameter1 -0.1578944 0.03390969 0.00114987
#> parameter2  0.1032584 0.01088495 0.00011848
#> parameter3 -0.3706614 0.15798998 0.02496083
#> parameter4  0.3091681 0.00406641 0.00001654
#> 
#> ------------------ iteration  10 ------------------
#> Function value  -1070.895 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 0.00369269 
#>                     : relative distance to maximum(RDM)= 734.1069 
#>                  coef    SE.coef   Var.coef
#> parameter1 -0.1578937 0.03390330 0.00114943
#> parameter2  0.1032586 0.01088495 0.00011848
#> parameter3 -0.3706514 0.15788498 0.02492767
#> parameter4  0.3091682 0.00406640 0.00001654
#> 
#> ------------------ iteration  11 ------------------
#> Function value  -1070.891 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 0.00379712 
#>                     : relative distance to maximum(RDM)= 732.7804 
#>                  coef    SE.coef   Var.coef
#> parameter1 -0.1578932 0.03389448 0.00114884
#> parameter2  0.1032587 0.01088496 0.00011848
#> parameter3 -0.3706414 0.15773736 0.02488108
#> parameter4  0.3091683 0.00406640 0.00001654
#> 
#> ------------------ iteration  12 ------------------
#> Function value  -1044.445 
#> Convergence criteria: parameters stability= 0.5764923 
#>                     : function stability= 26.44575 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                  coef    SE.coef    Var.coef
#> parameter1 -0.1270200   5.938905    35.27059
#> parameter2  0.1190798   9.395830    88.28163
#> parameter3  0.3877042   6.559355   -43.02514
#> parameter4  0.3233399 222.402347 49462.80385
#> 
#> ------------------ iteration  13 ------------------
#> Function value  -1044.442 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 0.00348247 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                  coef    SE.coef    Var.coef
#> parameter1 -0.1270206   5.938986    35.27155
#> parameter2  0.1190798   9.395828    88.28158
#> parameter3  0.3876942   6.555280   -42.97169
#> parameter4  0.3233399 222.402234 49462.75351
#> 
#> ------------------ iteration  14 ------------------
#> Function value  -1044.438 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 0.00354728 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                  coef    SE.coef    Var.coef
#> parameter1 -0.1270210   5.939060    35.27243
#> parameter2  0.1190798   9.395827    88.28156
#> parameter3  0.3876842   6.551120   -42.91718
#> parameter4  0.3233400 222.402055 49462.67395
#> 
#> ------------------ iteration  15 ------------------
#> Function value  -1044.435 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 0.00359666 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                  coef    SE.coef    Var.coef
#> parameter1 -0.1270214   5.939133    35.27331
#> parameter2  0.1190799   9.395822    88.28146
#> parameter3  0.3876742   6.547073   -42.86417
#> parameter4  0.3233402 222.401858 49462.58657
#> 
#> ------------------ iteration  16 ------------------
#> Function value  -990.0216 
#> Convergence criteria: parameters stability= 0.03537843 
#>                     : function stability= 54.41312 
#>                     : relative distance to maximum(RDM)= 12.63568 
#>                  coef    SE.coef   Var.coef
#> parameter1 -0.1312689 0.01972446 0.00038905
#> parameter2  0.1205435 0.01145365 0.00013119
#> parameter3  0.1996608 0.01309509 0.00017148
#> parameter4  0.3263760 0.00460804 0.00002123
#> 
#> ------------------ iteration  17 ------------------
#> Function value  -973.447 
#> Convergence criteria: parameters stability= 0.00032527 
#>                     : function stability= 16.5746 
#>                     : relative distance to maximum(RDM)= 5.923468 
#>                  coef    SE.coef   Var.coef
#> parameter1 -0.1424467 0.01988305 0.00039534
#> parameter2  0.1250521 0.01190601 0.00014175
#> parameter3  0.1966762 0.01313650 0.00017257
#> parameter4  0.3394560 0.00505861 0.00002559
#> 
#> ------------------ iteration  18 ------------------
#> Function value  -961.3927 
#> Convergence criteria: parameters stability= 0.00105038 
#>                     : function stability= 12.05426 
#>                     : relative distance to maximum(RDM)= 0.593486 
#>                  coef    SE.coef   Var.coef
#> parameter1 -0.1651658 0.02005362 0.00040215
#> parameter2  0.1362990 0.01255911 0.00015773
#> parameter3  0.1897438 0.01300009 0.00016900
#> parameter4  0.3584212 0.00577301 0.00003333
#> 
#> ------------------ iteration  19 ------------------
#> Function value  -960.1702 
#> Convergence criteria: parameters stability= 0.00024153 
#>                     : function stability= 1.222577 
#>                     : relative distance to maximum(RDM)= 0.00468292 
#>                  coef    SE.coef   Var.coef
#> parameter1 -0.1762450 0.02016492 0.00040662
#> parameter2  0.1435649 0.01282592 0.00016450
#> parameter3  0.1871356 0.01299865 0.00016896
#> parameter4  0.3661144 0.00608560 0.00003703
#> 
#> ------------------ iteration  20 ------------------
#> Function value  -960.1608 
#> Convergence criteria: parameters stability= 2.61e-06 
#>                     : function stability= 0.00939256 
#>                     : relative distance to maximum(RDM)= 5.3e-07 
#>                  coef    SE.coef   Var.coef
#> parameter1 -0.1773901 0.02017775 0.00040714
#> parameter2  0.1444374 0.01285068 0.00016514
#> parameter3  0.1869344 0.01300311 0.00016908
#> parameter4  0.3668221 0.00611506 0.00003739
#> 
#> ------------------ iteration  21 ------------------
#> Function value  -960.1608 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 1.06e-06 
#>                     : relative distance to maximum(RDM)= 0 
#>                  coef    SE.coef   Var.coef
#> parameter1 -0.1774069 0.02017785 0.00040715
#> parameter2  0.1444497 0.01285090 0.00016515
#> parameter3  0.1869323 0.01300313 0.00016908
#> parameter4  0.3668285 0.00611533 0.00003740
#> 
#> The program took 0.44 seconds 
#> ------------------ iteration  0 ------------------
#> Function value  -959.2723 
#> Convergence criteria: parameters stability= 1.0001 
#>                     : function stability= 1.0001 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                  coef   SE.coef  Var.coef
#> parameter1  0.0000000   0.00000     0.000
#> parameter2 -0.1874958  32.56135  1060.242
#> parameter3 -0.1673180  32.52617  1057.952
#> parameter4  0.1380242  86.33462 -7453.666
#> parameter5  0.1508751  86.34873 -7456.103
#> parameter6  0.1869323  76.50034  5852.302
#> parameter7  0.3668285 163.79515 26828.850
#> 
#> ------------------ iteration  1 ------------------
#> Function value  -957.172 
#> Convergence criteria: parameters stability= 6.145e-05 
#>                     : function stability= 2.100332 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                  coef    SE.coef     Var.coef
#> parameter1  0.0000005   2.873858     8.259062
#> parameter2 -0.1875478   5.743915    32.992561
#> parameter3 -0.1672656   3.956203    15.651543
#> parameter4  0.1324900  82.608245 -6824.122064
#> parameter5  0.1564099  82.612867 -6824.885748
#> parameter6  0.1867822  76.106068  5792.133534
#> parameter7  0.3664314 163.242233 26648.026657
#> 
#> ------------------ iteration  2 ------------------
#> Function value  -939.807 
#> Convergence criteria: parameters stability= 0.00087023 
#>                     : function stability= 17.36502 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef    SE.coef     Var.coef
#> parameter1  0.00005488   2.596824     6.743496
#> parameter2 -0.18764726   5.950719    35.411057
#> parameter3 -0.16715930   5.249387    27.556063
#> parameter4  0.11164707  54.870240 -3010.743218
#> parameter5  0.17725207  54.475787 -2967.611335
#> parameter6  0.18612558  74.101022  5490.961449
#> parameter7  0.36545858 157.038030 24660.942894
#> 
#> ------------------ iteration  3 ------------------
#> Function value  -920.3918 
#> Convergence criteria: parameters stability= 0.00046485 
#>                     : function stability= 19.41517 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef    SE.coef    Var.coef
#> parameter1 -0.00008149   3.371748   -11.36868
#> parameter2 -0.18638160  37.133125  1378.86898
#> parameter3 -0.16848824  36.505867  1332.67832
#> parameter4  0.09655969  20.983783  -440.31915
#> parameter5  0.19251663  20.000820  -400.03278
#> parameter6  0.18589413  73.030505  5333.45463
#> parameter7  0.36457657 151.289634 22888.55342
#> 
#> ------------------ iteration  4 ------------------
#> Function value  -870.4228 
#> Convergence criteria: parameters stability= 0.00225473 
#>                     : function stability= 49.96899 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef    SE.coef    Var.coef
#> parameter1 -0.00013076   6.743496   -45.47474
#> parameter2 -0.18534977  39.147602  1532.53477
#> parameter3 -0.16962638  37.912634  1437.36785
#> parameter4  0.06420283  53.303305  2841.24235
#> parameter5  0.22723004  50.873768  2588.14026
#> parameter6  0.18585681  71.701008  5141.03451
#> parameter7  0.36396508 138.031449 19052.68093
#> 
#> ------------------ iteration  5 ------------------
#> Function value  -807.3687 
#> Convergence criteria: parameters stability= 0.00288987 
#>                     : function stability= 63.05413 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef    SE.coef     Var.coef
#> parameter1  0.00154924   2.308902     5.331028
#> parameter2 -0.17307886   6.325157    40.007609
#> parameter3 -0.18296970   6.160603    37.953035
#> parameter4  0.03017905   6.362331    40.479255
#> parameter5  0.26391804   4.349280    18.916235
#> parameter6  0.18744110   7.997235    63.955775
#> parameter7  0.35673374 131.790811 17368.817851
#> 
#> ------------------ iteration  6 ------------------
#> Function value  -234.0025 
#> Convergence criteria: parameters stability= 0.2354196 
#>                     : function stability= 573.3662 
#>                     : relative distance to maximum(RDM)= 88.25512 
#>                   coef    SE.coef   Var.coef
#> parameter1  0.15233864 0.14201172 0.02016733
#> parameter2  0.03617987 0.02992324 0.00089540
#> parameter3 -0.40084166 0.03903369 0.00152363
#> parameter4 -0.17597225 0.01136866 0.00012925
#> parameter5  0.51667596 0.01643502 0.00027011
#> parameter6  0.24031620 0.01862848 0.00034702
#> parameter7  0.24608093 0.00589851 0.00003479
#> 
#> ------------------ iteration  7 ------------------
#> Function value  -116.8424 
#> Convergence criteria: parameters stability= 0.2750103 
#>                     : function stability= 117.1601 
#>                     : relative distance to maximum(RDM)= 14.02445 
#>                  coef    SE.coef   Var.coef
#> parameter1  0.3350642 0.14341104 0.02056673
#> parameter2  0.2213973 0.01957776 0.00038329
#> parameter3 -0.7824800 0.02390620 0.00057151
#> parameter4 -0.2305114 0.00800543 0.00006409
#> parameter5  0.7425480 0.00987910 0.00009760
#> parameter6  0.1864792 0.00860017 0.00007396
#> parameter7  0.1769633 0.00238969 0.00000571
#> 
#> ------------------ iteration  8 ------------------
#> Function value  -60.66954 
#> Convergence criteria: parameters stability= 0.00715973 
#>                     : function stability= 56.17281 
#>                     : relative distance to maximum(RDM)= 1.323995 
#>                  coef    SE.coef   Var.coef
#> parameter1  0.3942235 0.14417829 0.02078738
#> parameter2  0.1995742 0.02137132 0.00045673
#> parameter3 -0.7462780 0.02617159 0.00068495
#> parameter4 -0.2256329 0.00888236 0.00007890
#> parameter5  0.7082258 0.01088404 0.00011846
#> parameter6  0.2036181 0.01062426 0.00011287
#> parameter7  0.1963930 0.00305454 0.00000933
#> 
#> ------------------ iteration  9 ------------------
#> Function value  -55.73164 
#> Convergence criteria: parameters stability= 0.00050486 
#>                     : function stability= 4.937899 
#>                     : relative distance to maximum(RDM)= 0.01668516 
#>                  coef    SE.coef   Var.coef
#> parameter1  0.4052535 0.14433504 0.02083260
#> parameter2  0.1937467 0.02206570 0.00048689
#> parameter3 -0.7346090 0.02702485 0.00073034
#> parameter4 -0.2240926 0.00926405 0.00008582
#> parameter5  0.6980466 0.01134615 0.00012874
#> parameter6  0.2096016 0.01144289 0.00013094
#> parameter7  0.2048358 0.00338725 0.00001147
#> 
#> ------------------ iteration  10 ------------------
#> Function value  -55.67275 
#> Convergence criteria: parameters stability= 4.07e-06 
#>                     : function stability= 0.05888992 
#>                     : relative distance to maximum(RDM)= 3.06e-06 
#>                  coef    SE.coef   Var.coef
#> parameter1  0.4054646 0.14433800 0.02083346
#> parameter2  0.1932927 0.02213587 0.00049000
#> parameter3 -0.7334642 0.02711079 0.00073499
#> parameter4 -0.2239387 0.00931372 0.00008675
#> parameter5  0.6970418 0.01140694 0.00013012
#> parameter6  0.2101172 0.01152210 0.00013276
#> parameter7  0.2059342 0.00343271 0.00001178
#> 
#> ------------------ iteration  11 ------------------
#> Function value  -55.67274 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 1.071e-05 
#>                     : relative distance to maximum(RDM)= 0 
#>                  coef    SE.coef   Var.coef
#> parameter1  0.4054651 0.14433818 0.02083351
#> parameter2  0.1932891 0.02213651 0.00049003
#> parameter3 -0.7334510 0.02711158 0.00073504
#> parameter4 -0.2239371 0.00931442 0.00008676
#> parameter5  0.6970299 0.01140779 0.00013014
#> parameter6  0.2101202 0.01152271 0.00013277
#> parameter7  0.2059496 0.00343335 0.00001179
#> 
#> The program took 0.35 seconds 
#>              A            B
#> A 1.000000e+00 7.016739e-09
#> B 1.065267e-11 1.000000e+00