Compute the posterior confusion matrix

Compute the posterior confusion matrix (PCM). The entry \((i,j)\) represents the probability (or number, in case of scale = TRUE) of a trajectory belonging to cluster \(i\) is assigned to cluster \(j\) under the specified trajectory cluster assignment strategy.

confusionMatrix(object, strategy = which.max, scale = TRUE, ...)

Arguments

object

The model, of type lcModel.

strategy

The strategy for assigning trajectories to a specific cluster, see trajectoryAssignments(). If strategy = NULL, the posterior probabilities are used as weights (analogous to a repeated evaluation of strategy = which.weight).

scale

Whether to express the confusion in probabilities (scale = TRUE), or in terms of the number of trajectories.

...

Arguments passed on to trajectoryAssignments

Value

A K-by-K confusion matrix with K = nClusters(object).

See also

Examples

data(latrendData)
model = latrend(lcMethodLcmmGMM(
  fixed = Y ~ Time, mixture = ~ Time, random = ~ 1,
  id = "Id", time = "Time"),
  data=latrendData)
#> Be patient, hlme is running ... 
#> The program took 0.38 seconds 
confusionMatrix(model)
#>              A            B
#> A 9.998034e-01 0.0001965752
#> B 3.070110e-09 0.9999999969