lcMethod fit process: logic for fitting the method to the processed data

Note: this function should not be called directly, as it is part of the lcMethod fitting process. For fitting an lcMethod object to a dataset, see latrend().

The fit() function of the lcMethod object estimates the model with the evaluated method specification, processed training data, and prepared environment.

# S4 method for lcMethod
fit(method, data, envir, verbose)

Arguments

method

An object inheriting from lcMethod with all its arguments having been evaluated and finalized.

data

A data.frame representing the transformed training data.

envir

The environment containing variables generated by prepareData() and preFit().

verbose

A R.utils::Verbose object indicating the level of verbosity.

Value

The fitted object inheriting from lcModel.

Implementation

This method should be implemented for all lcMethod subclasses.

setMethod("fit", "lcMethodExample", function(method, data, envir, verbose) {
  # estimate the model or cluster parameters
  coefs <- FIT_CODE

  # create the lcModel object
  new("lcModelExample",
    method = method,
    data = data,
    model = coefs,
    clusterNames = make.clusterNames(method$nClusters)
  )
})

Fitting procedure

Each lcMethod subclass defines a type of methods in terms of a series of steps for estimating the method. These steps, as part of the fitting procedure, are executed by latrend() in the following order:

compose(): Evaluate and finalize the method argument values.

validate(): Check the validity of the method argument values in relation to the dataset.

prepareData(): Process the training data for fitting.

preFit(): Prepare environment for estimation, independent of training data.

fit(): Estimate the specified method on the training data, outputting an object inheriting from lcModel.

postFit(): Post-process the outputted lcModel object.

The result of the fitting procedure is an lcModel object that inherits from the lcModel class.