Extract the fitted trajectories for all strata

Extract the fitted trajectories for all strata

# S4 method for lcModel
fittedTrajectories(
  object,
  at = time(object),
  what = "mu",
  clusters = trajectoryAssignments(object),
  ...
)

Arguments

object

The model.

at

The time points at which to compute the id-specific trajectories. The default implementation merely filters the output of fitted(), so fitted values can only be outputted for times at which the model was trained.

what

The distributional parameter to compute the response for.

clusters

The cluster assignments for the strata to base the trajectories on.

...

Additional arguments.

Value

A data.frame representing the fitted response per trajectory per moment in time for the respective cluster.

Details

The default implementation uses the output of fitted() of the respective model.

See also

Examples

data(latrendData)
# Note: not a great example because the fitted trajectories
# are identical to the respective cluster trajectory
method <- lcMethodLMKM(Y ~ Time, id = "Id", time = "Time")
model <- latrend(method, latrendData)
fittedTrajectories(model)
#>        Id      Time           Y Cluster
#>    1:   1 0.0000000 -0.97975135       B
#>    2:   1 0.2222222 -0.63534769       B
#>    3:   1 0.4444444 -0.38525328       B
#>    4:   1 0.6666667 -0.16233773       B
#>    5:   1 0.8888889  0.05903586       B
#>   ---                                  
#> 1996: 200 1.1111111  0.01318545       A
#> 1997: 200 1.3333333 -0.05804444       A
#> 1998: 200 1.5555556 -0.13662404       A
#> 1999: 200 1.7777778 -0.24801613       A
#> 2000: 200 2.0000000 -0.34779054       A

fittedTrajectories(model, at = time(model)[c(1, 2)])
#>       Id      Time           Y Cluster
#>   1:   1 0.0000000 -0.97975135       B
#>   2:   1 0.2222222 -0.63534769       B
#>   3:   2 0.0000000 -0.97975135       B
#>   4:   2 0.2222222 -0.63534769       B
#>   5:   3 0.0000000 -0.97975135       B
#>  ---                                  
#> 396: 198 0.2222222  0.10921565       A
#> 397: 199 0.0000000  0.09202386       A
#> 398: 199 0.2222222  0.10921565       A
#> 399: 200 0.0000000  0.09202386       A
#> 400: 200 0.2222222  0.10921565       A