Method- and model-specific generics defined by the latrend package

List of S4 generic methods which have no general use other than supporting functions with signatures of lcMethod or lcModel.

Reduce the (serialized) memory footprint of an object.

clusterProportions(object, ...)

clusterTrajectories(object, ...)

compose(method, envir, ...)

converged(object, ...)

externalMetric(
  object,
  object2,
  name = getOption("latrend.externalMetric"),
  ...
)

estimationTime(object, unit = "secs", ...)

fit(method, data, envir, verbose, ...)

fittedTrajectories(
  object,
  at = time(object),
  what = "mu",
  clusters = trajectoryAssignments(object),
  ...
)

getArgumentDefaults(object, ...)

getArgumentExclusions(object, ...)

getLabel(object, ...)

getName(object, ...)

getShortName(object, ...)

idVariable(object, ...)

metric(object, name = getOption("latrend.metric", c("WRSS", "APPA.mean")), ...)

plotFittedTrajectories(object, ...)

plotClusterTrajectories(object, ...)

postFit(method, data, model, envir, verbose, ...)

postprob(object, ...)

predictAssignments(object, newdata = NULL, ...)

predictForCluster(object, newdata = NULL, cluster, ...)

predictPostprob(object, newdata = NULL, ...)

prepareData(method, data, verbose, ...)

preFit(method, data, envir, verbose, ...)

qqPlot(object, ...)

responseVariable(object, ...)

strip(object, ...)

timeVariable(object, ...)

trajectoryAssignments(object, ...)

validate(method, data, envir, ...)

# S4 method for `NULL`
getName(object, ...)

# S4 method for `NULL`
getShortName(object, ...)

Arguments

object

The object.

...

Arguments.

method

The method.

envir

environment.

object2

The model to compare with.

name

Metric name.

unit

Time unit.

data

data.frame.

verbose

R.utils::Verbose.

at

A vector of times at which to compute the cluster trajectories.

what

Parameter.

newdata

data.frame of newdata.

cluster

Cluster name.

Value

The stripped (i.e., updated) object.

Details

Serializing references to environments results in the serialization of the object together with any associated environments and references. This method removes those environments and references, greatly reducing the serialized object size.