Supported methods for longitudinal clustering

This page provides an overview of the currently supported methods for longitudinal clustering. For general recommendations on which method to apply to your dataset, see here.

Supported methods

Method	Description	Source

lcMethodAkmedoids	Anchored k-medoids (Adepeju et al. 2020)	akmedoids

lcMethodCrimCV	Group-based trajectory modeling of count data (Nielsen 2018)	crimCV

lcMethodDtwclust	Methods for distance-based clustering, including dynamic time warping (Sardá-Espinosa 2019)	dtwclust

lcMethodFeature	Feature-based clustering

lcMethodFlexmix	Interface to the FlexMix framework (Grün and Leisch 2008)	flexmix

lcMethodFlexmixGBTM	Group-based trajectory modeling	flexmix

lcMethodFunFEM	Model-based clustering using funFEM (Bouveyron 2015)	funFEM

lcMethodGCKM	Growth-curve modeling and k-means	lme4

lcMethodKML	Longitudinal k-means (Genolini et al. 2015)	kml

lcMethodLcmmGBTM	Group-based trajectory modeling (Proust-Lima et al. 2017)	lcmm

lcMethodLcmmGMM	Growth mixture modeling (Proust-Lima et al. 2017)	lcmm

lcMethodLMKM	Feature-based clustering using linear regression and k-means

lcMethodMclustLLPA	Longitudinal latent profile analysis (Scrucca et al. 2016)	mclust

lcMethodMixAK_GLMM	Mixture of generalized linear mixed models	mixAK

lcMethodMixtoolsGMM	Growth mixture modeling	mixtools

lcMethodMixtoolsNPRM	Non-parametric repeated measures clustering (Benaglia et al. 2009)	mixtools

lcMethodMixTVEM	Mixture of time-varying effects models

lcMethodRandom	Random partitioning

lcMethodStratify	Stratification rule

In addition, the functionality of any method can be extended via meta methods. This is used for extending the estimation procedure of a method, such as repeated fitting and selecting the best result, or fitting until convergence.

It is strongly encouraged to evaluate and compare several candidate methods in order to identify the most suitable method.

References

Adepeju M, Langton S, Bannister J (2020). akmedoids: Anchored Kmedoids for Longitudinal Data Clustering. R package version 0.1.5, https://CRAN.R-project.org/package=akmedoids.

Benaglia T, Chauveau D, Hunter DR, Young D (2009). “mixtools: An R Package for Analyzing Finite Mixture Models.” Journal of Statistical Software, 32(6), 1--29. doi:10.18637/jss.v032.i06 .

Bouveyron C (2015). funFEM: Clustering in the Discriminative Functional Subspace. R package version 1.1, https://CRAN.R-project.org/package=funFEM.

Genolini C, Alacoque X, Sentenac M, Arnaud C (2015). “kml and kml3d: R Packages to Cluster Longitudinal Data.” Journal of Statistical Software, 65(4), 1--34. doi:10.18637/jss.v065.i04 .

Grün B, Leisch F (2008). “FlexMix Version 2: Finite Mixtures with Concomitant Variables and Varying and Constant Parameters.” Journal of Statistical Software, 28(4), 1--35. doi:10.18637/jss.v028.i04 .

Nielsen JD (2018). crimCV: Group-Based Modelling of Longitudinal Data. R package version 0.9.6, https://CRAN.R-project.org/package=crimCV.

Proust-Lima C, Philipps V, Liquet B (2017). “Estimation of Extended Mixed Models Using Latent Classes and Latent Processes: The R Package lcmm.” Journal of Statistical Software, 78(2), 1--56. doi:10.18637/jss.v078.i02 .

Sardá-Espinosa A (2019). “Time-Series Clustering in R Using the dtwclust Package.” The R Journal. doi:10.32614/RJ-2019-023 .

Scrucca L, Fop M, Murphy TB, Raftery AE (2016). “mclust 5: clustering, classification and density estimation using Gaussian finite mixture models.” The R Journal, 8(1), 205--233.

See also

Examples

data(latrendData)
method <- lcMethodLMKM(Y ~ Time, id = "Id", time = "Time")
model <- latrend(method, data = latrendData)