Specify GBTM method

Group-based trajectory modeling through fixed-effects modeling.

lcMethodLcmmGBTM(
  fixed,
  mixture = ~1,
  classmb = ~1,
  time = getOption("latrend.time"),
  id = getOption("latrend.id"),
  nClusters = 2,
  init = "default",
  ...
)

Arguments

fixed

The fixed effects formula.

mixture

The mixture-specific effects formula. See lcmm::hlme for details.

classmb

The cluster membership formula for the multinomial logistic model. See lcmm::hlme for details.

time

The name of the time variable.

id

The name of the trajectory identifier variable. This replaces the subject argument of lcmm::hlme.

nClusters

The number of clusters to fit. This replaces the ng argument of lcmm::hlme.

init

Alternative for the B argument of lcmm::hlme, for initializing the hlme fitting procedure. If "lme.random" (default): random initialization through a standard linear mixed model. Assigns a fitted standard linear mixed model enclosed in a call to random() to the B argument. If "lme", fits a standard linear mixed model and passes this to the B argument. If NULL or "default", the default lcmm::hlme input for B is used.

The argument is ignored if the B argument is specified, or nClusters = 1.

...

Arguments passed to lcmm::hlme. The following arguments are ignored: data, fixed, random, mixture, subject, classmb, returndata, ng, verbose, subset.

References

Proust-Lima C, Philipps V, Liquet B (2017). “Estimation of Extended Mixed Models Using Latent Classes and Latent Processes: The R Package lcmm.” Journal of Statistical Software, 78(2), 1--56. doi:10.18637/jss.v078.i02 .

Proust-Lima C, Philipps V, Diakite A, Liquet B (2019). lcmm: Extended Mixed Models Using Latent Classes and Latent Processes. R package version: 1.8.1, https://cran.r-project.org/package=lcmm.

See also

Examples

data(latrendData)
if (rlang::is_installed("lcmm")) {
  method <- lcMethodLcmmGBTM(
    fixed = Y ~ Time,
    mixture = ~ 1,
   id = "Id",
   time = "Time",
   nClusters = 3
  )
  gbtm <- latrend(method, data = latrendData)
  summary(gbtm)

  method <- lcMethodLcmmGBTM(
    fixed = Y ~ Time,
    mixture = ~ Time,
    id = "Id",
    time = "Time",
    nClusters = 3
  )
}
#> ------------------ iteration  0 ------------------
#> Function value  -3305.592 
#> Convergence criteria: parameters stability= 1.0001 
#>                     : function stability= 1.0001 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef
#> parameter1 -0.03295723
#> parameter2  1.00000000
#> parameter3  1.00000000
#> 
#> ------------------ iteration  1 ------------------
#> Function value  -1734.439 
#> Convergence criteria: parameters stability= 0.4549852 
#>                     : function stability= 1571.153 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                  coef
#> parameter1 -0.4572794
#> parameter2  0.5232458
#> parameter3  0.7817312
#> 
#> ------------------ iteration  2 ------------------
#> Function value  -1734.421 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 0.01789124 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                  coef
#> parameter1 -0.4572795
#> parameter2  0.5232431
#> parameter3  0.7817212
#> 
#> ------------------ iteration  3 ------------------
#> Function value  -1514.243 
#> Convergence criteria: parameters stability= 2.098398 
#>                     : function stability= 220.1778 
#>                     : relative distance to maximum(RDM)= 4093.049 
#>                  coef
#> parameter1 -0.5198895
#> parameter2  0.2393904
#> parameter3 -0.6374000
#> 
#> ------------------ iteration  4 ------------------
#> Function value  -1514.22 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 0.02332214 
#>                     : relative distance to maximum(RDM)= 4090.921 
#>                  coef
#> parameter1 -0.5198816
#> parameter2  0.2393909
#> parameter3 -0.6373901
#> 
#> ------------------ iteration  5 ------------------
#> Function value  -1514.197 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 0.02332088 
#>                     : relative distance to maximum(RDM)= 4088.837 
#>                  coef
#> parameter1 -0.5198747
#> parameter2  0.2393924
#> parameter3 -0.6373800
#> 
#> ------------------ iteration  6 ------------------
#> Function value  -1383.633 
#> Convergence criteria: parameters stability= 0.155725 
#>                     : function stability= 130.5639 
#>                     : relative distance to maximum(RDM)= 149.4479 
#>                  coef
#> parameter1 -0.3091971
#> parameter2  0.3138116
#> parameter3 -0.3121082
#> 
#> ------------------ iteration  7 ------------------
#> Function value  -1140.381 
#> Convergence criteria: parameters stability= 0.00852576 
#>                     : function stability= 243.2518 
#>                     : relative distance to maximum(RDM)= 41.51194 
#>                  coef
#> parameter1 -0.2705361
#> parameter2  0.2424892
#> parameter3 -0.3562011
#> 
#> ------------------ iteration  8 ------------------
#> Function value  -1069.856 
#> Convergence criteria: parameters stability= 0.01032877 
#>                     : function stability= 70.52481 
#>                     : relative distance to maximum(RDM)= 4.209667 
#>                  coef
#> parameter1 -0.2060939
#> parameter2  0.1727871
#> parameter3 -0.3924999
#> 
#> ------------------ iteration  9 ------------------
#> Function value  -1063.11 
#> Convergence criteria: parameters stability= 0.00158953 
#>                     : function stability= 6.746472 
#>                     : relative distance to maximum(RDM)= 0.06222483 
#>                  coef
#> parameter1 -0.1803781
#> parameter2  0.1473150
#> parameter3 -0.4092151
#> 
#> ------------------ iteration  10 ------------------
#> Function value  -1063.015 
#> Convergence criteria: parameters stability= 2.256e-05 
#>                     : function stability= 0.09433224 
#>                     : relative distance to maximum(RDM)= 1.788e-05 
#>                  coef
#> parameter1 -0.1774510
#> parameter2  0.1444913
#> parameter3 -0.4116683
#> 
#> ------------------ iteration  11 ------------------
#> Function value  -1063.015 
#> Convergence criteria: parameters stability= 1e-08 
#>                     : function stability= 2.683e-05 
#>                     : relative distance to maximum(RDM)= 0 
#>                  coef
#> parameter1 -0.1774070
#> parameter2  0.1444497
#> parameter3 -0.4117121
#> 
#> The program took 0.17 seconds 
#> ------------------ iteration  0 ------------------
#> Function value  -1060.942 
#> Convergence criteria: parameters stability= 1.0001 
#>                     : function stability= 1.0001 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                  coef
#> parameter1  0.0000000
#> parameter2  0.0000000
#> parameter3 -0.1945179
#> parameter4 -0.1774070
#> parameter5 -0.1602960
#> parameter6  0.1444497
#> parameter7 -0.4117121
#> 
#> ------------------ iteration  1 ------------------
#> Function value  -1040.277 
#> Convergence criteria: parameters stability= 0.00439791 
#>                     : function stability= 20.6654 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef
#> parameter1  0.00023465
#> parameter2 -0.00045418
#> parameter3 -0.24152489
#> parameter4 -0.17692004
#> parameter5 -0.11377959
#> parameter6  0.14445231
#> parameter7 -0.40681583
#> 
#> ------------------ iteration  2 ------------------
#> Function value  -993.7736 
#> Convergence criteria: parameters stability= 0.03849861 
#>                     : function stability= 46.5031 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef
#> parameter1  0.00650092
#> parameter2 -0.01971082
#> parameter3 -0.32285769
#> parameter4 -0.21896099
#> parameter5  0.05840982
#> parameter6  0.14185636
#> parameter7 -0.39973532
#> 
#> ------------------ iteration  3 ------------------
#> Function value  -983.4672 
#> Convergence criteria: parameters stability= 0.05513769 
#>                     : function stability= 10.3064 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef
#> parameter1 -0.02714328
#> parameter2  0.20456900
#> parameter3 -0.36668886
#> parameter4 -0.18010908
#> parameter5  0.05940519
#> parameter6  0.14173240
#> parameter7 -0.38322232
#> 
#> ------------------ iteration  4 ------------------
#> Function value  -983.4671 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 0.0001445 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef
#> parameter1 -0.02714951
#> parameter2  0.20457900
#> parameter3 -0.36668985
#> parameter4 -0.18010934
#> parameter5  0.05940559
#> parameter6  0.14173237
#> parameter7 -0.38322226
#> 
#> ------------------ iteration  5 ------------------
#> Function value  -967.6235 
#> Convergence criteria: parameters stability= 2.02493 
#>                     : function stability= 15.84353 
#>                     : relative distance to maximum(RDM)= 1.131383 
#>                  coef
#> parameter1 -0.5824433
#> parameter2  1.5075284
#> parameter3 -0.4869535
#> parameter4 -0.2032992
#> parameter5  0.1204690
#> parameter6  0.1387673
#> parameter7 -0.3704416
#> 
#> ------------------ iteration  6 ------------------
#> Function value  -963.2659 
#> Convergence criteria: parameters stability= 0.0544265 
#>                     : function stability= 4.357653 
#>                     : relative distance to maximum(RDM)= 0.2916272 
#>                  coef
#> parameter1 -0.7906077
#> parameter2  1.5438908
#> parameter3 -0.5837541
#> parameter4 -0.2045180
#> parameter5  0.1397605
#> parameter6  0.1438441
#> parameter7 -0.3718892
#> 
#> ------------------ iteration  7 ------------------
#> Function value  -962.2846 
#> Convergence criteria: parameters stability= 0.4451523 
#>                     : function stability= 0.9812986 
#>                     : relative distance to maximum(RDM)= 0.00878651 
#>                  coef
#> parameter1 -1.4330421
#> parameter2  1.3909273
#> parameter3 -0.6746616
#> parameter4 -0.2230504
#> parameter5  0.1191933
#> parameter6  0.1442983
#> parameter7 -0.3731912
#> 
#> ------------------ iteration  8 ------------------
#> Function value  -962.2536 
#> Convergence criteria: parameters stability= 0.00174615 
#>                     : function stability= 0.03101885 
#>                     : relative distance to maximum(RDM)= 7.43e-06 
#>                  coef
#> parameter1 -1.4604543
#> parameter2  1.3611574
#> parameter3 -0.6849029
#> parameter4 -0.2227449
#> parameter5  0.1173345
#> parameter6  0.1444489
#> parameter7 -0.3733110
#> 
#> ------------------ iteration  9 ------------------
#> Function value  -962.2536 
#> Convergence criteria: parameters stability= 7.17e-06 
#>                     : function stability= 2.606e-05 
#>                     : relative distance to maximum(RDM)= 0 
#>                  coef
#> parameter1 -1.4630262
#> parameter2  1.3606610
#> parameter3 -0.6854437
#> parameter4 -0.2228127
#> parameter5  0.1172473
#> parameter6  0.1444497
#> parameter7 -0.3733210
#> 
#> The program took 0.42 seconds