Specify GBTM method

Group-based trajectory modeling through fixed-effects modeling.

lcMethodLcmmGBTM(
  fixed,
  mixture = ~1,
  classmb = ~1,
  time = getOption("latrend.time"),
  id = getOption("latrend.id"),
  nClusters = 2,
  init = "default",
  ...
)

Arguments

fixed

The fixed effects formula.

mixture

The mixture-specific effects formula. See lcmm::hlme for details.

classmb

The cluster membership formula for the multinomial logistic model. See lcmm::hlme for details.

time

The name of the time variable.

id

The name of the trajectory identifier variable. This replaces the subject argument of lcmm::hlme.

nClusters

The number of clusters to fit. This replaces the ng argument of lcmm::hlme.

init

Alternative for the B argument of lcmm::hlme, for initializing the hlme fitting procedure. If "lme.random" (default): random initialization through a standard linear mixed model. Assigns a fitted standard linear mixed model enclosed in a call to random() to the B argument. If "lme", fits a standard linear mixed model and passes this to the B argument. If NULL or "default", the default lcmm::hlme input for B is used.

The argument is ignored if the B argument is specified, or nClusters = 1.

...

Arguments passed to lcmm::hlme. The following arguments are ignored: data, fixed, random, mixture, subject, classmb, returndata, ng, verbose, subset.

References

Proust-Lima C, Philipps V, Liquet B (2017). “Estimation of Extended Mixed Models Using Latent Classes and Latent Processes: The R Package lcmm.” Journal of Statistical Software, 78(2), 1--56. doi: 10.18637/jss.v078.i02 .

Proust-Lima C, Philipps V, Diakite A, Liquet B (2019). lcmm: Extended Mixed Models Using Latent Classes and Latent Processes. R package version: 1.8.1, https://cran.r-project.org/package=lcmm.

See also

Examples

data(latrendData)
method <- lcMethodLcmmGBTM(fixed = Y ~ Time, mixture = ~ 1,
   id = "Id", time = "Time", nClusters = 3)
gbtm <- latrend(method, data = latrendData)
#> Be patient, hlme is running ... 
#> The program took 0.25 seconds 
summary(gbtm)
#> Longitudinal cluster model using group-based trajectory modeling using lcmm
#> lcMethodLcmmGBTM specifying "group-based trajectory modeling using lcmm"
#>  mixture:        ~1
#>  classmb:        ~1
#>  time:           "Time"
#>  id:             "Id"
#>  nClusters:      3
#>  init:           "default"
#>  idiag:          FALSE
#>  nwg:            FALSE
#>  cor:            NULL
#>  convB:          1e-04
#>  convL:          1e-04
#>  convG:          1e-04
#>  maxiter:        500
#>  subset:         NULL
#>  na.action:      1
#>  posfix:         NULL
#>  fixed:          Y ~ Time
#> 
#> Cluster sizes (K=3):
#>           A           B           C 
#>   13 (6.2%) 160 (78.3%)  27 (15.4%) 
#> 
#> Number of obs: 2000, strata (Id): 200
#> 
#> Scaled residuals:
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#> -4.2435 -0.5433  0.1265  0.0000  0.6868  2.5551 
#> 

method <- lcMethodLcmmGBTM(fixed = Y ~ Time, mixture = ~ Time,
    id = "Id", time = "Time", nClusters = 3)