Specify GBTM method

Group-based trajectory modeling through fixed-effects modeling.

lcMethodLcmmGBTM(
  fixed,
  mixture = ~1,
  classmb = ~1,
  time = getOption("latrend.time"),
  id = getOption("latrend.id"),
  nClusters = 2,
  init = "default",
  ...
)

Arguments

fixed

The fixed effects formula.

mixture

The mixture-specific effects formula. See lcmm::hlme for details.

classmb

The cluster membership formula for the multinomial logistic model. See lcmm::hlme for details.

time

The name of the time variable.

id

The name of the trajectory identifier variable. This replaces the subject argument of lcmm::hlme.

nClusters

The number of clusters to fit. This replaces the ng argument of lcmm::hlme.

init

Alternative for the B argument of lcmm::hlme, for initializing the hlme fitting procedure. If "lme.random" (default): random initialization through a standard linear mixed model. Assigns a fitted standard linear mixed model enclosed in a call to random() to the B argument. If "lme", fits a standard linear mixed model and passes this to the B argument. If NULL or "default", the default lcmm::hlme input for B is used.

The argument is ignored if the B argument is specified, or nClusters = 1.

...

Arguments passed to lcmm::hlme. The following arguments are ignored: data, fixed, random, mixture, subject, classmb, returndata, ng, verbose, subset.

References

Proust-Lima C, Philipps V, Liquet B (2017). “Estimation of Extended Mixed Models Using Latent Classes and Latent Processes: The R Package lcmm.” Journal of Statistical Software, 78(2), 1--56. doi:10.18637/jss.v078.i02 .

Proust-Lima C, Philipps V, Diakite A, Liquet B (2019). lcmm: Extended Mixed Models Using Latent Classes and Latent Processes. R package version: 1.8.1, https://cran.r-project.org/package=lcmm.

See also

Examples

data(latrendData)
if (rlang::is_installed("lcmm")) {
  method <- lcMethodLcmmGBTM(
    fixed = Y ~ Time,
    mixture = ~ 1,
   id = "Id",
   time = "Time",
   nClusters = 3
  )
  gbtm <- latrend(method, data = latrendData)
  summary(gbtm)

  method <- lcMethodLcmmGBTM(
    fixed = Y ~ Time,
    mixture = ~ Time,
    id = "Id",
    time = "Time",
    nClusters = 3
  )
}
#> ------------------ iteration  0 ------------------
#> Function value  -3305.592 
#> Convergence criteria: parameters stability= 1.0001 
#>                     : function stability= 1.0001 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef   SE.coef   Var.coef
#> parameter1 -0.03295723  6.687450   44.72198
#> parameter2  1.00000000  5.342837   28.54591
#> parameter3  1.00000000 82.481329 6803.16962
#> 
#> ------------------ iteration  1 ------------------
#> Function value  -1734.439 
#> Convergence criteria: parameters stability= 0.4549852 
#>                     : function stability= 1571.153 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                  coef   SE.coef  Var.coef
#> parameter1 -0.4572794  7.563604  57.20810
#> parameter2  0.5232458  6.042769  36.51505
#> parameter3  0.7817312 23.370405 546.17585
#> 
#> ------------------ iteration  2 ------------------
#> Function value  -1734.421 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 0.01789124 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                  coef   SE.coef  Var.coef
#> parameter1 -0.4572795  7.563653  57.20884
#> parameter2  0.5232431  6.042808  36.51553
#> parameter3  0.7817212 23.372253 546.26219
#> 
#> ------------------ iteration  3 ------------------
#> Function value  -1514.243 
#> Convergence criteria: parameters stability= 2.098398 
#>                     : function stability= 220.1778 
#>                     : relative distance to maximum(RDM)= 4093.049 
#>                  coef    SE.coef   Var.coef
#> parameter1 -0.5198895 0.05696642 0.00324517
#> parameter2  0.2393904 0.02634452 0.00069403
#> parameter3 -0.6374000 0.04692978 0.00220240
#> 
#> ------------------ iteration  4 ------------------
#> Function value  -1514.22 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 0.02332214 
#>                     : relative distance to maximum(RDM)= 4090.921 
#>                  coef    SE.coef   Var.coef
#> parameter1 -0.5198816 0.05695432 0.00324379
#> parameter2  0.2393909 0.02634241 0.00069392
#> parameter3 -0.6373901 0.04691769 0.00220127
#> 
#> ------------------ iteration  5 ------------------
#> Function value  -1514.197 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 0.02332088 
#>                     : relative distance to maximum(RDM)= 4088.837 
#>                  coef    SE.coef   Var.coef
#> parameter1 -0.5198747 0.05694269 0.00324247
#> parameter2  0.2393924 0.02634048 0.00069382
#> parameter3 -0.6373800 0.04690584 0.00220016
#> 
#> ------------------ iteration  6 ------------------
#> Function value  -1383.633 
#> Convergence criteria: parameters stability= 0.155725 
#>                     : function stability= 130.5639 
#>                     : relative distance to maximum(RDM)= 149.4479 
#>                  coef    SE.coef   Var.coef
#> parameter1 -0.3091971 0.01329500 0.00017676
#> parameter2  0.3138116 0.01155819 0.00013359
#> parameter3 -0.3121082 0.00345187 0.00001192
#> 
#> ------------------ iteration  7 ------------------
#> Function value  -1140.381 
#> Convergence criteria: parameters stability= 0.00852576 
#>                     : function stability= 243.2518 
#>                     : relative distance to maximum(RDM)= 41.51194 
#>                  coef    SE.coef   Var.coef
#> parameter1 -0.2705361 0.01499935 0.00022498
#> parameter2  0.2424892 0.01273472 0.00016217
#> parameter3 -0.3562011 0.00461531 0.00002130
#> 
#> ------------------ iteration  8 ------------------
#> Function value  -1069.856 
#> Convergence criteria: parameters stability= 0.01032877 
#>                     : function stability= 70.52481 
#>                     : relative distance to maximum(RDM)= 4.209667 
#>                  coef    SE.coef   Var.coef
#> parameter1 -0.2060939 0.01633440 0.00026681
#> parameter2  0.1727871 0.01377562 0.00018977
#> parameter3 -0.3924999 0.00578726 0.00003349
#> 
#> ------------------ iteration  9 ------------------
#> Function value  -1063.11 
#> Convergence criteria: parameters stability= 0.00158953 
#>                     : function stability= 6.746472 
#>                     : relative distance to maximum(RDM)= 0.06222483 
#>                  coef    SE.coef   Var.coef
#> parameter1 -0.1803781 0.01700743 0.00028925
#> parameter2  0.1473150 0.01433610 0.00020552
#> parameter3 -0.4092151 0.00641010 0.00004109
#> 
#> ------------------ iteration  10 ------------------
#> Function value  -1063.015 
#> Convergence criteria: parameters stability= 2.256e-05 
#>                     : function stability= 0.09433224 
#>                     : relative distance to maximum(RDM)= 1.788e-05 
#>                  coef    SE.coef   Var.coef
#> parameter1 -0.1774510 0.01710914 0.00029272
#> parameter2  0.1444913 0.01442175 0.00020799
#> parameter3 -0.4116683 0.00650638 0.00004233
#> 
#> ------------------ iteration  11 ------------------
#> Function value  -1063.015 
#> Convergence criteria: parameters stability= 1e-08 
#>                     : function stability= 2.683e-05 
#>                     : relative distance to maximum(RDM)= 0 
#>                  coef    SE.coef   Var.coef
#> parameter1 -0.1774070 0.01711095 0.00029278
#> parameter2  0.1444497 0.01442328 0.00020803
#> parameter3 -0.4117121 0.00650811 0.00004236
#> 
#> The program took 0.16 seconds 
#> ------------------ iteration  0 ------------------
#> Function value  -1060.942 
#> Convergence criteria: parameters stability= 1.0001 
#>                     : function stability= 1.0001 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                  coef    SE.coef    Var.coef
#> parameter1  0.0000000   3.265334    10.66240
#> parameter2  0.0000000   3.232179    10.44698
#> parameter3 -0.1945179  53.495770 -2861.79745
#> parameter4 -0.1774070  57.823727 -3343.58342
#> parameter5 -0.1602960  53.354566 -2846.70975
#> parameter6  0.1444497 128.349232 16473.52546
#> parameter7 -0.4117121 152.599555 23286.62429
#> 
#> ------------------ iteration  1 ------------------
#> Function value  -1040.277 
#> Convergence criteria: parameters stability= 0.00439791 
#>                     : function stability= 20.6654 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef    SE.coef     Var.coef
#> parameter1  0.00023465   3.672464    13.486992
#> parameter2 -0.00045418   2.843443     8.085167
#> parameter3 -0.24152489   5.298465    28.073731
#> parameter4 -0.17692004  38.629517 -1492.239572
#> parameter5 -0.11377959  21.784984   474.585538
#> parameter6  0.14445231 124.331144 15458.233451
#> parameter7 -0.40681583 149.108385 22233.310468
#> 
#> ------------------ iteration  2 ------------------
#> Function value  -993.7736 
#> Convergence criteria: parameters stability= 0.03849861 
#>                     : function stability= 46.5031 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef    SE.coef     Var.coef
#> parameter1  0.00650092   1.970094     3.881272
#> parameter2 -0.01971082   1.449795     2.101906
#> parameter3 -0.32285769  44.891869  2015.279875
#> parameter4 -0.21896099  39.799983  1584.038674
#> parameter5  0.05840982  44.373984  1969.050470
#> parameter6  0.14185636 102.985932 10606.102182
#> parameter7 -0.39973532 141.922797 20142.080172
#> 
#> ------------------ iteration  3 ------------------
#> Function value  -983.4672 
#> Convergence criteria: parameters stability= 0.05513769 
#>                     : function stability= 10.3064 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef    SE.coef     Var.coef
#> parameter1 -0.02714328   2.109782     4.451182
#> parameter2  0.20456900   1.829782     3.348103
#> parameter3 -0.36668886  45.889449  2105.841523
#> parameter4 -0.18010908  52.668051  2773.923545
#> parameter5  0.05940519  42.592876  1814.153094
#> parameter6  0.14173240 103.084869 10626.490219
#> parameter7 -0.38322232 155.329897 24127.376955
#> 
#> ------------------ iteration  4 ------------------
#> Function value  -983.4671 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 0.0001445 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef    SE.coef     Var.coef
#> parameter1 -0.02714951   2.111591     4.458815
#> parameter2  0.20457900   1.829640     3.347583
#> parameter3 -0.36668985  45.889162  2105.815208
#> parameter4 -0.18010934  52.668355  2773.955670
#> parameter5  0.05940559  42.593116  1814.173494
#> parameter6  0.14173237 103.084854 10626.487068
#> parameter7 -0.38322226 155.329908 24127.380233
#> 
#> ------------------ iteration  5 ------------------
#> Function value  -967.6235 
#> Convergence criteria: parameters stability= 2.02493 
#>                     : function stability= 15.84353 
#>                     : relative distance to maximum(RDM)= 1.131383 
#>                  coef    SE.coef   Var.coef
#> parameter1 -0.5824433 0.66324932 0.43989966
#> parameter2  1.5075284 0.37354755 0.13953777
#> parameter3 -0.4869535 0.05033031 0.00253314
#> parameter4 -0.2032992 0.02504848 0.00062743
#> parameter5  0.1204690 0.04188773 0.00175458
#> parameter6  0.1387673 0.01297878 0.00016845
#> parameter7 -0.3704416 0.00601708 0.00003621
#> 
#> ------------------ iteration  6 ------------------
#> Function value  -963.2659 
#> Convergence criteria: parameters stability= 0.0544265 
#>                     : function stability= 4.357653 
#>                     : relative distance to maximum(RDM)= 0.2916272 
#>                  coef    SE.coef   Var.coef
#> parameter1 -0.7906077 0.66538638 0.44273904
#> parameter2  1.5438908 0.35152426 0.12356930
#> parameter3 -0.5837541 0.07194599 0.00517622
#> parameter4 -0.2045180 0.02391926 0.00057213
#> parameter5  0.1397605 0.04464294 0.00199299
#> parameter6  0.1438441 0.01302819 0.00016973
#> parameter7 -0.3718892 0.00604106 0.00003649
#> 
#> ------------------ iteration  7 ------------------
#> Function value  -962.2846 
#> Convergence criteria: parameters stability= 0.4451523 
#>                     : function stability= 0.9812986 
#>                     : relative distance to maximum(RDM)= 0.00878651 
#>                  coef    SE.coef   Var.coef
#> parameter1 -1.4330421 0.64318492 0.41368684
#> parameter2  1.3909273 0.35074760 0.12302388
#> parameter3 -0.6746616 0.07477559 0.00559139
#> parameter4 -0.2230504 0.02275129 0.00051762
#> parameter5  0.1191933 0.04298579 0.00184778
#> parameter6  0.1442983 0.01307380 0.00017092
#> parameter7 -0.3731912 0.00609611 0.00003716
#> 
#> ------------------ iteration  8 ------------------
#> Function value  -962.2536 
#> Convergence criteria: parameters stability= 0.00174615 
#>                     : function stability= 0.03101885 
#>                     : relative distance to maximum(RDM)= 7.43e-06 
#>                  coef    SE.coef   Var.coef
#> parameter1 -1.4604543 0.63759497 0.40652734
#> parameter2  1.3611574 0.34502346 0.11904119
#> parameter3 -0.6849029 0.07741063 0.00599241
#> parameter4 -0.2227449 0.02271041 0.00051576
#> parameter5  0.1173345 0.04260285 0.00181500
#> parameter6  0.1444489 0.01307802 0.00017103
#> parameter7 -0.3733110 0.00610239 0.00003724
#> 
#> ------------------ iteration  9 ------------------
#> Function value  -962.2536 
#> Convergence criteria: parameters stability= 7.17e-06 
#>                     : function stability= 2.606e-05 
#>                     : relative distance to maximum(RDM)= 0 
#>                  coef    SE.coef   Var.coef
#> parameter1 -1.4630262 0.63798184 0.40702082
#> parameter2  1.3606610 0.34504755 0.11905781
#> parameter3 -0.6854437 0.07752316 0.00600984
#> parameter4 -0.2228127 0.02271418 0.00051593
#> parameter5  0.1172473 0.04259761 0.00181456
#> parameter6  0.1444497 0.01307835 0.00017104
#> parameter7 -0.3733210 0.00610291 0.00003725
#> 
#> The program took 0.34 seconds