Specify GMM method using lcmm

Growth mixture modeling through latent-class linear mixed modeling.

lcMethodLcmmGMM(
  fixed,
  mixture = ~1,
  random = ~1,
  classmb = ~1,
  time = getOption("latrend.time"),
  id = getOption("latrend.id"),
  init = "lme.random",
  nClusters = 2,
  ...
)

Arguments

fixed
The fixed effects formula.

mixture
The mixture-specific effects formula. See lcmm::hlme for details.

random
The random effects formula. See lcmm::hlme for details.

classmb
The cluster membership formula for the multinomial logistic model. See lcmm::hlme for details.

time
The name of the time variable.

id
The name of the trajectory identifier variable. This replaces the subject argument of lcmm::hlme.

init
Alternative for the B argument of lcmm::hlme, for initializing the hlme fitting procedure. If "lme.random" (default): random initialization through a standard linear mixed model. Assigns a fitted standard linear mixed model enclosed in a call to random() to the B argument. If "lme", fits a standard linear mixed model and passes this to the B argument. If NULL or "default", the default lcmm::hlme input for B is used.

The argument is ignored if the B argument is specified, or nClusters = 1.

nClusters
The number of clusters to fit. This replaces the ng argument of lcmm::hlme.

...
Arguments passed to lcmm::hlme. The following arguments are ignored: data, fixed, random, mixture, subject, classmb, returndata, ng, verbose, subset.

References

Proust-Lima C, Philipps V, Liquet B (2017). “Estimation of Extended Mixed Models Using Latent Classes and Latent Processes: The R Package lcmm.” Journal of Statistical Software, 78(2), 1--56. doi: 10.18637/jss.v078.i02 .

Proust-Lima C, Philipps V, Diakite A, Liquet B (2019). lcmm: Extended Mixed Models Using Latent Classes and Latent Processes. R package version: 1.8.1, https://cran.r-project.org/package=lcmm.

See also

Examples

data(latrendData)
method <- lcMethodLcmmGMM(fixed = Y ~ Time,
   mixture = ~ Time, random = ~ 1,
   id = "Id", time = "Time", , nClusters = 3)
gmm <- latrend(method, data = latrendData)

#> Be patient, hlme is running ... 
#> The program took 0.08 seconds 
#> Be patient, hlme is running ... 
#> The program took 0.66 seconds 

summary(gmm)

#> Longitudinal cluster model using growth mixture model
#> lcMethodLcmmGMM as "growth mixture model"
#>  nClusters:      3
#>  init:           "lme.random"
#>  id:             "Id"
#>  time:           "Time"
#>  classmb:        ~1
#>  random:         ~1
#>  mixture:        ~Time
#>  fixed:          Y ~ Time
#>  posfix:         NULL
#>  na.action:      1
#>  maxiter:        500
#>  convG:          1e-04
#>  convL:          1e-04
#>  convB:          1e-04
#>  cor:            NULL
#>  nwg:            FALSE
#>  idiag:          FALSE
#> 
#> Cluster sizes (K=3):
#>          A          B          C 
#>   90 (45%) 78 (39.1%) 32 (15.9%) 
#> 
#> Number of obs: 2000, strata (Id): 200
#> 
#> Scaled residuals:
#>     Min.  1st Qu.   Median     Mean  3rd Qu.     Max. 
#> -3.44863 -0.61421  0.06616  0.00000  0.67860  3.26730 
#> 

method <- lcMethodLcmmGMM(fixed = Y ~ Time,
    mixture = ~ Time, random = ~ Time,
    id = "Id", time = "Time", nClusters = 3)