Specify a GLMM iwht a normal mixture in the random effects

Specify a GLMM iwht a normal mixture in the random effects

lcMethodMixAK_GLMM(
  fixed,
  random,
  time = getOption("latrend.time"),
  id = getOption("latrend.id"),
  nClusters = 2,
  ...
)

Arguments

fixed

A formula specifying the fixed effects of the model, including the response. Creates the y and x arguments for the call to mixAK::GLMM_MCMC.

random

A formula specifying the random effects of the model, including the random intercept. Creates the z and random.intercept arguments for the call to mixAK::GLMM_MCMC.

time

The name of the time variable.

id

The name of the trajectory identifier variable. This is used to generate the id vector argument for the call to mixAK::GLMM_MCMC.

nClusters

The number of clusters.

...

Arguments passed to mixAK::GLMM_MCMC. The following external arguments are ignored: y, x, z, random.intercept, silent.

Note

This method currently does not appear to work under R 4.2 due to an error triggered by the mixAK package during fitting.

References

komarek2009newlatrend

See also

Examples

data(latrendData)
# this example only runs when the mixAK package is installed
try({
 method <- lcMethodMixAK_GLMM(fixed = Y ~ 1, random = ~ Time,
  id = "Id", time = "Time", nClusters = 3)
 model <- latrend(method, latrendData)
 summary(model)
})
#> Longitudinal cluster model using generalized linear mixed model with normal random effects mixture
#> lcMethodMixAK_GLMM specifying "generalized linear mixed model with normal random effects mixture"
#>  time:           "Time"
#>  id:             "Id"
#>  nClusters:      3
#>  dist:           "gaussian"
#>  nMCMC:          10, 10, 1, 10
#>  tuneMCMC:       list(alpha = 1, b = 1)
#>  store:          FALSE
#>  PED:            TRUE
#>  keep.chains:    TRUE
#>  dens.zero:      1e-300
#>  parallel:       FALSE
#>  fixed:          Y ~ 1
#>  random:         ~Time
#> 
#> Cluster sizes (K=3):
#>           A           B           C 
#>    25 (18%)  54 (42.8%) 121 (39.2%) 
#> 
#> Number of obs: 2000, strata (Id): 200
#> 
#> Scaled residuals:
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#> -3.7195 -0.6955 -0.1383  0.0000  0.5541  3.4917 
#>