Extract the log-likelihood of a lcModel

Extract the log-likelihood of a lcModel

# S3 method for lcModel
logLik(object, ...)

Arguments

object
The lcModel object.

...
Additional arguments.

Value

A numeric with the computed log-likelihood. If unavailable, NA is returned.

Details

The default implementation checks for the existence of the logLik() function for the internal model, and returns the output, if available.

See also

Examples

data(latrendData)
method <- lcMethodLcmmGBTM(fixed = Y ~ Time, mixture = ~ 1,
  id = "Id", time = "Time", nClusters = 3)
gbtm <- latrend(method, data = latrendData)

#> Be patient, hlme is running ... 
#> The program took 0.07 seconds 
#> Be patient, hlme is running ... 
#> The program took 0.32 seconds 

logLik(gbtm)

#> 'log Lik.' -882.5348 (df=7)