Posterior probability per fitted trajectory

Get the posterior probability matrix with element \((i,j)\) indicating the probability of trajectory \(i\) belonging to cluster \(j\).

# S4 method for lcModel
postprob(object, ...)

Arguments

object

The lcModel.

...

Additional arguments.

Value

A I-by-K matrix with I = nIds(object) and K = nClusters(object).

Details

This method should be extended by lcModel implementations. The default implementation returns uniform probabilities for all observations.

Implementation

Classes extending lcModel should override this method.

setMethod("postprob", "lcModelExt", function(object, ...) {
  # return trajectory-specific posterior probability matrix
})

Troubleshooting

If you are getting errors about undefined model signatures when calling postprob(model), check whether the postprob() function is still the one defined by the latrend package. It may have been overridden when attaching another package (e.g., lcmm). If you need to attach conflicting packages, load them first.

See also

Examples

data(latrendData)
method <- lcMethodLMKM(Y ~ Time, id = "Id", time = "Time")
model <- latrend(method, latrendData)

postprob(model)
#>        A B
#>   [1,] 1 0
#>   [2,] 1 0
#>   [3,] 1 0
#>   [4,] 1 0
#>   [5,] 1 0
#>   [6,] 1 0
#>   [7,] 1 0
#>   [8,] 1 0
#>   [9,] 1 0
#>  [10,] 1 0
#>  [11,] 1 0
#>  [12,] 1 0
#>  [13,] 1 0
#>  [14,] 1 0
#>  [15,] 1 0
#>  [16,] 1 0
#>  [17,] 1 0
#>  [18,] 1 0
#>  [19,] 1 0
#>  [20,] 1 0
#>  [21,] 1 0
#>  [22,] 1 0
#>  [23,] 1 0
#>  [24,] 1 0
#>  [25,] 1 0
#>  [26,] 1 0
#>  [27,] 1 0
#>  [28,] 1 0
#>  [29,] 1 0
#>  [30,] 1 0
#>  [31,] 1 0
#>  [32,] 1 0
#>  [33,] 1 0
#>  [34,] 1 0
#>  [35,] 1 0
#>  [36,] 1 0
#>  [37,] 1 0
#>  [38,] 1 0
#>  [39,] 1 0
#>  [40,] 1 0
#>  [41,] 1 0
#>  [42,] 1 0
#>  [43,] 1 0
#>  [44,] 1 0
#>  [45,] 1 0
#>  [46,] 1 0
#>  [47,] 1 0
#>  [48,] 1 0
#>  [49,] 1 0
#>  [50,] 1 0
#>  [51,] 1 0
#>  [52,] 1 0
#>  [53,] 1 0
#>  [54,] 1 0
#>  [55,] 1 0
#>  [56,] 1 0
#>  [57,] 1 0
#>  [58,] 1 0
#>  [59,] 1 0
#>  [60,] 1 0
#>  [61,] 1 0
#>  [62,] 1 0
#>  [63,] 1 0
#>  [64,] 1 0
#>  [65,] 1 0
#>  [66,] 1 0
#>  [67,] 1 0
#>  [68,] 1 0
#>  [69,] 1 0
#>  [70,] 1 0
#>  [71,] 1 0
#>  [72,] 1 0
#>  [73,] 1 0
#>  [74,] 1 0
#>  [75,] 1 0
#>  [76,] 1 0
#>  [77,] 1 0
#>  [78,] 1 0
#>  [79,] 1 0
#>  [80,] 1 0
#>  [81,] 0 1
#>  [82,] 0 1
#>  [83,] 0 1
#>  [84,] 0 1
#>  [85,] 0 1
#>  [86,] 0 1
#>  [87,] 0 1
#>  [88,] 0 1
#>  [89,] 0 1
#>  [90,] 0 1
#>  [91,] 0 1
#>  [92,] 0 1
#>  [93,] 0 1
#>  [94,] 0 1
#>  [95,] 0 1
#>  [96,] 0 1
#>  [97,] 0 1
#>  [98,] 0 1
#>  [99,] 0 1
#> [100,] 0 1
#> [101,] 0 1
#> [102,] 0 1
#> [103,] 0 1
#> [104,] 0 1
#> [105,] 0 1
#> [106,] 0 1
#> [107,] 0 1
#> [108,] 0 1
#> [109,] 0 1
#> [110,] 0 1
#> [111,] 0 1
#> [112,] 0 1
#> [113,] 0 1
#> [114,] 0 1
#> [115,] 0 1
#> [116,] 0 1
#> [117,] 0 1
#> [118,] 0 1
#> [119,] 0 1
#> [120,] 0 1
#> [121,] 0 1
#> [122,] 0 1
#> [123,] 0 1
#> [124,] 0 1
#> [125,] 0 1
#> [126,] 0 1
#> [127,] 0 1
#> [128,] 0 1
#> [129,] 0 1
#> [130,] 0 1
#> [131,] 0 1
#> [132,] 0 1
#> [133,] 0 1
#> [134,] 0 1
#> [135,] 0 1
#> [136,] 0 1
#> [137,] 0 1
#> [138,] 0 1
#> [139,] 0 1
#> [140,] 0 1
#> [141,] 0 1
#> [142,] 0 1
#> [143,] 0 1
#> [144,] 0 1
#> [145,] 0 1
#> [146,] 0 1
#> [147,] 0 1
#> [148,] 0 1
#> [149,] 0 1
#> [150,] 0 1
#> [151,] 0 1
#> [152,] 0 1
#> [153,] 0 1
#> [154,] 0 1
#> [155,] 0 1
#> [156,] 0 1
#> [157,] 0 1
#> [158,] 0 1
#> [159,] 0 1
#> [160,] 0 1
#> [161,] 0 1
#> [162,] 0 1
#> [163,] 0 1
#> [164,] 0 1
#> [165,] 0 1
#> [166,] 0 1
#> [167,] 0 1
#> [168,] 0 1
#> [169,] 0 1
#> [170,] 0 1
#> [171,] 0 1
#> [172,] 0 1
#> [173,] 0 1
#> [174,] 0 1
#> [175,] 0 1
#> [176,] 0 1
#> [177,] 0 1
#> [178,] 0 1
#> [179,] 0 1
#> [180,] 0 1
#> [181,] 0 1
#> [182,] 0 1
#> [183,] 0 1
#> [184,] 0 1
#> [185,] 0 1
#> [186,] 0 1
#> [187,] 0 1
#> [188,] 0 1
#> [189,] 0 1
#> [190,] 0 1
#> [191,] 0 1
#> [192,] 0 1
#> [193,] 0 1
#> [194,] 0 1
#> [195,] 0 1
#> [196,] 0 1
#> [197,] 0 1
#> [198,] 0 1
#> [199,] 0 1
#> [200,] 0 1

if (rlang::is_installed("lcmm")) {
  gmmMethod = lcMethodLcmmGMM(
    fixed = Y ~ Time,
    mixture = ~ Time,
    id = "Id",
    time = "Time"
  )
  gmmModel <- latrend(gmmMethod, data = latrendData)
  postprob(gmmModel)
}
#> ------------------ iteration  0 ------------------
#> Function value  -2211.674 
#> Convergence criteria: parameters stability= 1.0001 
#>                     : function stability= 1.0001 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef   SE.coef    Var.coef
#> parameter1 -0.03295723  3.671255    13.47811
#> parameter2  0.00000000  4.663262    21.74601
#> parameter3  1.00000000 11.298389  -127.65358
#> parameter4  1.00000000 31.748083 -1007.94077
#> 
#> ------------------ iteration  1 ------------------
#> Function value  -2211.661 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 0.01365446 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef   SE.coef    Var.coef
#> parameter1 -0.03295781  3.672703    13.48875
#> parameter2  0.00000137  5.416630    29.33988
#> parameter3  0.99999000 11.298556  -127.65737
#> parameter4  0.99999246 31.748115 -1007.94279
#> 
#> ------------------ iteration  2 ------------------
#> Function value  -2211.643 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 0.01733755 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef   SE.coef    Var.coef
#> parameter1 -0.03295808  3.671208    13.47777
#> parameter2  0.00000233  4.946477    24.46763
#> parameter3  0.99998055 11.298637  -127.65919
#> parameter4  0.99998246 31.748227 -1007.94995
#> 
#> ------------------ iteration  3 ------------------
#> Function value  -1439.862 
#> Convergence criteria: parameters stability= 0.280072 
#>                     : function stability= 771.7816 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef   SE.coef   Var.coef
#> parameter1 -0.04031975  4.568936   20.87517
#> parameter2  0.04576697  6.864072   47.11549
#> parameter3  0.64986537 15.463936 -239.13331
#> parameter4  0.60584696 26.426104  698.33896
#> 
#> ------------------ iteration  4 ------------------
#> Function value  -1439.841 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 0.02074656 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef   SE.coef   Var.coef
#> parameter1 -0.04032027  4.569053   20.87625
#> parameter2  0.04576783  6.864238   47.11776
#> parameter3  0.64985797 15.464033 -239.13633
#> parameter4  0.60583696 26.429943  698.54191
#> 
#> ------------------ iteration  5 ------------------
#> Function value  -1259.308 
#> Convergence criteria: parameters stability= 0.1423089 
#>                     : function stability= 180.5326 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef    SE.coef     Var.coef
#> parameter1 -0.05020431   5.429509     29.47957
#> parameter2  0.06607079  10.206393    104.17045
#> parameter3  0.47182147  16.189455   -262.09845
#> parameter4  0.27402077 328.006561 107588.30435
#> 
#> ------------------ iteration  6 ------------------
#> Function value  -1182.724 
#> Convergence criteria: parameters stability= 1.186942 
#>                     : function stability= 76.58411 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef    SE.coef    Var.coef
#> parameter1 -0.17261108   4.785824    22.90411
#> parameter2  0.09024304   9.785319    95.75248
#> parameter3 -0.61021076  17.228002  -296.80404
#> parameter4  0.29811228 269.571111 72668.58395
#> 
#> ------------------ iteration  7 ------------------
#> Function value  -1182.721 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 0.00300208 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef    SE.coef    Var.coef
#> parameter1 -0.17261028   4.785870    22.90455
#> parameter2  0.09024312   9.785316    95.75240
#> parameter3 -0.61020076  17.228015  -296.80448
#> parameter4  0.29811233 269.571036 72668.54354
#> 
#> ------------------ iteration  8 ------------------
#> Function value  -1182.718 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 0.00380941 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef    SE.coef    Var.coef
#> parameter1 -0.17260961   4.785916    22.90499
#> parameter2  0.09024344   9.785311    95.75231
#> parameter3 -0.61019076  17.228004  -296.80414
#> parameter4  0.29811260 269.570458 72668.23197
#> 
#> ------------------ iteration  9 ------------------
#> Function value  -1070.899 
#> Convergence criteria: parameters stability= 0.05788247 
#>                     : function stability= 111.819 
#>                     : relative distance to maximum(RDM)= 735.0503 
#>                  coef    SE.coef   Var.coef
#> parameter1 -0.1578944 0.03390969 0.00114987
#> parameter2  0.1032584 0.01088495 0.00011848
#> parameter3 -0.3706614 0.15798998 0.02496083
#> parameter4  0.3091681 0.00406641 0.00001654
#> 
#> ------------------ iteration  10 ------------------
#> Function value  -1070.895 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 0.00369269 
#>                     : relative distance to maximum(RDM)= 734.1069 
#>                  coef    SE.coef   Var.coef
#> parameter1 -0.1578937 0.03390330 0.00114943
#> parameter2  0.1032586 0.01088495 0.00011848
#> parameter3 -0.3706514 0.15788498 0.02492767
#> parameter4  0.3091682 0.00406640 0.00001654
#> 
#> ------------------ iteration  11 ------------------
#> Function value  -1070.891 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 0.00379712 
#>                     : relative distance to maximum(RDM)= 732.7804 
#>                  coef    SE.coef   Var.coef
#> parameter1 -0.1578932 0.03389448 0.00114884
#> parameter2  0.1032587 0.01088496 0.00011848
#> parameter3 -0.3706414 0.15773736 0.02488108
#> parameter4  0.3091683 0.00406640 0.00001654
#> 
#> ------------------ iteration  12 ------------------
#> Function value  -1044.445 
#> Convergence criteria: parameters stability= 0.5764923 
#>                     : function stability= 26.44575 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                  coef    SE.coef    Var.coef
#> parameter1 -0.1270200   5.938905    35.27059
#> parameter2  0.1190798   9.395830    88.28163
#> parameter3  0.3877042   6.559355   -43.02514
#> parameter4  0.3233399 222.402347 49462.80385
#> 
#> ------------------ iteration  13 ------------------
#> Function value  -1044.442 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 0.00348247 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                  coef    SE.coef    Var.coef
#> parameter1 -0.1270206   5.938986    35.27155
#> parameter2  0.1190798   9.395828    88.28158
#> parameter3  0.3876942   6.555280   -42.97169
#> parameter4  0.3233399 222.402234 49462.75351
#> 
#> ------------------ iteration  14 ------------------
#> Function value  -1044.438 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 0.00354728 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                  coef    SE.coef    Var.coef
#> parameter1 -0.1270210   5.939060    35.27243
#> parameter2  0.1190798   9.395827    88.28156
#> parameter3  0.3876842   6.551120   -42.91718
#> parameter4  0.3233400 222.402055 49462.67395
#> 
#> ------------------ iteration  15 ------------------
#> Function value  -1044.435 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 0.00359666 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                  coef    SE.coef    Var.coef
#> parameter1 -0.1270214   5.939133    35.27331
#> parameter2  0.1190799   9.395822    88.28146
#> parameter3  0.3876742   6.547073   -42.86417
#> parameter4  0.3233402 222.401858 49462.58657
#> 
#> ------------------ iteration  16 ------------------
#> Function value  -990.0216 
#> Convergence criteria: parameters stability= 0.03537843 
#>                     : function stability= 54.41312 
#>                     : relative distance to maximum(RDM)= 12.63568 
#>                  coef    SE.coef   Var.coef
#> parameter1 -0.1312689 0.01972446 0.00038905
#> parameter2  0.1205435 0.01145365 0.00013119
#> parameter3  0.1996608 0.01309509 0.00017148
#> parameter4  0.3263760 0.00460804 0.00002123
#> 
#> ------------------ iteration  17 ------------------
#> Function value  -973.447 
#> Convergence criteria: parameters stability= 0.00032527 
#>                     : function stability= 16.5746 
#>                     : relative distance to maximum(RDM)= 5.923468 
#>                  coef    SE.coef   Var.coef
#> parameter1 -0.1424467 0.01988305 0.00039534
#> parameter2  0.1250521 0.01190601 0.00014175
#> parameter3  0.1966762 0.01313650 0.00017257
#> parameter4  0.3394560 0.00505861 0.00002559
#> 
#> ------------------ iteration  18 ------------------
#> Function value  -961.3927 
#> Convergence criteria: parameters stability= 0.00105038 
#>                     : function stability= 12.05426 
#>                     : relative distance to maximum(RDM)= 0.593486 
#>                  coef    SE.coef   Var.coef
#> parameter1 -0.1651658 0.02005362 0.00040215
#> parameter2  0.1362990 0.01255911 0.00015773
#> parameter3  0.1897438 0.01300009 0.00016900
#> parameter4  0.3584212 0.00577301 0.00003333
#> 
#> ------------------ iteration  19 ------------------
#> Function value  -960.1702 
#> Convergence criteria: parameters stability= 0.00024153 
#>                     : function stability= 1.222577 
#>                     : relative distance to maximum(RDM)= 0.00468292 
#>                  coef    SE.coef   Var.coef
#> parameter1 -0.1762450 0.02016492 0.00040662
#> parameter2  0.1435649 0.01282592 0.00016450
#> parameter3  0.1871356 0.01299865 0.00016896
#> parameter4  0.3661144 0.00608560 0.00003703
#> 
#> ------------------ iteration  20 ------------------
#> Function value  -960.1608 
#> Convergence criteria: parameters stability= 2.61e-06 
#>                     : function stability= 0.00939256 
#>                     : relative distance to maximum(RDM)= 5.3e-07 
#>                  coef    SE.coef   Var.coef
#> parameter1 -0.1773901 0.02017775 0.00040714
#> parameter2  0.1444374 0.01285068 0.00016514
#> parameter3  0.1869344 0.01300311 0.00016908
#> parameter4  0.3668221 0.00611506 0.00003739
#> 
#> ------------------ iteration  21 ------------------
#> Function value  -960.1608 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 1.06e-06 
#>                     : relative distance to maximum(RDM)= 0 
#>                  coef    SE.coef   Var.coef
#> parameter1 -0.1774069 0.02017785 0.00040715
#> parameter2  0.1444497 0.01285090 0.00016515
#> parameter3  0.1869323 0.01300313 0.00016908
#> parameter4  0.3668285 0.00611533 0.00003740
#> 
#> The program took 0.46 seconds 
#> ------------------ iteration  0 ------------------
#> Function value  -959.2723 
#> Convergence criteria: parameters stability= 1.0001 
#>                     : function stability= 1.0001 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                  coef   SE.coef  Var.coef
#> parameter1  0.0000000   0.00000     0.000
#> parameter2 -0.1874958  32.56135  1060.242
#> parameter3 -0.1673180  32.52617  1057.952
#> parameter4  0.1380242  86.33462 -7453.666
#> parameter5  0.1508751  86.34873 -7456.103
#> parameter6  0.1869323  76.50034  5852.302
#> parameter7  0.3668285 163.79515 26828.850
#> 
#> ------------------ iteration  1 ------------------
#> Function value  -957.172 
#> Convergence criteria: parameters stability= 6.145e-05 
#>                     : function stability= 2.100332 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                  coef    SE.coef     Var.coef
#> parameter1  0.0000005   2.873858     8.259062
#> parameter2 -0.1875478   5.743915    32.992561
#> parameter3 -0.1672656   3.956203    15.651543
#> parameter4  0.1324900  82.608245 -6824.122064
#> parameter5  0.1564099  82.612867 -6824.885748
#> parameter6  0.1867822  76.106068  5792.133534
#> parameter7  0.3664314 163.242233 26648.026657
#> 
#> ------------------ iteration  2 ------------------
#> Function value  -939.807 
#> Convergence criteria: parameters stability= 0.00087023 
#>                     : function stability= 17.36502 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef    SE.coef     Var.coef
#> parameter1  0.00005488   2.596824     6.743496
#> parameter2 -0.18764726   5.950719    35.411057
#> parameter3 -0.16715930   5.249387    27.556063
#> parameter4  0.11164707  54.870240 -3010.743218
#> parameter5  0.17725207  54.475787 -2967.611335
#> parameter6  0.18612558  74.101022  5490.961449
#> parameter7  0.36545858 157.038030 24660.942894
#> 
#> ------------------ iteration  3 ------------------
#> Function value  -920.3918 
#> Convergence criteria: parameters stability= 0.00046485 
#>                     : function stability= 19.41517 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef    SE.coef    Var.coef
#> parameter1 -0.00008149   3.371748   -11.36868
#> parameter2 -0.18638160  37.133125  1378.86898
#> parameter3 -0.16848824  36.505867  1332.67832
#> parameter4  0.09655969  20.983783  -440.31915
#> parameter5  0.19251663  20.000820  -400.03278
#> parameter6  0.18589413  73.030505  5333.45463
#> parameter7  0.36457657 151.289634 22888.55342
#> 
#> ------------------ iteration  4 ------------------
#> Function value  -870.4228 
#> Convergence criteria: parameters stability= 0.00225473 
#>                     : function stability= 49.96899 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef    SE.coef    Var.coef
#> parameter1 -0.00013076   6.743496   -45.47474
#> parameter2 -0.18534977  39.147602  1532.53477
#> parameter3 -0.16962638  37.912634  1437.36785
#> parameter4  0.06420283  53.303305  2841.24235
#> parameter5  0.22723004  50.873768  2588.14026
#> parameter6  0.18585681  71.701008  5141.03451
#> parameter7  0.36396508 138.031449 19052.68093
#> 
#> ------------------ iteration  5 ------------------
#> Function value  -807.3687 
#> Convergence criteria: parameters stability= 0.00288987 
#>                     : function stability= 63.05413 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef    SE.coef     Var.coef
#> parameter1  0.00154924   2.308902     5.331028
#> parameter2 -0.17307886   6.325157    40.007609
#> parameter3 -0.18296970   6.160603    37.953035
#> parameter4  0.03017905   6.362331    40.479255
#> parameter5  0.26391804   4.349280    18.916235
#> parameter6  0.18744110   7.997235    63.955775
#> parameter7  0.35673374 131.790811 17368.817851
#> 
#> ------------------ iteration  6 ------------------
#> Function value  -234.0025 
#> Convergence criteria: parameters stability= 0.2354196 
#>                     : function stability= 573.3662 
#>                     : relative distance to maximum(RDM)= 88.25512 
#>                   coef    SE.coef   Var.coef
#> parameter1  0.15233864 0.14201172 0.02016733
#> parameter2  0.03617987 0.02992324 0.00089540
#> parameter3 -0.40084166 0.03903369 0.00152363
#> parameter4 -0.17597225 0.01136866 0.00012925
#> parameter5  0.51667596 0.01643502 0.00027011
#> parameter6  0.24031620 0.01862848 0.00034702
#> parameter7  0.24608093 0.00589851 0.00003479
#> 
#> ------------------ iteration  7 ------------------
#> Function value  -116.8424 
#> Convergence criteria: parameters stability= 0.2750103 
#>                     : function stability= 117.1601 
#>                     : relative distance to maximum(RDM)= 14.02445 
#>                  coef    SE.coef   Var.coef
#> parameter1  0.3350642 0.14341104 0.02056673
#> parameter2  0.2213973 0.01957776 0.00038329
#> parameter3 -0.7824800 0.02390620 0.00057151
#> parameter4 -0.2305114 0.00800543 0.00006409
#> parameter5  0.7425480 0.00987910 0.00009760
#> parameter6  0.1864792 0.00860017 0.00007396
#> parameter7  0.1769633 0.00238969 0.00000571
#> 
#> ------------------ iteration  8 ------------------
#> Function value  -60.66954 
#> Convergence criteria: parameters stability= 0.00715973 
#>                     : function stability= 56.17281 
#>                     : relative distance to maximum(RDM)= 1.323995 
#>                  coef    SE.coef   Var.coef
#> parameter1  0.3942235 0.14417829 0.02078738
#> parameter2  0.1995742 0.02137132 0.00045673
#> parameter3 -0.7462780 0.02617159 0.00068495
#> parameter4 -0.2256329 0.00888236 0.00007890
#> parameter5  0.7082258 0.01088404 0.00011846
#> parameter6  0.2036181 0.01062426 0.00011287
#> parameter7  0.1963930 0.00305454 0.00000933
#> 
#> ------------------ iteration  9 ------------------
#> Function value  -55.73164 
#> Convergence criteria: parameters stability= 0.00050486 
#>                     : function stability= 4.937899 
#>                     : relative distance to maximum(RDM)= 0.01668516 
#>                  coef    SE.coef   Var.coef
#> parameter1  0.4052535 0.14433504 0.02083260
#> parameter2  0.1937467 0.02206570 0.00048689
#> parameter3 -0.7346090 0.02702485 0.00073034
#> parameter4 -0.2240926 0.00926405 0.00008582
#> parameter5  0.6980466 0.01134615 0.00012874
#> parameter6  0.2096016 0.01144289 0.00013094
#> parameter7  0.2048358 0.00338725 0.00001147
#> 
#> ------------------ iteration  10 ------------------
#> Function value  -55.67275 
#> Convergence criteria: parameters stability= 4.07e-06 
#>                     : function stability= 0.05888992 
#>                     : relative distance to maximum(RDM)= 3.06e-06 
#>                  coef    SE.coef   Var.coef
#> parameter1  0.4054646 0.14433800 0.02083346
#> parameter2  0.1932927 0.02213587 0.00049000
#> parameter3 -0.7334642 0.02711079 0.00073499
#> parameter4 -0.2239387 0.00931372 0.00008675
#> parameter5  0.6970418 0.01140694 0.00013012
#> parameter6  0.2101172 0.01152210 0.00013276
#> parameter7  0.2059342 0.00343271 0.00001178
#> 
#> ------------------ iteration  11 ------------------
#> Function value  -55.67274 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 1.071e-05 
#>                     : relative distance to maximum(RDM)= 0 
#>                  coef    SE.coef   Var.coef
#> parameter1  0.4054651 0.14433818 0.02083351
#> parameter2  0.1932891 0.02213651 0.00049003
#> parameter3 -0.7334510 0.02711158 0.00073504
#> parameter4 -0.2239371 0.00931442 0.00008676
#> parameter5  0.6970299 0.01140779 0.00013014
#> parameter6  0.2101202 0.01152271 0.00013277
#> parameter7  0.2059496 0.00343335 0.00001179
#> 
#> The program took 0.42 seconds 
#>                A            B
#> 1   2.547309e-17 1.000000e+00
#> 2   5.411522e-22 1.000000e+00
#> 3   3.041642e-19 1.000000e+00
#> 4   2.235736e-20 1.000000e+00
#> 5   8.499348e-10 1.000000e+00
#> 6   1.428899e-22 1.000000e+00
#> 7   7.217099e-20 1.000000e+00
#> 8   4.601011e-22 1.000000e+00
#> 9   2.465789e-18 1.000000e+00
#> 10  1.487810e-17 1.000000e+00
#> 11  4.431033e-18 1.000000e+00
#> 12  1.048311e-23 1.000000e+00
#> 13  1.624114e-19 1.000000e+00
#> 14  8.535574e-17 1.000000e+00
#> 15  9.154080e-19 1.000000e+00
#> 16  2.775977e-20 1.000000e+00
#> 17  3.751780e-19 1.000000e+00
#> 18  1.074618e-14 1.000000e+00
#> 19  3.429499e-19 1.000000e+00
#> 20  5.383558e-18 1.000000e+00
#> 21  8.954243e-15 1.000000e+00
#> 22  8.071920e-17 1.000000e+00
#> 23  3.636239e-18 1.000000e+00
#> 24  5.106287e-20 1.000000e+00
#> 25  3.644339e-17 1.000000e+00
#> 26  1.171056e-21 1.000000e+00
#> 27  1.186961e-19 1.000000e+00
#> 28  1.506209e-14 1.000000e+00
#> 29  5.704580e-19 1.000000e+00
#> 30  5.336798e-19 1.000000e+00
#> 31  1.202215e-17 1.000000e+00
#> 32  1.410158e-12 1.000000e+00
#> 33  1.841029e-16 1.000000e+00
#> 34  1.571870e-21 1.000000e+00
#> 35  2.531197e-23 1.000000e+00
#> 36  1.134933e-21 1.000000e+00
#> 37  8.624020e-20 1.000000e+00
#> 38  8.734216e-18 1.000000e+00
#> 39  3.199024e-20 1.000000e+00
#> 40  1.589164e-15 1.000000e+00
#> 41  6.669018e-21 1.000000e+00
#> 42  1.013342e-18 1.000000e+00
#> 43  7.838956e-18 1.000000e+00
#> 44  9.027838e-15 1.000000e+00
#> 45  4.032962e-18 1.000000e+00
#> 46  1.606951e-13 1.000000e+00
#> 47  1.326813e-15 1.000000e+00
#> 48  6.837974e-17 1.000000e+00
#> 49  1.112480e-18 1.000000e+00
#> 50  5.127967e-18 1.000000e+00
#> 51  6.227700e-20 1.000000e+00
#> 52  5.071706e-22 1.000000e+00
#> 53  1.910355e-16 1.000000e+00
#> 54  1.315540e-15 1.000000e+00
#> 55  1.489717e-20 1.000000e+00
#> 56  3.086665e-17 1.000000e+00
#> 57  1.051878e-14 1.000000e+00
#> 58  2.876305e-21 1.000000e+00
#> 59  1.715890e-19 1.000000e+00
#> 60  1.321604e-15 1.000000e+00
#> 61  9.535077e-18 1.000000e+00
#> 62  9.190271e-20 1.000000e+00
#> 63  2.424449e-16 1.000000e+00
#> 64  7.904951e-23 1.000000e+00
#> 65  1.599877e-22 1.000000e+00
#> 66  2.884337e-19 1.000000e+00
#> 67  6.949903e-17 1.000000e+00
#> 68  1.967578e-14 1.000000e+00
#> 69  1.408407e-17 1.000000e+00
#> 70  3.518169e-15 1.000000e+00
#> 71  2.106024e-17 1.000000e+00
#> 72  2.756480e-16 1.000000e+00
#> 73  4.318563e-15 1.000000e+00
#> 74  1.844605e-23 1.000000e+00
#> 75  1.737188e-17 1.000000e+00
#> 76  1.314989e-21 1.000000e+00
#> 77  6.142480e-13 1.000000e+00
#> 78  2.824823e-19 1.000000e+00
#> 79  4.868013e-15 1.000000e+00
#> 80  1.192237e-17 1.000000e+00
#> 81  9.999999e-01 5.541483e-08
#> 82  1.000000e+00 1.687982e-09
#> 83  1.000000e+00 1.237469e-08
#> 84  1.000000e+00 2.646910e-11
#> 85  1.000000e+00 2.743526e-11
#> 86  1.000000e+00 2.225112e-12
#> 87  1.000000e+00 3.498192e-09
#> 88  1.000000e+00 3.840670e-10
#> 89  1.000000e+00 4.912743e-09
#> 90  1.000000e+00 4.347079e-10
#> 91  1.000000e+00 6.570054e-11
#> 92  1.000000e+00 9.025892e-10
#> 93  1.000000e+00 7.647599e-11
#> 94  1.000000e+00 2.499976e-10
#> 95  1.000000e+00 1.745679e-08
#> 96  1.000000e+00 3.376375e-10
#> 97  1.000000e+00 7.724647e-09
#> 98  1.000000e+00 3.138129e-09
#> 99  1.000000e+00 7.196848e-09
#> 100 1.000000e+00 3.432416e-09
#> 101 1.000000e+00 3.378190e-11
#> 102 1.000000e+00 6.094949e-10
#> 103 1.000000e+00 3.903088e-10
#> 104 1.000000e+00 1.463033e-09
#> 105 1.000000e+00 3.266176e-08
#> 106 9.999997e-01 2.691745e-07
#> 107 1.000000e+00 2.165185e-10
#> 108 1.000000e+00 4.417720e-08
#> 109 1.000000e+00 2.613364e-10
#> 110 1.000000e+00 1.855565e-10
#> 111 1.000000e+00 1.247697e-08
#> 112 9.999999e-01 1.177873e-07
#> 113 1.000000e+00 2.975804e-09
#> 114 1.000000e+00 1.558266e-09
#> 115 1.000000e+00 1.849239e-09
#> 116 1.000000e+00 1.031145e-09
#> 117 1.000000e+00 7.985078e-10
#> 118 1.000000e+00 4.550104e-09
#> 119 1.000000e+00 4.156114e-10
#> 120 1.000000e+00 3.206136e-10
#> 121 1.000000e+00 2.233346e-08
#> 122 9.999999e-01 8.120601e-08
#> 123 1.000000e+00 5.954044e-10
#> 124 1.000000e+00 3.026458e-11
#> 125 1.000000e+00 1.424766e-09
#> 126 1.000000e+00 8.130364e-10
#> 127 1.000000e+00 6.150024e-10
#> 128 1.000000e+00 9.444138e-09
#> 129 1.000000e+00 2.328948e-11
#> 130 1.000000e+00 1.563133e-09
#> 131 1.000000e+00 7.070482e-12
#> 132 1.000000e+00 2.430530e-12
#> 133 1.000000e+00 1.825761e-09
#> 134 1.000000e+00 6.650521e-09
#> 135 1.000000e+00 1.370103e-11
#> 136 1.000000e+00 1.025182e-09
#> 137 1.000000e+00 1.227042e-09
#> 138 1.000000e+00 2.044575e-09
#> 139 1.000000e+00 5.253970e-09
#> 140 1.000000e+00 3.989508e-08
#> 141 1.000000e+00 6.667898e-10
#> 142 1.000000e+00 6.031687e-11
#> 143 1.000000e+00 9.281413e-11
#> 144 1.000000e+00 3.580572e-09
#> 145 1.000000e+00 5.328300e-13
#> 146 1.000000e+00 2.029222e-12
#> 147 1.000000e+00 3.769147e-09
#> 148 1.000000e+00 2.288466e-09
#> 149 1.000000e+00 1.052566e-08
#> 150 1.000000e+00 3.274686e-08
#> 151 1.000000e+00 2.471732e-26
#> 152 1.000000e+00 8.469882e-29
#> 153 1.000000e+00 1.695607e-31
#> 154 1.000000e+00 1.193504e-30
#> 155 1.000000e+00 8.066766e-25
#> 156 1.000000e+00 3.073320e-26
#> 157 1.000000e+00 1.101321e-34
#> 158 1.000000e+00 2.275016e-31
#> 159 1.000000e+00 2.911821e-35
#> 160 1.000000e+00 2.330361e-32
#> 161 1.000000e+00 1.275205e-41
#> 162 1.000000e+00 3.527564e-25
#> 163 1.000000e+00 5.453888e-33
#> 164 1.000000e+00 5.219841e-32
#> 165 1.000000e+00 4.904099e-30
#> 166 1.000000e+00 4.439443e-29
#> 167 1.000000e+00 1.565171e-28
#> 168 1.000000e+00 4.165055e-28
#> 169 1.000000e+00 7.260726e-33
#> 170 1.000000e+00 2.191088e-36
#> 171 1.000000e+00 1.604883e-33
#> 172 1.000000e+00 7.948705e-31
#> 173 1.000000e+00 2.416388e-34
#> 174 1.000000e+00 2.302697e-31
#> 175 1.000000e+00 1.701660e-35
#> 176 1.000000e+00 1.088227e-32
#> 177 1.000000e+00 4.202725e-30
#> 178 1.000000e+00 2.132957e-32
#> 179 1.000000e+00 1.985949e-32
#> 180 1.000000e+00 2.624164e-24
#> 181 1.000000e+00 1.952052e-29
#> 182 1.000000e+00 6.747480e-26
#> 183 1.000000e+00 7.691045e-35
#> 184 1.000000e+00 2.337360e-30
#> 185 1.000000e+00 3.671426e-27
#> 186 1.000000e+00 1.975783e-30
#> 187 1.000000e+00 2.035570e-39
#> 188 1.000000e+00 4.969321e-28
#> 189 1.000000e+00 4.456556e-38
#> 190 1.000000e+00 1.971674e-33
#> 191 1.000000e+00 5.329904e-25
#> 192 1.000000e+00 1.312493e-29
#> 193 1.000000e+00 1.176425e-25
#> 194 1.000000e+00 3.285059e-31
#> 195 1.000000e+00 5.007539e-31
#> 196 1.000000e+00 7.281392e-36
#> 197 1.000000e+00 5.466925e-24
#> 198 1.000000e+00 4.488509e-31
#> 199 1.000000e+00 4.130772e-33
#> 200 1.000000e+00 5.307268e-28