Posterior probability per fitted trajectory

Get the posterior probability matrix with element \((i,j)\) indicating the probability of trajectory \(i\) belonging to cluster \(j\).

# S4 method for lcModel
postprob(object, ...)

Arguments

object

The lcModel.

...

Additional arguments.

Value

A I-by-K matrix with I = nIds(object) and K = nClusters(object).

Details

This method should be extended by lcModel implementations. The default implementation returns uniform probabilities for all observations.

Implementation

Classes extending lcModel should override this method.

setMethod("postprob", "lcModelExt", function(object, ...) {
  # return trajectory-specific posterior probability matrix
})

Troubleshooting

If you are getting errors about undefined model signatures when calling postprob(model), check whether the postprob() function is still the one defined by the latrend package. It may have been overridden when attaching another package (e.g., lcmm). If you need to attach conflicting packages, load them first.

See also

Examples

data(latrendData)
method <- lcMethodLMKM(Y ~ Time, id = "Id", time = "Time")
model <- latrend(method, latrendData)

postprob(model)
#>        A B
#>   [1,] 0 1
#>   [2,] 0 1
#>   [3,] 0 1
#>   [4,] 0 1
#>   [5,] 0 1
#>   [6,] 0 1
#>   [7,] 0 1
#>   [8,] 0 1
#>   [9,] 0 1
#>  [10,] 0 1
#>  [11,] 0 1
#>  [12,] 0 1
#>  [13,] 0 1
#>  [14,] 0 1
#>  [15,] 0 1
#>  [16,] 0 1
#>  [17,] 0 1
#>  [18,] 0 1
#>  [19,] 0 1
#>  [20,] 0 1
#>  [21,] 0 1
#>  [22,] 0 1
#>  [23,] 0 1
#>  [24,] 0 1
#>  [25,] 0 1
#>  [26,] 0 1
#>  [27,] 0 1
#>  [28,] 0 1
#>  [29,] 0 1
#>  [30,] 0 1
#>  [31,] 0 1
#>  [32,] 0 1
#>  [33,] 0 1
#>  [34,] 0 1
#>  [35,] 0 1
#>  [36,] 0 1
#>  [37,] 0 1
#>  [38,] 0 1
#>  [39,] 0 1
#>  [40,] 0 1
#>  [41,] 0 1
#>  [42,] 0 1
#>  [43,] 0 1
#>  [44,] 0 1
#>  [45,] 0 1
#>  [46,] 0 1
#>  [47,] 0 1
#>  [48,] 0 1
#>  [49,] 0 1
#>  [50,] 0 1
#>  [51,] 0 1
#>  [52,] 0 1
#>  [53,] 0 1
#>  [54,] 0 1
#>  [55,] 0 1
#>  [56,] 0 1
#>  [57,] 0 1
#>  [58,] 0 1
#>  [59,] 0 1
#>  [60,] 0 1
#>  [61,] 0 1
#>  [62,] 0 1
#>  [63,] 0 1
#>  [64,] 0 1
#>  [65,] 0 1
#>  [66,] 0 1
#>  [67,] 0 1
#>  [68,] 0 1
#>  [69,] 0 1
#>  [70,] 0 1
#>  [71,] 0 1
#>  [72,] 0 1
#>  [73,] 0 1
#>  [74,] 0 1
#>  [75,] 0 1
#>  [76,] 0 1
#>  [77,] 0 1
#>  [78,] 0 1
#>  [79,] 0 1
#>  [80,] 0 1
#>  [81,] 0 1
#>  [82,] 0 1
#>  [83,] 0 1
#>  [84,] 0 1
#>  [85,] 0 1
#>  [86,] 0 1
#>  [87,] 0 1
#>  [88,] 0 1
#>  [89,] 0 1
#>  [90,] 0 1
#>  [91,] 1 0
#>  [92,] 1 0
#>  [93,] 1 0
#>  [94,] 1 0
#>  [95,] 1 0
#>  [96,] 1 0
#>  [97,] 1 0
#>  [98,] 1 0
#>  [99,] 1 0
#> [100,] 1 0
#> [101,] 1 0
#> [102,] 1 0
#> [103,] 1 0
#> [104,] 1 0
#> [105,] 1 0
#> [106,] 1 0
#> [107,] 1 0
#> [108,] 1 0
#> [109,] 1 0
#> [110,] 1 0
#> [111,] 1 0
#> [112,] 1 0
#> [113,] 1 0
#> [114,] 1 0
#> [115,] 1 0
#> [116,] 1 0
#> [117,] 1 0
#> [118,] 1 0
#> [119,] 1 0
#> [120,] 1 0
#> [121,] 1 0
#> [122,] 1 0
#> [123,] 1 0
#> [124,] 1 0
#> [125,] 1 0
#> [126,] 1 0
#> [127,] 1 0
#> [128,] 1 0
#> [129,] 1 0
#> [130,] 1 0
#> [131,] 1 0
#> [132,] 1 0
#> [133,] 1 0
#> [134,] 1 0
#> [135,] 1 0
#> [136,] 1 0
#> [137,] 1 0
#> [138,] 1 0
#> [139,] 1 0
#> [140,] 1 0
#> [141,] 1 0
#> [142,] 1 0
#> [143,] 1 0
#> [144,] 1 0
#> [145,] 1 0
#> [146,] 1 0
#> [147,] 1 0
#> [148,] 1 0
#> [149,] 1 0
#> [150,] 1 0
#> [151,] 1 0
#> [152,] 1 0
#> [153,] 1 0
#> [154,] 1 0
#> [155,] 1 0
#> [156,] 1 0
#> [157,] 1 0
#> [158,] 1 0
#> [159,] 1 0
#> [160,] 1 0
#> [161,] 1 0
#> [162,] 1 0
#> [163,] 1 0
#> [164,] 1 0
#> [165,] 1 0
#> [166,] 1 0
#> [167,] 1 0
#> [168,] 1 0
#> [169,] 1 0
#> [170,] 1 0
#> [171,] 1 0
#> [172,] 1 0
#> [173,] 1 0
#> [174,] 1 0
#> [175,] 1 0
#> [176,] 1 0
#> [177,] 1 0
#> [178,] 1 0
#> [179,] 1 0
#> [180,] 1 0
#> [181,] 1 0
#> [182,] 1 0
#> [183,] 1 0
#> [184,] 1 0
#> [185,] 1 0
#> [186,] 1 0
#> [187,] 1 0
#> [188,] 1 0
#> [189,] 1 0
#> [190,] 1 0
#> [191,] 1 0
#> [192,] 1 0
#> [193,] 1 0
#> [194,] 1 0
#> [195,] 1 0
#> [196,] 1 0
#> [197,] 1 0
#> [198,] 1 0
#> [199,] 1 0
#> [200,] 1 0

if (rlang::is_installed("lcmm")) {
  gmmMethod = lcMethodLcmmGMM(
    fixed = Y ~ Time,
    mixture = ~ Time,
    id = "Id",
    time = "Time"
  )
  gmmModel <- latrend(gmmMethod, data = latrendData)
  postprob(gmmModel)
}
#> ------------------ iteration  0 ------------------
#> Function value  -2214.962 
#> Convergence criteria: parameters stability= 1.0001 
#>                     : function stability= 1.0001 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef   SE.coef   Var.coef
#> parameter1 -0.02626082  3.672290   13.48571
#> parameter2  0.00000000  5.009514   25.09523
#> parameter3  1.00000000 11.170348 -124.77667
#> parameter4  1.00000000 31.494353 -991.89424
#> 
#> ------------------ iteration  1 ------------------
#> Function value  -2214.946 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 0.01528693 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef   SE.coef   Var.coef
#> parameter1 -0.02626170  3.669229   13.46324
#> parameter2  0.00000196  4.698516   22.07605
#> parameter3  0.99999000 11.170594 -124.78217
#> parameter4  0.99999151 31.494423 -991.89866
#> 
#> ------------------ iteration  2 ------------------
#> Function value  -2214.929 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 0.01761045 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef   SE.coef   Var.coef
#> parameter1 -0.02626236  3.673180   13.49225
#> parameter2  0.00000453  4.258475   18.13461
#> parameter3  0.99998085 11.170692 -124.78436
#> parameter4  0.99998151 31.494499 -991.90344
#> 
#> ------------------ iteration  3 ------------------
#> Function value  -1384.589 
#> Convergence criteria: parameters stability= 0.3102247 
#>                     : function stability= 830.34 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef   SE.coef   Var.coef
#> parameter1 -0.04536491  4.601141   21.17050
#> parameter2  0.11953703  6.947529   48.26816
#> parameter3  0.64134243 14.223751 -202.31508
#> parameter4  0.59138599 22.660503  513.49841
#> 
#> ------------------ iteration  4 ------------------
#> Function value  -1111.675 
#> Convergence criteria: parameters stability= 1.057287 
#>                     : function stability= 272.9141 
#>                     : relative distance to maximum(RDM)= 77.40388 
#>                  coef    SE.coef   Var.coef
#> parameter1 -0.1229702 0.01534980 0.00023562
#> parameter2  0.2002027 0.00989126 0.00009784
#> parameter3  0.1084268 0.00580726 0.00003372
#> parameter4 -0.2808289 0.00360838 0.00001302
#> 
#> ------------------ iteration  5 ------------------
#> Function value  -956.2987 
#> Convergence criteria: parameters stability= 0.00358795 
#>                     : function stability= 155.3761 
#>                     : relative distance to maximum(RDM)= 29.15557 
#>                  coef    SE.coef   Var.coef
#> parameter1 -0.1705625 0.01636548 0.00026783
#> parameter2  0.2037941 0.01086859 0.00011813
#> parameter3  0.1313634 0.00705489 0.00004977
#> parameter4 -0.3088278 0.00441799 0.00001952
#> 
#> ------------------ iteration  6 ------------------
#> Function value  -889.2709 
#> Convergence criteria: parameters stability= 0.00542335 
#>                     : function stability= 67.0278 
#>                     : relative distance to maximum(RDM)= 5.203307 
#>                  coef    SE.coef   Var.coef
#> parameter1 -0.2265610 0.01824680 0.00033295
#> parameter2  0.2269771 0.01175687 0.00013822
#> parameter3  0.1636821 0.00939243 0.00008822
#> parameter4 -0.3353904 0.00532726 0.00002838
#> 
#> ------------------ iteration  7 ------------------
#> Function value  -877.4981 
#> Convergence criteria: parameters stability= 0.00204746 
#>                     : function stability= 11.77276 
#>                     : relative distance to maximum(RDM)= 0.388275 
#>                  coef    SE.coef   Var.coef
#> parameter1 -0.2575585 0.01981871 0.00039278
#> parameter2  0.2387577 0.01212680 0.00014706
#> parameter3  0.1925249 0.01187135 0.00014093
#> parameter4 -0.3461571 0.00573541 0.00003289
#> 
#> ------------------ iteration  8 ------------------
#> Function value  -876.6699 
#> Convergence criteria: parameters stability= 0.00023298 
#>                     : function stability= 0.8281948 
#>                     : relative distance to maximum(RDM)= 0.00495316 
#>                  coef    SE.coef   Var.coef
#> parameter1 -0.2655036 0.02048768 0.00041975
#> parameter2  0.2400265 0.01217954 0.00014834
#> parameter3  0.2054081 0.01313407 0.00017250
#> parameter4 -0.3476645 0.00579217 0.00003355
#> 
#> ------------------ iteration  9 ------------------
#> Function value  -876.6599 
#> Convergence criteria: parameters stability= 3.59e-06 
#>                     : function stability= 0.00999739 
#>                     : relative distance to maximum(RDM)= 1.01e-06 
#>                  coef    SE.coef   Var.coef
#> parameter1 -0.2662980 0.02057557 0.00042335
#> parameter2  0.2400478 0.01218149 0.00014839
#> parameter3  0.2071285 0.01331116 0.00017719
#> parameter4 -0.3477202 0.00579388 0.00003357
#> 
#> ------------------ iteration  10 ------------------
#> Function value  -876.6599 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 2.03e-06 
#>                     : relative distance to maximum(RDM)= 0 
#>                  coef    SE.coef   Var.coef
#> parameter1 -0.2663087 0.02057687 0.00042341
#> parameter2  0.2400479 0.01218151 0.00014839
#> parameter3  0.2071537 0.01331377 0.00017726
#> parameter4 -0.3477207 0.00579390 0.00003357
#> 
#> The program took 0.27 seconds 
#> ------------------ iteration  0 ------------------
#> Function value  -875.8352 
#> Convergence criteria: parameters stability= 1.0001 
#>                     : function stability= 1.0001 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                  coef    SE.coef     Var.coef
#> parameter1  0.0000000   3.551890    12.615925
#> parameter2 -0.2765972   5.516118    30.427559
#> parameter3 -0.2560203   3.037092     9.223927
#> parameter4  0.2339572  88.466961 -7826.403106
#> parameter5  0.2461387  88.440425 -7821.708779
#> parameter6  0.2071537  74.680812  5577.223668
#> parameter7 -0.3477207 172.636101 29803.223496
#> 
#> ------------------ iteration  1 ------------------
#> Function value  -873.994 
#> Convergence criteria: parameters stability= 5.01e-05 
#>                     : function stability= 1.841189 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef    SE.coef    Var.coef
#> parameter1 -0.00000022   3.180447    10.11524
#> parameter2 -0.27663816   5.543917    30.73501
#> parameter3 -0.25597947   3.588780    12.87934
#> parameter4  0.22895886  85.113657 -7244.33456
#> parameter5  0.25113660  84.963084 -7218.72560
#> parameter6  0.20703717  74.367497  5530.52458
#> parameter7 -0.34737316 172.138076 29631.51721
#> 
#> ------------------ iteration  2 ------------------
#> Function value  -859.0964 
#> Convergence criteria: parameters stability= 0.00069688 
#>                     : function stability= 14.89763 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef   SE.coef   Var.coef
#> parameter1 -0.00029738  10.66240  -113.6868
#> parameter2 -0.27701579  32.56422  1060.4287
#> parameter3 -0.25565151  32.65059  1066.0608
#> parameter4  0.21033159  61.16099 -3740.6665
#> parameter5  0.26980804  58.95450 -3475.6335
#> parameter6  0.20648825  72.69177  5284.0932
#> parameter7 -0.34657255 166.43302 27699.9491
#> 
#> ------------------ iteration  3 ------------------
#> Function value  -844.5946 
#> Convergence criteria: parameters stability= 0.00028433 
#>                     : function stability= 14.50182 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef    SE.coef    Var.coef
#> parameter1 -0.00078499   7.539457   -56.84342
#> parameter2 -0.27629067  33.638259  1131.53250
#> parameter3 -0.25655277  33.852721  1146.00673
#> parameter4  0.19880810  40.073633 -1605.89610
#> parameter5  0.28203338  34.527363 -1192.13878
#> parameter6  0.20626924  71.765445  5150.27909
#> parameter7 -0.34589921 161.720276 26153.44766
#> 
#> ------------------ iteration  4 ------------------
#> Function value  -716.1336 
#> Convergence criteria: parameters stability= 0.01387405 
#>                     : function stability= 128.4609 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef    SE.coef     Var.coef
#> parameter1 -0.00501281   2.460415     6.053643
#> parameter2 -0.26937013   6.133523    37.620109
#> parameter3 -0.26627370   5.978390    35.741152
#> parameter4  0.12455005   6.265785    39.260063
#> parameter5  0.37247275   4.847660    23.499806
#> parameter6  0.20587208   7.767472    60.333627
#> parameter7 -0.34141714 129.548623 16782.845797
#> 
#> ------------------ iteration  5 ------------------
#> Function value  -334.7966 
#> Convergence criteria: parameters stability= 0.086728 
#>                     : function stability= 381.337 
#>                     : relative distance to maximum(RDM)= 157.6855 
#>                   coef    SE.coef   Var.coef
#> parameter1  0.01143070 0.14453836 0.02089134
#> parameter2 -0.13485216 0.03282602 0.00107755
#> parameter3 -0.40648825 0.03667541 0.00134509
#> parameter4 -0.01886364 0.01447015 0.00020939
#> parameter5  0.51773548 0.01843621 0.00033989
#> parameter6  0.24073163 0.01691481 0.00028611
#> parameter7 -0.26513822 0.00768029 0.00005899
#> 
#> ------------------ iteration  6 ------------------
#> Function value  -291.6813 
#> Convergence criteria: parameters stability= 0.3068446 
#>                     : function stability= 43.11529 
#>                     : relative distance to maximum(RDM)= 40.46282 
#>                  coef    SE.coef   Var.coef
#> parameter1  0.2032109 0.14215247 0.02020733
#> parameter2  0.1284274 0.02383865 0.00056828
#> parameter3 -0.7343482 0.02694022 0.00072578
#> parameter4 -0.1703930 0.00755994 0.00005715
#> parameter5  0.7605108 0.00843444 0.00007114
#> parameter6  0.2306920 0.01231619 0.00015169
#> parameter7 -0.1590496 0.00185199 0.00000343
#> 
#> ------------------ iteration  7 ------------------
#> Function value  -123.7402 
#> Convergence criteria: parameters stability= 0.00331586 
#>                     : function stability= 167.9411 
#>                     : relative distance to maximum(RDM)= 7.79339 
#>                  coef    SE.coef   Var.coef
#> parameter1  0.2013452 0.14214392 0.02020489
#> parameter2  0.1261218 0.02369173 0.00056130
#> parameter3 -0.7528443 0.02620270 0.00068658
#> parameter4 -0.1466805 0.00871161 0.00007589
#> parameter5  0.7190565 0.00964421 0.00009301
#> parameter6  0.2236083 0.01169290 0.00013672
#> parameter7 -0.1842296 0.00258201 0.00000667
#> 
#> ------------------ iteration  8 ------------------
#> Function value  -92.77588 
#> Convergence criteria: parameters stability= 0.00118282 
#>                     : function stability= 30.96434 
#>                     : relative distance to maximum(RDM)= 0.49567 
#>                  coef    SE.coef   Var.coef
#> parameter1  0.2005246 0.14215307 0.02020749
#> parameter2  0.1153279 0.02430082 0.00059053
#> parameter3 -0.7339871 0.02689008 0.00072308
#> parameter4 -0.1368313 0.00954859 0.00009118
#> parameter5  0.7018656 0.01056544 0.00011163
#> parameter6  0.2255056 0.01217817 0.00014831
#> parameter7 -0.2019473 0.00321731 0.00001035
#> 
#> ------------------ iteration  9 ------------------
#> Function value  -90.96085 
#> Convergence criteria: parameters stability= 7.884e-05 
#>                     : function stability= 1.815033 
#>                     : relative distance to maximum(RDM)= 0.00272533 
#>                  coef    SE.coef   Var.coef
#> parameter1  0.2002665 0.14216235 0.02021013
#> parameter2  0.1129522 0.02442020 0.00059635
#> parameter3 -0.7297423 0.02703135 0.00073069
#> parameter4 -0.1345810 0.00982520 0.00009653
#> parameter5  0.6978214 0.01087440 0.00011825
#> parameter6  0.2250899 0.01222048 0.00014934
#> parameter7 -0.2077364 0.00345066 0.00001191
#> 
#> ------------------ iteration  10 ------------------
#> Function value  -90.95127 
#> Convergence criteria: parameters stability= 4e-07 
#>                     : function stability= 0.00957939 
#>                     : relative distance to maximum(RDM)= 1.1e-07 
#>                  coef    SE.coef   Var.coef
#> parameter1  0.2002341 0.14216300 0.02021032
#> parameter2  0.1128111 0.02442858 0.00059676
#> parameter3 -0.7294758 0.02704144 0.00073124
#> parameter4 -0.1344466 0.00984756 0.00009697
#> parameter5  0.6975642 0.01089945 0.00011880
#> parameter6  0.2250392 0.01222199 0.00014938
#> parameter7 -0.2082035 0.00347009 0.00001204
#> 
#> ------------------ iteration  11 ------------------
#> Function value  -90.95127 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 3.8e-07 
#>                     : relative distance to maximum(RDM)= 0 
#>                  coef    SE.coef   Var.coef
#> parameter1  0.2002339 0.14216243 0.02021016
#> parameter2  0.1128105 0.02442864 0.00059676
#> parameter3 -0.7294744 0.02704150 0.00073124
#> parameter4 -0.1344460 0.00984770 0.00009698
#> parameter5  0.6975628 0.01089961 0.00011880
#> parameter6  0.2250389 0.01222200 0.00014938
#> parameter7 -0.2082065 0.00347022 0.00001204
#> 
#> The program took 0.5 seconds 
#>                A            B
#> 1   6.094830e-14 1.000000e+00
#> 2   4.630003e-18 1.000000e+00
#> 3   1.212453e-15 1.000000e+00
#> 4   1.271356e-16 1.000000e+00
#> 5   2.743807e-07 9.999997e-01
#> 6   1.388856e-18 1.000000e+00
#> 7   3.523930e-16 1.000000e+00
#> 8   4.040637e-18 1.000000e+00
#> 9   7.943657e-15 1.000000e+00
#> 10  3.822145e-14 1.000000e+00
#> 11  1.363469e-14 1.000000e+00
#> 12  1.411162e-19 1.000000e+00
#> 13  6.952657e-16 1.000000e+00
#> 14  1.696443e-13 1.000000e+00
#> 15  3.366902e-15 1.000000e+00
#> 16  1.488731e-16 1.000000e+00
#> 17  1.479891e-15 1.000000e+00
#> 18  1.315933e-11 1.000000e+00
#> 19  1.405337e-15 1.000000e+00
#> 20  1.581632e-14 1.000000e+00
#> 21  1.117063e-11 1.000000e+00
#> 22  1.741506e-13 1.000000e+00
#> 23  1.105955e-14 1.000000e+00
#> 24  2.412696e-16 1.000000e+00
#> 25  8.518368e-14 1.000000e+00
#> 26  9.077320e-18 1.000000e+00
#> 27  5.331399e-16 1.000000e+00
#> 28  1.660181e-11 1.000000e+00
#> 29  2.124229e-15 1.000000e+00
#> 30  2.043707e-15 1.000000e+00
#> 31  3.272026e-14 1.000000e+00
#> 32  9.490982e-10 1.000000e+00
#> 33  3.570117e-13 1.000000e+00
#> 34  1.175206e-17 1.000000e+00
#> 35  2.948949e-19 1.000000e+00
#> 36  8.683956e-18 1.000000e+00
#> 37  4.020959e-16 1.000000e+00
#> 38  2.383720e-14 1.000000e+00
#> 39  1.736211e-16 1.000000e+00
#> 40  2.424831e-12 1.000000e+00
#> 41  4.193581e-17 1.000000e+00
#> 42  3.559845e-15 1.000000e+00
#> 43  2.219354e-14 1.000000e+00
#> 44  1.121084e-11 1.000000e+00
#> 45  1.197758e-14 1.000000e+00
#> 46  1.368138e-10 1.000000e+00
#> 47  2.046871e-12 1.000000e+00
#> 48  1.506056e-13 1.000000e+00
#> 49  3.872326e-15 1.000000e+00
#> 50  1.534281e-14 1.000000e+00
#> 51  3.059323e-16 1.000000e+00
#> 52  4.266342e-18 1.000000e+00
#> 53  3.698762e-13 1.000000e+00
#> 54  1.958998e-12 1.000000e+00
#> 55  8.932947e-17 1.000000e+00
#> 56  7.716031e-14 1.000000e+00
#> 57  1.239058e-11 1.000000e+00
#> 58  1.953540e-17 1.000000e+00
#> 59  7.556522e-16 1.000000e+00
#> 60  2.009338e-12 1.000000e+00
#> 61  2.753635e-14 1.000000e+00
#> 62  4.250818e-16 1.000000e+00
#> 63  4.594170e-13 1.000000e+00
#> 64  8.394654e-19 1.000000e+00
#> 65  1.543557e-18 1.000000e+00
#> 66  1.186362e-15 1.000000e+00
#> 67  1.423975e-13 1.000000e+00
#> 68  2.268949e-11 1.000000e+00
#> 69  3.671334e-14 1.000000e+00
#> 70  5.063660e-12 1.000000e+00
#> 71  5.269645e-14 1.000000e+00
#> 72  4.978410e-13 1.000000e+00
#> 73  5.806369e-12 1.000000e+00
#> 74  2.276530e-19 1.000000e+00
#> 75  4.271275e-14 1.000000e+00
#> 76  1.010145e-17 1.000000e+00
#> 77  4.500632e-10 1.000000e+00
#> 78  1.160327e-15 1.000000e+00
#> 79  6.417437e-12 1.000000e+00
#> 80  3.111558e-14 1.000000e+00
#> 81  1.529621e-18 1.000000e+00
#> 82  1.383492e-15 1.000000e+00
#> 83  1.084538e-17 1.000000e+00
#> 84  2.631487e-12 1.000000e+00
#> 85  2.314740e-12 1.000000e+00
#> 86  2.723653e-10 1.000000e+00
#> 87  4.724998e-16 1.000000e+00
#> 88  6.625740e-15 1.000000e+00
#> 89  1.099063e-16 1.000000e+00
#> 90  5.120131e-15 1.000000e+00
#> 91  1.000000e+00 8.612416e-13
#> 92  1.000000e+00 3.210322e-09
#> 93  1.000000e+00 2.211643e-11
#> 94  1.000000e+00 7.241164e-14
#> 95  9.995726e-01 4.274148e-04
#> 96  1.000000e+00 1.823676e-10
#> 97  9.999991e-01 8.775068e-07
#> 98  9.999996e-01 4.306488e-07
#> 99  1.000000e+00 1.698033e-08
#> 100 1.000000e+00 1.391716e-09
#> 101 1.000000e+00 1.079550e-10
#> 102 1.000000e+00 2.040772e-12
#> 103 9.999987e-01 1.250269e-06
#> 104 1.000000e+00 2.131705e-10
#> 105 9.992816e-01 7.183910e-04
#> 106 9.847819e-01 1.521812e-02
#> 107 1.000000e+00 6.518123e-11
#> 108 9.999924e-01 7.639357e-06
#> 109 1.000000e+00 5.103399e-13
#> 110 1.000000e+00 1.530179e-10
#> 111 9.999925e-01 7.482198e-06
#> 112 9.999952e-01 4.795215e-06
#> 113 1.000000e+00 5.454513e-10
#> 114 1.000000e+00 6.683491e-11
#> 115 1.000000e+00 1.653210e-09
#> 116 9.999999e-01 1.028519e-07
#> 117 1.000000e+00 4.591747e-11
#> 118 1.000000e+00 5.444763e-09
#> 119 1.000000e+00 3.294277e-13
#> 120 1.000000e+00 8.930416e-10
#> 121 9.999278e-01 7.216735e-05
#> 122 9.982712e-01 1.728810e-03
#> 123 1.000000e+00 1.192406e-08
#> 124 1.000000e+00 2.166077e-16
#> 125 1.000000e+00 1.169246e-08
#> 126 1.000000e+00 3.883080e-09
#> 127 1.000000e+00 5.354953e-11
#> 128 9.999999e-01 7.742748e-08
#> 129 1.000000e+00 1.496865e-14
#> 130 1.000000e+00 3.487916e-08
#> 131 1.000000e+00 2.258823e-14
#> 132 1.000000e+00 1.019790e-17
#> 133 1.000000e+00 3.426510e-10
#> 134 9.999890e-01 1.098875e-05
#> 135 1.000000e+00 7.001812e-14
#> 136 1.000000e+00 2.110542e-08
#> 137 9.999999e-01 7.898090e-08
#> 138 1.000000e+00 2.384011e-09
#> 139 1.000000e+00 2.383921e-09
#> 140 1.000000e+00 3.749213e-08
#> 141 1.000000e+00 1.010198e-08
#> 142 1.000000e+00 3.194983e-14
#> 143 1.000000e+00 3.078339e-11
#> 144 1.000000e+00 1.617521e-09
#> 145 1.000000e+00 2.671032e-20
#> 146 1.000000e+00 1.569012e-17
#> 147 9.999995e-01 5.042910e-07
#> 148 1.000000e+00 1.167889e-09
#> 149 9.999982e-01 1.795880e-06
#> 150 9.978728e-01 2.127222e-03
#> 151 9.999994e-01 6.359506e-07
#> 152 1.000000e+00 1.425801e-08
#> 153 1.000000e+00 2.472028e-12
#> 154 1.000000e+00 6.889220e-09
#> 155 9.998621e-01 1.379466e-04
#> 156 9.990900e-01 9.099658e-04
#> 157 1.000000e+00 1.053469e-12
#> 158 1.000000e+00 2.746278e-10
#> 159 1.000000e+00 3.539671e-11
#> 160 1.000000e+00 6.383873e-12
#> 161 1.000000e+00 3.817563e-20
#> 162 9.997543e-01 2.456540e-04
#> 163 1.000000e+00 1.559017e-11
#> 164 1.000000e+00 1.692386e-10
#> 165 9.999999e-01 6.955004e-08
#> 166 1.000000e+00 7.289017e-09
#> 167 1.000000e+00 4.184024e-08
#> 168 9.999994e-01 5.681451e-07
#> 169 1.000000e+00 1.613878e-11
#> 170 1.000000e+00 4.635950e-15
#> 171 1.000000e+00 5.645835e-32
#> 172 1.000000e+00 1.780517e-29
#> 173 1.000000e+00 5.888222e-32
#> 174 1.000000e+00 3.013637e-29
#> 175 1.000000e+00 7.994967e-33
#> 176 1.000000e+00 8.473971e-32
#> 177 1.000000e+00 1.084935e-27
#> 178 1.000000e+00 8.183935e-31
#> 179 1.000000e+00 8.323865e-31
#> 180 1.000000e+00 5.378962e-23
#> 181 1.000000e+00 1.157007e-28
#> 182 1.000000e+00 1.369166e-23
#> 183 1.000000e+00 1.702039e-32
#> 184 1.000000e+00 2.906181e-28
#> 185 1.000000e+00 5.648319e-26
#> 186 1.000000e+00 1.216591e-27
#> 187 1.000000e+00 1.715859e-36
#> 188 1.000000e+00 9.290557e-25
#> 189 1.000000e+00 6.923938e-35
#> 190 1.000000e+00 2.195182e-30
#> 191 1.000000e+00 9.979930e-23
#> 192 1.000000e+00 1.802344e-28
#> 193 1.000000e+00 4.256601e-23
#> 194 1.000000e+00 2.988467e-30
#> 195 1.000000e+00 1.623734e-28
#> 196 1.000000e+00 2.651893e-34
#> 197 1.000000e+00 7.514509e-22
#> 198 1.000000e+00 3.992683e-29
#> 199 1.000000e+00 3.842338e-30
#> 200 1.000000e+00 1.775559e-26