Posterior probability per fitted trajectory

Get the posterior probability matrix with element \((i,j)\) indicating the probability of trajectory \(i\) belonging to cluster \(j\).

# S4 method for lcModel
postprob(object, ...)

Arguments

object

The lcModel.

...

Additional arguments.

Value

A I-by-K matrix with I = nIds(object) and K = nClusters(object).

Details

This method should be extended by lcModel implementations. The default implementation returns uniform probabilities for all observations.

Implementation

Classes extending lcModel should override this method.

setMethod("postprob", "lcModelExt", function(object, ...) {
  # return trajectory-specific posterior probability matrix
})

Troubleshooting

If you are getting errors about undefined model signatures when calling postprob(model), check whether the postprob() function is still the one defined by the latrend package. It may have been overridden when attaching another package (e.g., lcmm). If you need to attach conflicting packages, load them first.

See also

Examples

data(latrendData)
method <- lcMethodLMKM(Y ~ Time, id = "Id", time = "Time")
model <- latrend(method, latrendData)

postprob(model)
#>        A B
#>   [1,] 0 1
#>   [2,] 0 1
#>   [3,] 0 1
#>   [4,] 0 1
#>   [5,] 0 1
#>   [6,] 0 1
#>   [7,] 0 1
#>   [8,] 0 1
#>   [9,] 0 1
#>  [10,] 0 1
#>  [11,] 0 1
#>  [12,] 0 1
#>  [13,] 0 1
#>  [14,] 0 1
#>  [15,] 0 1
#>  [16,] 0 1
#>  [17,] 0 1
#>  [18,] 0 1
#>  [19,] 0 1
#>  [20,] 0 1
#>  [21,] 0 1
#>  [22,] 0 1
#>  [23,] 0 1
#>  [24,] 0 1
#>  [25,] 0 1
#>  [26,] 0 1
#>  [27,] 0 1
#>  [28,] 0 1
#>  [29,] 0 1
#>  [30,] 0 1
#>  [31,] 0 1
#>  [32,] 0 1
#>  [33,] 0 1
#>  [34,] 0 1
#>  [35,] 0 1
#>  [36,] 0 1
#>  [37,] 0 1
#>  [38,] 0 1
#>  [39,] 0 1
#>  [40,] 0 1
#>  [41,] 0 1
#>  [42,] 0 1
#>  [43,] 0 1
#>  [44,] 0 1
#>  [45,] 0 1
#>  [46,] 0 1
#>  [47,] 0 1
#>  [48,] 0 1
#>  [49,] 0 1
#>  [50,] 0 1
#>  [51,] 0 1
#>  [52,] 0 1
#>  [53,] 0 1
#>  [54,] 0 1
#>  [55,] 0 1
#>  [56,] 0 1
#>  [57,] 0 1
#>  [58,] 0 1
#>  [59,] 0 1
#>  [60,] 0 1
#>  [61,] 0 1
#>  [62,] 0 1
#>  [63,] 0 1
#>  [64,] 0 1
#>  [65,] 0 1
#>  [66,] 0 1
#>  [67,] 0 1
#>  [68,] 0 1
#>  [69,] 0 1
#>  [70,] 0 1
#>  [71,] 0 1
#>  [72,] 0 1
#>  [73,] 0 1
#>  [74,] 0 1
#>  [75,] 0 1
#>  [76,] 0 1
#>  [77,] 0 1
#>  [78,] 0 1
#>  [79,] 0 1
#>  [80,] 0 1
#>  [81,] 1 0
#>  [82,] 1 0
#>  [83,] 1 0
#>  [84,] 1 0
#>  [85,] 1 0
#>  [86,] 1 0
#>  [87,] 1 0
#>  [88,] 1 0
#>  [89,] 1 0
#>  [90,] 1 0
#>  [91,] 1 0
#>  [92,] 1 0
#>  [93,] 1 0
#>  [94,] 1 0
#>  [95,] 1 0
#>  [96,] 1 0
#>  [97,] 1 0
#>  [98,] 1 0
#>  [99,] 1 0
#> [100,] 1 0
#> [101,] 1 0
#> [102,] 1 0
#> [103,] 1 0
#> [104,] 1 0
#> [105,] 1 0
#> [106,] 1 0
#> [107,] 1 0
#> [108,] 1 0
#> [109,] 1 0
#> [110,] 1 0
#> [111,] 1 0
#> [112,] 1 0
#> [113,] 1 0
#> [114,] 1 0
#> [115,] 1 0
#> [116,] 1 0
#> [117,] 1 0
#> [118,] 1 0
#> [119,] 1 0
#> [120,] 1 0
#> [121,] 1 0
#> [122,] 1 0
#> [123,] 1 0
#> [124,] 1 0
#> [125,] 1 0
#> [126,] 1 0
#> [127,] 1 0
#> [128,] 1 0
#> [129,] 1 0
#> [130,] 1 0
#> [131,] 1 0
#> [132,] 1 0
#> [133,] 1 0
#> [134,] 1 0
#> [135,] 1 0
#> [136,] 1 0
#> [137,] 1 0
#> [138,] 1 0
#> [139,] 1 0
#> [140,] 1 0
#> [141,] 1 0
#> [142,] 1 0
#> [143,] 1 0
#> [144,] 1 0
#> [145,] 1 0
#> [146,] 1 0
#> [147,] 1 0
#> [148,] 1 0
#> [149,] 1 0
#> [150,] 1 0
#> [151,] 1 0
#> [152,] 1 0
#> [153,] 1 0
#> [154,] 1 0
#> [155,] 1 0
#> [156,] 1 0
#> [157,] 1 0
#> [158,] 1 0
#> [159,] 1 0
#> [160,] 1 0
#> [161,] 1 0
#> [162,] 1 0
#> [163,] 1 0
#> [164,] 1 0
#> [165,] 1 0
#> [166,] 1 0
#> [167,] 1 0
#> [168,] 1 0
#> [169,] 1 0
#> [170,] 1 0
#> [171,] 1 0
#> [172,] 1 0
#> [173,] 1 0
#> [174,] 1 0
#> [175,] 1 0
#> [176,] 1 0
#> [177,] 1 0
#> [178,] 1 0
#> [179,] 1 0
#> [180,] 1 0
#> [181,] 1 0
#> [182,] 1 0
#> [183,] 1 0
#> [184,] 1 0
#> [185,] 1 0
#> [186,] 1 0
#> [187,] 1 0
#> [188,] 1 0
#> [189,] 1 0
#> [190,] 1 0
#> [191,] 1 0
#> [192,] 1 0
#> [193,] 1 0
#> [194,] 1 0
#> [195,] 1 0
#> [196,] 1 0
#> [197,] 1 0
#> [198,] 1 0
#> [199,] 1 0
#> [200,] 1 0

if (rlang::is_installed("lcmm")) {
  gmmMethod = lcMethodLcmmGMM(
    fixed = Y ~ Time,
    mixture = ~ Time,
    id = "Id",
    time = "Time"
  )
  gmmModel <- latrend(gmmMethod, data = latrendData)
  postprob(gmmModel)
}
#> ------------------ iteration  0 ------------------
#> Function value  -2211.674 
#> Convergence criteria: parameters stability= 1.0001 
#>                     : function stability= 1.0001 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef
#> parameter1 -0.03295723
#> parameter2  0.00000000
#> parameter3  1.00000000
#> parameter4  1.00000000
#> 
#> ------------------ iteration  1 ------------------
#> Function value  -2211.661 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 0.01365446 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef
#> parameter1 -0.03295781
#> parameter2  0.00000137
#> parameter3  0.99999000
#> parameter4  0.99999246
#> 
#> ------------------ iteration  2 ------------------
#> Function value  -2211.643 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 0.01733755 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef
#> parameter1 -0.03295808
#> parameter2  0.00000233
#> parameter3  0.99998055
#> parameter4  0.99998246
#> 
#> ------------------ iteration  3 ------------------
#> Function value  -1439.862 
#> Convergence criteria: parameters stability= 0.280072 
#>                     : function stability= 771.7816 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef
#> parameter1 -0.04031975
#> parameter2  0.04576697
#> parameter3  0.64986537
#> parameter4  0.60584696
#> 
#> ------------------ iteration  4 ------------------
#> Function value  -1439.841 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 0.02074656 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef
#> parameter1 -0.04032027
#> parameter2  0.04576783
#> parameter3  0.64985797
#> parameter4  0.60583696
#> 
#> ------------------ iteration  5 ------------------
#> Function value  -1259.308 
#> Convergence criteria: parameters stability= 0.1423089 
#>                     : function stability= 180.5326 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef
#> parameter1 -0.05020431
#> parameter2  0.06607079
#> parameter3  0.47182147
#> parameter4  0.27402077
#> 
#> ------------------ iteration  6 ------------------
#> Function value  -1182.724 
#> Convergence criteria: parameters stability= 1.186942 
#>                     : function stability= 76.58411 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef
#> parameter1 -0.17261108
#> parameter2  0.09024304
#> parameter3 -0.61021076
#> parameter4  0.29811228
#> 
#> ------------------ iteration  7 ------------------
#> Function value  -1182.721 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 0.00300208 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef
#> parameter1 -0.17261028
#> parameter2  0.09024312
#> parameter3 -0.61020076
#> parameter4  0.29811233
#> 
#> ------------------ iteration  8 ------------------
#> Function value  -1182.718 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 0.00380941 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef
#> parameter1 -0.17260961
#> parameter2  0.09024344
#> parameter3 -0.61019076
#> parameter4  0.29811260
#> 
#> ------------------ iteration  9 ------------------
#> Function value  -1070.899 
#> Convergence criteria: parameters stability= 0.05788247 
#>                     : function stability= 111.819 
#>                     : relative distance to maximum(RDM)= 735.0503 
#>                  coef
#> parameter1 -0.1578944
#> parameter2  0.1032584
#> parameter3 -0.3706614
#> parameter4  0.3091681
#> 
#> ------------------ iteration  10 ------------------
#> Function value  -1070.895 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 0.00369269 
#>                     : relative distance to maximum(RDM)= 734.1069 
#>                  coef
#> parameter1 -0.1578937
#> parameter2  0.1032586
#> parameter3 -0.3706514
#> parameter4  0.3091682
#> 
#> ------------------ iteration  11 ------------------
#> Function value  -1070.891 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 0.00379712 
#>                     : relative distance to maximum(RDM)= 732.7804 
#>                  coef
#> parameter1 -0.1578932
#> parameter2  0.1032587
#> parameter3 -0.3706414
#> parameter4  0.3091683
#> 
#> ------------------ iteration  12 ------------------
#> Function value  -1044.445 
#> Convergence criteria: parameters stability= 0.5764923 
#>                     : function stability= 26.44575 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                  coef
#> parameter1 -0.1270200
#> parameter2  0.1190798
#> parameter3  0.3877042
#> parameter4  0.3233399
#> 
#> ------------------ iteration  13 ------------------
#> Function value  -1044.442 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 0.00348247 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                  coef
#> parameter1 -0.1270206
#> parameter2  0.1190798
#> parameter3  0.3876942
#> parameter4  0.3233399
#> 
#> ------------------ iteration  14 ------------------
#> Function value  -1044.438 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 0.00354728 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                  coef
#> parameter1 -0.1270210
#> parameter2  0.1190798
#> parameter3  0.3876842
#> parameter4  0.3233400
#> 
#> ------------------ iteration  15 ------------------
#> Function value  -1044.435 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 0.00359666 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                  coef
#> parameter1 -0.1270214
#> parameter2  0.1190799
#> parameter3  0.3876742
#> parameter4  0.3233402
#> 
#> ------------------ iteration  16 ------------------
#> Function value  -990.0216 
#> Convergence criteria: parameters stability= 0.03537843 
#>                     : function stability= 54.41312 
#>                     : relative distance to maximum(RDM)= 12.63568 
#>                  coef
#> parameter1 -0.1312689
#> parameter2  0.1205435
#> parameter3  0.1996608
#> parameter4  0.3263760
#> 
#> ------------------ iteration  17 ------------------
#> Function value  -973.447 
#> Convergence criteria: parameters stability= 0.00032527 
#>                     : function stability= 16.5746 
#>                     : relative distance to maximum(RDM)= 5.923468 
#>                  coef
#> parameter1 -0.1424467
#> parameter2  0.1250521
#> parameter3  0.1966762
#> parameter4  0.3394560
#> 
#> ------------------ iteration  18 ------------------
#> Function value  -961.3927 
#> Convergence criteria: parameters stability= 0.00105038 
#>                     : function stability= 12.05426 
#>                     : relative distance to maximum(RDM)= 0.593486 
#>                  coef
#> parameter1 -0.1651658
#> parameter2  0.1362990
#> parameter3  0.1897438
#> parameter4  0.3584212
#> 
#> ------------------ iteration  19 ------------------
#> Function value  -960.1702 
#> Convergence criteria: parameters stability= 0.00024153 
#>                     : function stability= 1.222577 
#>                     : relative distance to maximum(RDM)= 0.00468292 
#>                  coef
#> parameter1 -0.1762450
#> parameter2  0.1435649
#> parameter3  0.1871356
#> parameter4  0.3661144
#> 
#> ------------------ iteration  20 ------------------
#> Function value  -960.1608 
#> Convergence criteria: parameters stability= 2.61e-06 
#>                     : function stability= 0.00939256 
#>                     : relative distance to maximum(RDM)= 5.3e-07 
#>                  coef
#> parameter1 -0.1773901
#> parameter2  0.1444374
#> parameter3  0.1869344
#> parameter4  0.3668221
#> 
#> ------------------ iteration  21 ------------------
#> Function value  -960.1608 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 1.06e-06 
#>                     : relative distance to maximum(RDM)= 0 
#>                  coef
#> parameter1 -0.1774069
#> parameter2  0.1444497
#> parameter3  0.1869323
#> parameter4  0.3668285
#> 
#> The program took 0.47 seconds 
#> ------------------ iteration  0 ------------------
#> Function value  -959.2723 
#> Convergence criteria: parameters stability= 1.0001 
#>                     : function stability= 1.0001 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                  coef
#> parameter1  0.0000000
#> parameter2 -0.1874958
#> parameter3 -0.1673180
#> parameter4  0.1380242
#> parameter5  0.1508751
#> parameter6  0.1869323
#> parameter7  0.3668285
#> 
#> ------------------ iteration  1 ------------------
#> Function value  -957.172 
#> Convergence criteria: parameters stability= 6.145e-05 
#>                     : function stability= 2.100332 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                  coef
#> parameter1  0.0000005
#> parameter2 -0.1875478
#> parameter3 -0.1672656
#> parameter4  0.1324900
#> parameter5  0.1564099
#> parameter6  0.1867822
#> parameter7  0.3664314
#> 
#> ------------------ iteration  2 ------------------
#> Function value  -939.807 
#> Convergence criteria: parameters stability= 0.00087023 
#>                     : function stability= 17.36502 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef
#> parameter1  0.00005488
#> parameter2 -0.18764726
#> parameter3 -0.16715930
#> parameter4  0.11164707
#> parameter5  0.17725207
#> parameter6  0.18612558
#> parameter7  0.36545858
#> 
#> ------------------ iteration  3 ------------------
#> Function value  -920.3918 
#> Convergence criteria: parameters stability= 0.00046485 
#>                     : function stability= 19.41517 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef
#> parameter1 -0.00008149
#> parameter2 -0.18638160
#> parameter3 -0.16848824
#> parameter4  0.09655969
#> parameter5  0.19251663
#> parameter6  0.18589413
#> parameter7  0.36457657
#> 
#> ------------------ iteration  4 ------------------
#> Function value  -870.4228 
#> Convergence criteria: parameters stability= 0.00225473 
#>                     : function stability= 49.96899 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef
#> parameter1 -0.00013076
#> parameter2 -0.18534977
#> parameter3 -0.16962638
#> parameter4  0.06420283
#> parameter5  0.22723004
#> parameter6  0.18585681
#> parameter7  0.36396508
#> 
#> ------------------ iteration  5 ------------------
#> Function value  -807.3687 
#> Convergence criteria: parameters stability= 0.00288987 
#>                     : function stability= 63.05413 
#>                     : relative distance to maximum(RDM)= 1.0001 
#>                   coef
#> parameter1  0.00154924
#> parameter2 -0.17307886
#> parameter3 -0.18296970
#> parameter4  0.03017905
#> parameter5  0.26391804
#> parameter6  0.18744110
#> parameter7  0.35673374
#> 
#> ------------------ iteration  6 ------------------
#> Function value  -234.0025 
#> Convergence criteria: parameters stability= 0.2354196 
#>                     : function stability= 573.3662 
#>                     : relative distance to maximum(RDM)= 88.25512 
#>                   coef
#> parameter1  0.15233864
#> parameter2  0.03617987
#> parameter3 -0.40084166
#> parameter4 -0.17597225
#> parameter5  0.51667596
#> parameter6  0.24031620
#> parameter7  0.24608093
#> 
#> ------------------ iteration  7 ------------------
#> Function value  -116.8424 
#> Convergence criteria: parameters stability= 0.2750103 
#>                     : function stability= 117.1601 
#>                     : relative distance to maximum(RDM)= 14.02445 
#>                  coef
#> parameter1  0.3350642
#> parameter2  0.2213973
#> parameter3 -0.7824800
#> parameter4 -0.2305114
#> parameter5  0.7425480
#> parameter6  0.1864792
#> parameter7  0.1769633
#> 
#> ------------------ iteration  8 ------------------
#> Function value  -60.66954 
#> Convergence criteria: parameters stability= 0.00715973 
#>                     : function stability= 56.17281 
#>                     : relative distance to maximum(RDM)= 1.323995 
#>                  coef
#> parameter1  0.3942235
#> parameter2  0.1995742
#> parameter3 -0.7462780
#> parameter4 -0.2256329
#> parameter5  0.7082258
#> parameter6  0.2036181
#> parameter7  0.1963930
#> 
#> ------------------ iteration  9 ------------------
#> Function value  -55.73164 
#> Convergence criteria: parameters stability= 0.00050486 
#>                     : function stability= 4.937899 
#>                     : relative distance to maximum(RDM)= 0.01668516 
#>                  coef
#> parameter1  0.4052535
#> parameter2  0.1937467
#> parameter3 -0.7346090
#> parameter4 -0.2240926
#> parameter5  0.6980466
#> parameter6  0.2096016
#> parameter7  0.2048358
#> 
#> ------------------ iteration  10 ------------------
#> Function value  -55.67275 
#> Convergence criteria: parameters stability= 4.07e-06 
#>                     : function stability= 0.05888992 
#>                     : relative distance to maximum(RDM)= 3.06e-06 
#>                  coef
#> parameter1  0.4054646
#> parameter2  0.1932927
#> parameter3 -0.7334642
#> parameter4 -0.2239387
#> parameter5  0.6970418
#> parameter6  0.2101172
#> parameter7  0.2059342
#> 
#> ------------------ iteration  11 ------------------
#> Function value  -55.67274 
#> Convergence criteria: parameters stability= 0 
#>                     : function stability= 1.071e-05 
#>                     : relative distance to maximum(RDM)= 0 
#>                  coef
#> parameter1  0.4054651
#> parameter2  0.1932891
#> parameter3 -0.7334510
#> parameter4 -0.2239371
#> parameter5  0.6970299
#> parameter6  0.2101202
#> parameter7  0.2059496
#> 
#> The program took 0.44 seconds 
#>                A            B
#> 1   2.547309e-17 1.000000e+00
#> 2   5.411522e-22 1.000000e+00
#> 3   3.041642e-19 1.000000e+00
#> 4   2.235736e-20 1.000000e+00
#> 5   8.499348e-10 1.000000e+00
#> 6   1.428899e-22 1.000000e+00
#> 7   7.217099e-20 1.000000e+00
#> 8   4.601011e-22 1.000000e+00
#> 9   2.465789e-18 1.000000e+00
#> 10  1.487810e-17 1.000000e+00
#> 11  4.431033e-18 1.000000e+00
#> 12  1.048311e-23 1.000000e+00
#> 13  1.624114e-19 1.000000e+00
#> 14  8.535574e-17 1.000000e+00
#> 15  9.154080e-19 1.000000e+00
#> 16  2.775977e-20 1.000000e+00
#> 17  3.751780e-19 1.000000e+00
#> 18  1.074618e-14 1.000000e+00
#> 19  3.429499e-19 1.000000e+00
#> 20  5.383558e-18 1.000000e+00
#> 21  8.954243e-15 1.000000e+00
#> 22  8.071920e-17 1.000000e+00
#> 23  3.636239e-18 1.000000e+00
#> 24  5.106287e-20 1.000000e+00
#> 25  3.644339e-17 1.000000e+00
#> 26  1.171056e-21 1.000000e+00
#> 27  1.186961e-19 1.000000e+00
#> 28  1.506209e-14 1.000000e+00
#> 29  5.704580e-19 1.000000e+00
#> 30  5.336798e-19 1.000000e+00
#> 31  1.202215e-17 1.000000e+00
#> 32  1.410158e-12 1.000000e+00
#> 33  1.841029e-16 1.000000e+00
#> 34  1.571870e-21 1.000000e+00
#> 35  2.531197e-23 1.000000e+00
#> 36  1.134933e-21 1.000000e+00
#> 37  8.624020e-20 1.000000e+00
#> 38  8.734216e-18 1.000000e+00
#> 39  3.199024e-20 1.000000e+00
#> 40  1.589164e-15 1.000000e+00
#> 41  6.669018e-21 1.000000e+00
#> 42  1.013342e-18 1.000000e+00
#> 43  7.838956e-18 1.000000e+00
#> 44  9.027838e-15 1.000000e+00
#> 45  4.032962e-18 1.000000e+00
#> 46  1.606951e-13 1.000000e+00
#> 47  1.326813e-15 1.000000e+00
#> 48  6.837974e-17 1.000000e+00
#> 49  1.112480e-18 1.000000e+00
#> 50  5.127967e-18 1.000000e+00
#> 51  6.227700e-20 1.000000e+00
#> 52  5.071706e-22 1.000000e+00
#> 53  1.910355e-16 1.000000e+00
#> 54  1.315540e-15 1.000000e+00
#> 55  1.489717e-20 1.000000e+00
#> 56  3.086665e-17 1.000000e+00
#> 57  1.051878e-14 1.000000e+00
#> 58  2.876305e-21 1.000000e+00
#> 59  1.715890e-19 1.000000e+00
#> 60  1.321604e-15 1.000000e+00
#> 61  9.535077e-18 1.000000e+00
#> 62  9.190271e-20 1.000000e+00
#> 63  2.424449e-16 1.000000e+00
#> 64  7.904951e-23 1.000000e+00
#> 65  1.599877e-22 1.000000e+00
#> 66  2.884337e-19 1.000000e+00
#> 67  6.949903e-17 1.000000e+00
#> 68  1.967578e-14 1.000000e+00
#> 69  1.408407e-17 1.000000e+00
#> 70  3.518169e-15 1.000000e+00
#> 71  2.106024e-17 1.000000e+00
#> 72  2.756480e-16 1.000000e+00
#> 73  4.318563e-15 1.000000e+00
#> 74  1.844605e-23 1.000000e+00
#> 75  1.737188e-17 1.000000e+00
#> 76  1.314989e-21 1.000000e+00
#> 77  6.142480e-13 1.000000e+00
#> 78  2.824823e-19 1.000000e+00
#> 79  4.868013e-15 1.000000e+00
#> 80  1.192237e-17 1.000000e+00
#> 81  9.999999e-01 5.541483e-08
#> 82  1.000000e+00 1.687982e-09
#> 83  1.000000e+00 1.237469e-08
#> 84  1.000000e+00 2.646910e-11
#> 85  1.000000e+00 2.743526e-11
#> 86  1.000000e+00 2.225112e-12
#> 87  1.000000e+00 3.498192e-09
#> 88  1.000000e+00 3.840670e-10
#> 89  1.000000e+00 4.912743e-09
#> 90  1.000000e+00 4.347079e-10
#> 91  1.000000e+00 6.570054e-11
#> 92  1.000000e+00 9.025892e-10
#> 93  1.000000e+00 7.647599e-11
#> 94  1.000000e+00 2.499976e-10
#> 95  1.000000e+00 1.745679e-08
#> 96  1.000000e+00 3.376375e-10
#> 97  1.000000e+00 7.724647e-09
#> 98  1.000000e+00 3.138129e-09
#> 99  1.000000e+00 7.196848e-09
#> 100 1.000000e+00 3.432416e-09
#> 101 1.000000e+00 3.378190e-11
#> 102 1.000000e+00 6.094949e-10
#> 103 1.000000e+00 3.903088e-10
#> 104 1.000000e+00 1.463033e-09
#> 105 1.000000e+00 3.266176e-08
#> 106 9.999997e-01 2.691745e-07
#> 107 1.000000e+00 2.165185e-10
#> 108 1.000000e+00 4.417720e-08
#> 109 1.000000e+00 2.613364e-10
#> 110 1.000000e+00 1.855565e-10
#> 111 1.000000e+00 1.247697e-08
#> 112 9.999999e-01 1.177873e-07
#> 113 1.000000e+00 2.975804e-09
#> 114 1.000000e+00 1.558266e-09
#> 115 1.000000e+00 1.849239e-09
#> 116 1.000000e+00 1.031145e-09
#> 117 1.000000e+00 7.985078e-10
#> 118 1.000000e+00 4.550104e-09
#> 119 1.000000e+00 4.156114e-10
#> 120 1.000000e+00 3.206136e-10
#> 121 1.000000e+00 2.233346e-08
#> 122 9.999999e-01 8.120601e-08
#> 123 1.000000e+00 5.954044e-10
#> 124 1.000000e+00 3.026458e-11
#> 125 1.000000e+00 1.424766e-09
#> 126 1.000000e+00 8.130364e-10
#> 127 1.000000e+00 6.150024e-10
#> 128 1.000000e+00 9.444138e-09
#> 129 1.000000e+00 2.328948e-11
#> 130 1.000000e+00 1.563133e-09
#> 131 1.000000e+00 7.070482e-12
#> 132 1.000000e+00 2.430530e-12
#> 133 1.000000e+00 1.825761e-09
#> 134 1.000000e+00 6.650521e-09
#> 135 1.000000e+00 1.370103e-11
#> 136 1.000000e+00 1.025182e-09
#> 137 1.000000e+00 1.227042e-09
#> 138 1.000000e+00 2.044575e-09
#> 139 1.000000e+00 5.253970e-09
#> 140 1.000000e+00 3.989508e-08
#> 141 1.000000e+00 6.667898e-10
#> 142 1.000000e+00 6.031687e-11
#> 143 1.000000e+00 9.281413e-11
#> 144 1.000000e+00 3.580572e-09
#> 145 1.000000e+00 5.328300e-13
#> 146 1.000000e+00 2.029222e-12
#> 147 1.000000e+00 3.769147e-09
#> 148 1.000000e+00 2.288466e-09
#> 149 1.000000e+00 1.052566e-08
#> 150 1.000000e+00 3.274686e-08
#> 151 1.000000e+00 2.471732e-26
#> 152 1.000000e+00 8.469882e-29
#> 153 1.000000e+00 1.695607e-31
#> 154 1.000000e+00 1.193504e-30
#> 155 1.000000e+00 8.066766e-25
#> 156 1.000000e+00 3.073320e-26
#> 157 1.000000e+00 1.101321e-34
#> 158 1.000000e+00 2.275016e-31
#> 159 1.000000e+00 2.911821e-35
#> 160 1.000000e+00 2.330361e-32
#> 161 1.000000e+00 1.275205e-41
#> 162 1.000000e+00 3.527564e-25
#> 163 1.000000e+00 5.453888e-33
#> 164 1.000000e+00 5.219841e-32
#> 165 1.000000e+00 4.904099e-30
#> 166 1.000000e+00 4.439443e-29
#> 167 1.000000e+00 1.565171e-28
#> 168 1.000000e+00 4.165055e-28
#> 169 1.000000e+00 7.260726e-33
#> 170 1.000000e+00 2.191088e-36
#> 171 1.000000e+00 1.604883e-33
#> 172 1.000000e+00 7.948705e-31
#> 173 1.000000e+00 2.416388e-34
#> 174 1.000000e+00 2.302697e-31
#> 175 1.000000e+00 1.701660e-35
#> 176 1.000000e+00 1.088227e-32
#> 177 1.000000e+00 4.202725e-30
#> 178 1.000000e+00 2.132957e-32
#> 179 1.000000e+00 1.985949e-32
#> 180 1.000000e+00 2.624164e-24
#> 181 1.000000e+00 1.952052e-29
#> 182 1.000000e+00 6.747480e-26
#> 183 1.000000e+00 7.691045e-35
#> 184 1.000000e+00 2.337360e-30
#> 185 1.000000e+00 3.671426e-27
#> 186 1.000000e+00 1.975783e-30
#> 187 1.000000e+00 2.035570e-39
#> 188 1.000000e+00 4.969321e-28
#> 189 1.000000e+00 4.456556e-38
#> 190 1.000000e+00 1.971674e-33
#> 191 1.000000e+00 5.329904e-25
#> 192 1.000000e+00 1.312493e-29
#> 193 1.000000e+00 1.176425e-25
#> 194 1.000000e+00 3.285059e-31
#> 195 1.000000e+00 5.007539e-31
#> 196 1.000000e+00 7.281392e-36
#> 197 1.000000e+00 5.466925e-24
#> 198 1.000000e+00 4.488509e-31
#> 199 1.000000e+00 4.130772e-33
#> 200 1.000000e+00 5.307268e-28