lcModel prediction conditional on a cluster

Predicts the expected trajectory observations at the given time under the assumption that the trajectory belongs to the specified cluster.

The same result can be obtained by calling predict() with the newdata data.frame having a "Cluster" assignment column. The main purpose of this function is to make it easier to implement the prediction computations for custom lcModel classes.

# S4 method for lcModel
predictForCluster(object, newdata = NULL, cluster, ..., what = "mu")

Arguments

object

The lcModel object.

newdata

Optional data.frame for which to compute the model predictions. If omitted, the model training data is used. Cluster trajectory predictions are made when ids are not specified.

cluster

The cluster name (as character) to predict for.

...

Additional arguments.

what

The distributional parameter to predict. By default, the mean response 'mu' is predicted. The cluster membership predictions can be obtained by specifying what = 'mb'.

Value

A vector with the predictions per newdata observation, or a data.frame with the predictions and newdata alongside.

Details

The default predictForCluster() method makes use of predict.lcModel(), and vice versa. For this to work, any extending lcModel classes, e.g., lcModelExample, should implement either predictForCluster(lcModelExample) or predict.lcModelExample(). When implementing new models, it is advisable to implement predictForCluster as the cluster-specific computation generally results in shorter and simpler code.

Implementation

Classes extending lcModel should override this method, unless predict.lcModel() is preferred.

setMethod("predictForCluster", "lcModelExt",
 function(object, newdata = NULL, cluster, ..., what = "mu") {
  # return model predictions for the given data under the
  # assumption of the data belonging to the given cluster
})

See also

Examples

data(latrendData)
method <- lcMethodLMKM(Y ~ Time, id = "Id", time = "Time")
model <- latrend(method, latrendData)

predictForCluster(
  model,
  newdata = data.frame(Time = c(0, 1)),
  cluster = "B"
)
#> [1] -0.9797514  0.1191113

# all fitted values under cluster B
predictForCluster(model, cluster = "B")
#>    [1] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#>    [7]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#>   [13] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#>   [19]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#>   [25]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#>   [31] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#>   [37]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#>   [43] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#>   [49]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#>   [55]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#>   [61] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#>   [67]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#>   [73] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#>   [79]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#>   [85]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#>   [91] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#>   [97]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#>  [103] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#>  [109]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#>  [115]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#>  [121] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#>  [127]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#>  [133] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#>  [139]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#>  [145]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#>  [151] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#>  [157]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#>  [163] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#>  [169]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#>  [175]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#>  [181] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#>  [187]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#>  [193] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#>  [199]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#>  [205]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#>  [211] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#>  [217]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#>  [223] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#>  [229]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#>  [235]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#>  [241] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#>  [247]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#>  [253] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#>  [259]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#>  [265]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#>  [271] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#>  [277]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#>  [283] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#>  [289]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#>  [295]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#>  [301] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#>  [307]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#>  [313] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#>  [319]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#>  [325]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#>  [331] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#>  [337]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#>  [343] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#>  [349]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#>  [355]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#>  [361] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#>  [367]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#>  [373] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#>  [379]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#>  [385]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#>  [391] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#>  [397]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#>  [403] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#>  [409]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#>  [415]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#>  [421] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#>  [427]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#>  [433] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#>  [439]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#>  [445]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#>  [451] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#>  [457]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#>  [463] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#>  [469]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#>  [475]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#>  [481] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#>  [487]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#>  [493] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#>  [499]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#>  [505]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#>  [511] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#>  [517]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#>  [523] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#>  [529]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#>  [535]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#>  [541] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#>  [547]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#>  [553] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#>  [559]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#>  [565]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#>  [571] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#>  [577]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#>  [583] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#>  [589]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#>  [595]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#>  [601] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#>  [607]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#>  [613] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#>  [619]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#>  [625]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#>  [631] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#>  [637]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#>  [643] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#>  [649]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#>  [655]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#>  [661] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#>  [667]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#>  [673] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#>  [679]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#>  [685]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#>  [691] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#>  [697]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#>  [703] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#>  [709]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#>  [715]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#>  [721] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#>  [727]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#>  [733] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#>  [739]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#>  [745]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#>  [751] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#>  [757]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#>  [763] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#>  [769]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#>  [775]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#>  [781] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#>  [787]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#>  [793] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#>  [799]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#>  [805]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#>  [811] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#>  [817]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#>  [823] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#>  [829]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#>  [835]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#>  [841] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#>  [847]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#>  [853] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#>  [859]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#>  [865]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#>  [871] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#>  [877]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#>  [883] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#>  [889]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#>  [895]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#>  [901] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#>  [907]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#>  [913] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#>  [919]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#>  [925]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#>  [931] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#>  [937]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#>  [943] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#>  [949]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#>  [955]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#>  [961] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#>  [967]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#>  [973] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#>  [979]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#>  [985]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#>  [991] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#>  [997]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#> [1003] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#> [1009]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#> [1015]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#> [1021] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#> [1027]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#> [1033] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#> [1039]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#> [1045]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#> [1051] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#> [1057]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#> [1063] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#> [1069]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#> [1075]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#> [1081] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#> [1087]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#> [1093] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#> [1099]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#> [1105]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#> [1111] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#> [1117]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#> [1123] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#> [1129]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#> [1135]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#> [1141] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#> [1147]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#> [1153] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#> [1159]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#> [1165]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#> [1171] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#> [1177]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#> [1183] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#> [1189]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#> [1195]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#> [1201] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#> [1207]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#> [1213] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#> [1219]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#> [1225]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#> [1231] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#> [1237]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#> [1243] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#> [1249]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#> [1255]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#> [1261] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#> [1267]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#> [1273] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#> [1279]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#> [1285]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#> [1291] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#> [1297]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#> [1303] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#> [1309]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#> [1315]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#> [1321] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#> [1327]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#> [1333] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#> [1339]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#> [1345]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#> [1351] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#> [1357]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#> [1363] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#> [1369]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#> [1375]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#> [1381] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#> [1387]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#> [1393] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#> [1399]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#> [1405]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#> [1411] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#> [1417]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#> [1423] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#> [1429]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#> [1435]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#> [1441] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#> [1447]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#> [1453] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#> [1459]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#> [1465]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#> [1471] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#> [1477]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#> [1483] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#> [1489]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#> [1495]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#> [1501] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#> [1507]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#> [1513] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#> [1519]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#> [1525]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#> [1531] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#> [1537]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#> [1543] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#> [1549]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#> [1555]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#> [1561] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#> [1567]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#> [1573] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#> [1579]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#> [1585]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#> [1591] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#> [1597]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#> [1603] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#> [1609]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#> [1615]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#> [1621] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#> [1627]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#> [1633] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#> [1639]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#> [1645]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#> [1651] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#> [1657]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#> [1663] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#> [1669]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#> [1675]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#> [1681] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#> [1687]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#> [1693] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#> [1699]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#> [1705]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#> [1711] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#> [1717]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#> [1723] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#> [1729]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#> [1735]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#> [1741] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#> [1747]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#> [1753] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#> [1759]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#> [1765]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#> [1771] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#> [1777]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#> [1783] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#> [1789]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#> [1795]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#> [1801] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#> [1807]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#> [1813] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#> [1819]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#> [1825]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#> [1831] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#> [1837]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#> [1843] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#> [1849]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#> [1855]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#> [1861] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#> [1867]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#> [1873] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#> [1879]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#> [1885]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#> [1891] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#> [1897]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#> [1903] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#> [1909]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#> [1915]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#> [1921] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#> [1927]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#> [1933] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#> [1939]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#> [1945]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#> [1951] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#> [1957]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#> [1963] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#> [1969]  0.42939022  0.43011898 -0.97975135 -0.63534769 -0.38525328 -0.16233773
#> [1975]  0.05903586  0.17918673  0.31069966  0.39004675  0.42939022  0.43011898
#> [1981] -0.97975135 -0.63534769 -0.38525328 -0.16233773  0.05903586  0.17918673
#> [1987]  0.31069966  0.39004675  0.42939022  0.43011898 -0.97975135 -0.63534769
#> [1993] -0.38525328 -0.16233773  0.05903586  0.17918673  0.31069966  0.39004675
#> [1999]  0.42939022  0.43011898